### 2023-03-13 #5250 @editorialbot
[PRE REVIEW]: Reducing the efforts to create reproducible analysis code with FieldTrip
**Submitting author:** @matsvanes (<a href="http://orcid.org/0000-0002-7133-509X">Mats W.J. van Es</a>)
**Repository:** https://github.com/fieldtrip/fieldtrip
**Branch with paper.md** (empty if default branch): JOSS
**Version:** https://github.com/fieldtrip/fieldtrip/commit/9cfc0a7e7b7871a1d9938552ce60041c59f3a553
