### 2023-03-14 3fdb8d5
document the SimNIBS mesh_load_gmsh4 reader, commented-out
