### 2023-03-14 #5250 @Kevin-Mattheus-Moerman
@chartgerink do you think you can help edit this work? I thought that the "reproducible science" aspect might connect with you. Note, there are a number of EMG/ECG/EEG/signal processing papers in JOSS already, and you might check if reviewers for those are applicable to this submission too.
