### 2023-03-14 9eaec59
fix: accommodate a change to in_fopen_nk
