+++ fieldtrip2fiff.m
@@ -1,8 +1,7 @@
 function fieldtrip2fiff(filename, data)
 
 % FIELDTRIP2FIFF saves a FieldTrip raw data structure as a fiff-file, allowing it
-% to be further analyzed by the Neuromag/Elekta/Megin software, or in the MNE suite
-% software.
+% to be further analyzed by the Neuromag/Elekta/Megin software, or in MNE-python.
 %
 % Use as
 %   fieldtrip2fiff(filename, data)
@@ -12,11 +11,11 @@ function fieldtrip2fiff(filename, data)
 %
 % If the data comes from preprocessing and has only one trial, then it writes the
 % data into raw continuous format. If present in the data, the original header
-% from neuromag is reused (also removing the non-used channels). Otherwise, the
-% function tries to create a correct header, which might or might not contain the
-% correct scaling and channel location. If the data contains events in the cfg
-% structure, it writes the events in the MNE format (three columns) into a file
-% based on "filename", ending with "-eve.fif"
+% is reused (also removing the non-used channels). Otherwise, the function 
+% attempts to create the header, which might or might not be correct (e.g. with
+% respect to the scaling and the sensor locations. If the data contains events in 
+% the cfg structure, it writes the events in the MNE format (three columns) into 
+% a file based on "filename", ending with "-eve.fif"
 %
 % See also FT_DATATYPE_RAW, FT_DATATYPE_TIMELOCK
 
@@ -48,73 +47,57 @@ ft_defaults
 if ~strcmp(ext, '.fif')
   ft_error('if the filename is specified with extension, this should read .fif');
 end
-fifffile = fullfile(pathstr ,[name '.fif']);
+fifffile  = fullfile(pathstr ,[name '.fif']);
 eventfile = fullfile(pathstr ,[name '-eve.fif']);
 
 % ensure the mne-toolbox to be on the path
 ft_hastoolbox('mne', 1);
 
 % check if the input data is valid for this function
-data   = ft_checkdata(data, 'datatype', {'raw', 'timelock'}, 'feedback', 'yes');
+data   = ft_checkdata(data, 'datatype', {'raw', 'timelock'}, 'hassampleinfo', 'yes', 'feedback', 'yes');
 istlck = ft_datatype(data, 'timelock');
 isepch = ft_datatype(data, 'raw');
-israw = false;
+israw  = false;
 if isepch && numel(data.trial) == 1
   isepch = false;
-  israw = true;
+  israw  = true;
+end
+if istlck
+  fsample = 1./mean(diff(data.time));
+else
+  fsample = 1./mean(diff(data.time{1}));
 end
 
-% Create a fiff-header, or take it from the original header if possible
-if ft_senstype(data, 'neuromag') && isfield(data, 'hdr')
-  fprintf('Using the original FIFF header, but channel locations are read \nfrom .grad and .elec in the data, if they exist\n')
+% Create a fiff-header, or take information from the original header if possible
+if isfield(data, 'hdr') && isfield(data.hdr, 'orig') && isfield(data.hdr.orig, 'meas_id')
[…]
