### 2023-03-15 b45cdef
ENH - allow writing raw files in double precision
