+++ external/openmeeg/openmeeg_msm.m
@@ -1,0 +1,114 @@
+function [dsm] = openmeeg_msm(pos, headmodel, NAflag, source)
+
+% FT_SYSMAT_OPENMEEG creates a volume conduction model of the
+% head using the boundary element method (BEM). This function takes
+% as input the triangulated surfaces that describe the boundaries and
+% returns as output a volume conduction model which can be used to
+% compute leadfields.
+%
+% This function implements
+%   Gramfort et al. OpenMEEG: opensource software for quasistatic
+%   bioelectromagnetics. Biomedical engineering online (2010) vol. 9 (1) pp. 45
+%   http://www.biomedical-engineering-online.com/content/9/1/45
+%   doi:10.1186/1475-925X-9-45
+% and
+%   Kybic et al. Generalized head models for MEG/EEG: boundary element method
+%   beyond nested volumes. Phys. Med. Biol. (2006) vol. 51 pp. 1333-1346
+%   doi:10.1088/0031-9155/51/5/021
+%
+% This link with FieldTrip is derived from the OpenMEEG project
+% with contributions from Daniel Wong and Sarang Dalal, and uses external
+% command-line executables. See http://openmeeg.github.io/
+%
+% Use as
+%   dsm = ft_sysmat_openmeeg(pos, headmodel, sens, nonadaptive_flag)
+%
+%
+% See also FT_HEADMODEL_OPENMEEG, FT_SENSINTERP_OPENMEEG, FT_PREPARE_LEADFIELD
+
+%$Id$
+
+ft_hastoolbox('openmeeg', 1);  % add to path (if not yet on path)
+openmeeg_license;              % show the license (only once)
+prefix = om_checkombin;        % check the installation of the binaries
+if(~ispc) % if Linux/Mac, set number of threads
+    omp_num_threads = feature('numCores');
+    prefix = ['export OMP_NUM_THREADS=' num2str(omp_num_threads) ';' prefix];
+end
+
+ecog = ft_getopt(headmodel,'ecog','no');
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+% this uses an implementation that was contributed by INRIA Odyssee Team
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+% store the current path and change folder to the temporary one
+cwd = pwd;
+bndom = headmodel.bnd;
+
+workdir = fullfile(tempdir,['ft_om_' datestr(now,'ddmmyyHHMMSSFFF') '_' num2str(randi(1000000))]);
+mkdir(workdir);
+
+try
+    % Write the triangulations to file, named after tissue type.
+    % OpenMEEG v2.3 and up internally adjusts the convention for surface
+    % normals, but OpenMEEG v2.2 expects surface normals to point inwards;
+    % this checks and corrects if needed
+    bndfile = fullfile(workdir,strcat(headmodel.tissue,'.tri'));
+    for ii=1:length(bndom)
+        om_save_tri(bndfile{ii}, bndom(ii).pos, bndom(ii).tri);
+    end
+    
+    condfile  = fullfile(workdir,'om.cond');
+    geomfile  = fullfile(workdir,'om.geom');
+    dipfile = fullfile(workdir,'dip.txt');
+    msmfile = fullfile(workdir,'msm.bin'); % extension added directly in om_assemble function call below
+    
+    % write conductivity and mesh files
+    bndlabel = {};
+    for i=1:length(bndom)
+        [dum,bndlabel{i}] = fileparts(bndfile{i});
[…]
