### 2023-03-17 e1ec03c
add test for corehistration of headshape with spherical headmodel
