+++ plotting/ft_plot_ortho.m
@@ -19,7 +19,7 @@ function [hx, hy, hz] = ft_plot_ortho(dat, varargin)
 %
 % The following options are supported and passed on to FT_PLOT_SLICE
 %   'clim'                = [min max], lower and upper color limits
-%   'facealpha'           = 
+%   'facealpha'           = transparency when no mask is specified, between 0 and 1 (default = 1)
 %   'transform'           = 4x4 homogeneous transformation matrix specifying the mapping from voxel space to the coordinate system in which the data are plotted
 %   'location'            = 1x3 vector specifying the intersection point at which the three slices will be plotted. The coordinates should be expressed in the coordinate system of the data. 
 %   'datmask'             = 3D-matrix with the same size as the matrix dat, serving as opacitymap if the second input argument to the function contains a matrix, this will be used as the mask
