### 2023-03-17 42e2ed3
implemented gmsh reader based on SimNIBS code
