+++ fileio/ft_read_headshape.m
@@ -31,6 +31,8 @@ function [shape] = ft_read_headshape(filename, varargin)
 %
 % Supported input file formats include
 %   'gifti'           see https://www.nitrc.org/projects/gifti/
+%   'gmsh_ascii'      see https://gmsh.info
+%   'gmsh_binary'     see https://gmsh.info
 %   'matlab'          containing FieldTrip or BrainStorm headshapes or cortical meshes
 %   'mne_tri'         MNE surface description in ASCII format
 %   'mne_pos'         MNE source grid in ascii format, described as 3D points
@@ -1346,12 +1348,22 @@ switch fileformat
     shape.hex = shape.hex(:,1:8);
     shape.hex = shape.hex + 1; % this should be one-offset
 
-    %   case 'gmsh_binary'
-    %     % use the SimNIBS reader, this is an alternative to READ_GMSH_BINARY but it has some limitations
-    %     shape = mesh_load_gmsh4(filename);
-    %     shape = fixpos(shape);
+  case {'gmsh_ascii' 'gmsh_binary'}
+    % use the SimNIBS reader, this does not read all gmsh properties/tags
+    ft_hastoolbox('simnibs', 1);
+    shape = mesh_load_gmsh4(filename);
+    shape = fixpos(shape);
 
-  case 'gmsh_binary'
+    % remove empty fields
+    fn = fieldnames(shape);
+    for i=1:numel(fn)
+      if isempty(shape.(fn{i}))
+        shape = rmfield(shape, fn{i});
+      end
+    end
+
+  case 'gmsh_binary_v1'
+    % use Jan-Mathijs' reader, this only works for binary files but does read all gmsh properties/tags
     [nodes, elements] = read_gmsh_binary(filename);
     shape.pos = nodes.nodes(nodes.indx, :);
 
@@ -1380,7 +1392,7 @@ switch fileformat
           ft_warning('skipping element field %s', fnames{k});
       end
     end
-    
+
   case {'neurojson_jmesh' 'neurojson_bmesh'}
     % see https://github.com/NeuroJSON/jmesh/blob/master/JMesh_specification.md
     ft_hastoolbox('jsonlab', 1);
