### 2023-03-17 44bcbb2
add external/simnibs for reading and writing gmsh files
