### 2023-03-17 239873b
print some information about the conversion to edges
