### 2023-03-17 16a72cd
added another case to the test script
