### 2023-03-18 ade836b
ENH - update for double precision support, PR submitted to MNE folks as well
