+++ external/mne/fiff_write_raw_buffer.m
@@ -1,10 +1,11 @@
-function fiff_write_raw_buffer(fid,buf,cals)
+function fiff_write_raw_buffer(fid,buf,cals,datatype)
 %
-% function fiff_write_raw_buffer(fid,info,buf)
+% function fiff_write_raw_buffer(fid,buf,cals,datatype)
 %
 % fid        of an open raw data file
 % buf        the buffer to write
 % cals       calibration factors
+% datatype   (optional) datatype to write, default float
 %
 %
 
@@ -17,7 +18,7 @@ function fiff_write_raw_buffer(fid,buf,cals)
 %
 
 me='MNE:fiff_write_raw_buffer';
-if nargin ~= 3
+if nargin < 3
     error(me,'Incorrect number of arguments');
 end
 
@@ -30,6 +31,25 @@ if isempty(FIFF)
     FIFF = fiff_define_constants();
 end
 
-fiff_write_float(fid,FIFF.FIFF_DATA_BUFFER,inv(diag(cals))*buf); % XXX why not diag(1./cals) ???
+if nargin < 4
+    datatype = FIFF.FIFFT_FLOAT;
+end
+
+if datatype ~= FIFF.FIFFT_FLOAT;
+    warning(me, 'reading and writing of data in numeric precision ~= float is only supported in FieldTrip and MNE-Python');
+end
+
+switch datatype
+    case FIFF.FIFFT_FLOAT
+        fiff_write_float(fid,FIFF.FIFF_DATA_BUFFER,diag(1./cals)*buf);
+    case FIFF.FIFFT_DOUBLE
+        fiff_write_double(fid,FIFF.FIFF_DATA_BUFFER,diag(1./cals)*buf);
+    case FIFF.FIFFT_COMPLEX_FLOAT
+        fiff_write_complex(fid,FIFF.FIFF_DATA_BUFFER,diag(1./cals)*buf);
+    case FIFF.FIFFT_COMPLEX_DOUBLE
+        fiff_write_double_complex(fid,FIFF.FIFF_DATA_BUFFER,diag(1./cals)*buf);
+    otherwise
+        error(me,'unsupported datatype requested for writing of the buffer');
+end
 
 return;
