+++ external/mne/mne_rt_data_client.m
@@ -31,7 +31,7 @@ classdef mne_rt_data_client < mne_rt_client
             if (nargin < 3)
                 numOfRetries = 20; % set to -1 for infinite
             end
-            obj = obj@mne_rt_client(host, port, numOfRetries); %Superclass call
+            obj = obj@mne_rt_client(host, port, numOfRetries);%Superclass call
             obj.getClientId();
         end % mne_rt_data_client
         
@@ -386,14 +386,14 @@ classdef mne_rt_data_client < mne_rt_client
             
             kind = FIFF.FIFF_MNE_RT_COMMAND;
             type = FIFF.FIFFT_VOID;
-            size = 4+length(data); % first 4 bytes are the command code
+            size = 4+length(data);% first 4 bytes are the command code
             next = 0;
             
             p_DataOutputStream.writeInt(kind);
             p_DataOutputStream.writeInt(type);
             p_DataOutputStream.writeInt(size);
             p_DataOutputStream.writeInt(next);
-            p_DataOutputStream.writeInt(p_Cmd); % first 4 bytes are the command code
+            p_DataOutputStream.writeInt(p_Cmd);% first 4 bytes are the command code
             if(~isempty(data))
                 p_DataOutputStream.writeBytes(data);
             end
@@ -507,17 +507,17 @@ classdef mne_rt_data_client < mne_rt_client
 %                             case FIFF.FIFFT_INT
 %                                 tag.data = zeros(el_size/4, 1);
 %                                 for i = 1:el_size/4
-%                                     tag.data(i) = p_DataInputStream.readInt; %idata = fread(fid,dims(1)*dims(2),'int32=>int32');
+%                                     tag.data(i) = p_DataInputStream.readInt;%idata = fread(fid,dims(1)*dims(2),'int32=>int32');
 %                                 end
 %                             case FIFF.FIFFT_JULIAN
 %                                 tag.data = zeros(el_size/4, 1);
 %                                 for i = 1:el_size/4
-%                                     tag.data(i) = p_DataInputStream.readInt; %idata = fread(fid,dims(1)*dims(2),'int32=>int32');
+%                                     tag.data(i) = p_DataInputStream.readInt;%idata = fread(fid,dims(1)*dims(2),'int32=>int32');
 %                                 end
                             case FIFF.FIFFT_FLOAT
                                 t_MNERTBufferReader = mne_rt_buffer_reader(p_DataInputStream);
                                 tmp = t_MNERTBufferReader.readBuffer(el_size);
-                                tag.data = typecast(tmp, 'single'); %fdata = fread(fid,dims(1)*dims(2),'single=>double');
+                                tag.data = typecast(tmp, 'single');%fdata = fread(fid,dims(1)*dims(2),'single=>double');
                                 tag.data = swapbytes(tag.data);
                             otherwise
                                 error(me,'Cannot handle a matrix of type %d yet',matrix_type)
@@ -546,47 +546,47 @@ classdef mne_rt_data_client < mne_rt_client
                         case FIFF.FIFFT_INT
 %                             tag.data = zeros(tag.size/4, 1);
 %                             for i = 1:tag.size/4
-%                                 tag.data(i) = p_DataInputStream.readInt; %fread(fid,tag.size/4,'int32=>int32');
+%                                 tag.data(i) = p_DataInputStream.readInt;%fread(fid,tag.size/4,'int32=>int32');
 %                             end
                             t_MNERTBufferReader = mne_rt_buffer_reader(p_DataInputStream);
                             tmp = t_MNERTBufferReader.readBuffer(tag.size);
-                            tag.data = typecast(tmp, 'int32'); %fread(fid,tag.size/4,'int32=>int32');
+                            tag.data = typecast(tmp, 'int32');%fread(fid,tag.size/4,'int32=>int32');
[…]
