+++ utilities/private/ft_notification.m
@@ -235,10 +235,8 @@ switch varargin{1}
     varargout{1} = getstate(s, 'last', ident);
     
   case {'clear' '-clear'}
-    default = getstate(s, 'all');
     % reset the notification system
-    % but keep default state
-    s = struct('identifier','all','state',default,'timestamp',NaN);
+    s = [];
     
   case {'query' '-query'}
     if numel(varargin)>1
