### 2023-03-23 c5ac97b
Revert "topoplot3d prevent crash when isolines out of data bounds"
This reverts commit c7405646aaaebb7d0285c6df72db0e9187993e04.
