### 2023-03-23 5998231
Revert "add support for linearly spaced isolines"
This reverts commit fcb61139f693d8f84e7e2e3451506f62eac00470.
