### 2023-03-23 8a8a267
Revert "add ft_artifact_sliding_threshold.m"
This reverts commit 7304977134bbbe0e223fdb7cc0b47b173dfeb28e.
