+++ ft_artifact_sliding_threshold.m
@@ -1,219 +1,0 @@
-function [cfg, artifact,badwin] = ft_artifact_sliding_threshold(cfg, data)
-
-% FT_ARTIFACT_SLIDING_THRESHOLD scans for time points in which the range, 
-% (min-max difference) of the signal in any within a sliding time window
-% exceeds a specified threshold on any channel.
-% The sliding window is defined by its width (in s) and the steps (in s)
-% with which it moves.
-% With the 'trial' method, artifacts are whole trials. With the 'window'
-% method, artifacts are time points corresponding to individual sliding
-% window points.
-%
-% Use as
-%   [cfg, artifact] = ft_artifact_threshold(cfg)
-% with the configuration options
-%   cfg.dataset     = string with the filename
-% or
-%   cfg.headerfile  = string with the filename
-%   cfg.datafile    = string with the filename
-%
-% Alternatively you can use it as
-%   [cfg, artifact] = ft_artifact_threshold(cfg, data)
-%
-% In both cases the configuration should also contain
-%   cfg.trl        = structure that defines the data segments of interest, see FT_DEFINETRIAL
-%   cfg.continuous = 'yes' or 'no' whether the file contains continuous data
-%
-% The following configuration options can be specified
-%   cfg.artfctdef.slidethresh.method    = 'trial' or 'window'
-%   cfg.artfctdef.slidethresh.channel   = cell-array with channel labels
-%   cfg.artfctdef.slidethresh.bpfilter  = 'no' or 'yes'
-%   cfg.artfctdef.slidethresh.bpfreq    = [0.3 30]
-%   cfg.artfctdef.slidethresh.bpfiltord = 4
-%
-% The detection of artifacts is done according to the following settings.
-%   cfg.artfctdef.slidethresh.range     = value in uV/T, default  inf
-%   cfg.artfctdef.slidethresh.stdrange  = value in stdev, default  inf
-%   cfg.artfctdef.slidethresh.winsize   = value in s, default = 0.1
-%   cfg.artfctdef.slidethresh.winstep   = value in s, default = 0.05
-%
-%
-% To facilitate data-handling and distributed computing you can use
-%   cfg.inputfile   =  ...
-% If you specify one of these (or both) the input data will be read from a *.mat
-% file on disk and/or the output data will be written to a *.mat file. These mat
-% files should contain only a single variable, corresponding with the
-% input/output structure.
-%
-% See also FT_REJECTARTIFACT, FT_ARTIFACT_CLIP, FT_ARTIFACT_ECG, FT_ARTIFACT_EOG,
-% FT_ARTIFACT_JUMP, FT_ARTIFACT_MUSCLE, FT_ARTIFACT_THRESHOLD, FT_ARTIFACT_ZVALUE
-
-% Copyright (C) 2003-2011, Robert Oostenveld, SMI, FCDC
-%
-% This file is part of FieldTrip, see http://www.fieldtriptoolbox.org
-% for the documentation and details.
-%
-%    FieldTrip is free software: you can redistribute it and/or modify
-%    it under the terms of the GNU General Public License as published by
-%    the Free Software Foundation, either version 3 of the License, or
-%    (at your option) any later version.
-%
-%    FieldTrip is distributed in the hope that it will be useful,
-%    but WITHOUT ANY WARRANTY; without even the implied warranty of
[…]
