+++ test/test_tutorial_natmeg_preprocessing.m
@@ -12,7 +12,7 @@ function test_tutorial_natmeg_preprocessing
 clear all
 close all
 
-cd(dccnpath('/home/common/matlab/fieldtrip/data/ftp/workshop/natmeg'));
+cd(dccnpath('/home/common/matlab/fieldtrip/data/ftp/workshop/natmeg2014'));
 
 cfg = [];
 cfg.dataset    = 'oddball1_mc_downsampled.fif';
