### 2023-03-27 c2039be
FIX #2221 - 3 bytes per sample for BDF, thanks to @vlitvak for reporting
