+++ fileio/ft_write_mri.m
@@ -33,8 +33,9 @@ function [V] = ft_write_mri(filename, dat, varargin)
 %   'vista'       SIMBIO specific format
 %   'vmr'         Brainvoyager specific format
 %   'vmp'         Brainvoyager specific format
+%   'vtk'         Visualization ToolKit file format, for use with Paraview
 %
-% See also FT_READ_MRI, FT_WRITE_DATA, FT_WRITE_HEADSHAPE
+% See also FT_READ_MRI, FT_DATATYPE_VOLUME, FT_WRITE_DATA, FT_WRITE_HEADSHAPE, FT_WRITE_SENS
 
 % Copyright (C) 2011-2021, Jan-Mathijs Schoffelen
 %
@@ -70,7 +71,7 @@ if ~ft_hastoolbox('spm') && isempty(spmversion)
   spmversion = 'spm12';
 end
 
-%% ensure that the input data is consistent
+% ensure that the input data is consistent
 if isnumeric(dat) || islogical(dat)
   % convert the data to a structure according to FT_DATATYPE_VOLUME and FT_READ_MRI
   mri.anatomy = dat;
@@ -117,7 +118,7 @@ dat         = mri.anatomy;
 transform   = mri.transform;
 clear mri % to avoid confusion
 
-%% determine the details of the output format
+% determine the details of the output format
 if isempty(dataformat)
   % only do the autodetection if the format was not specified
   dataformat = ft_filetype(filename);
@@ -152,28 +153,28 @@ if ~ismember(dataformat, {'empty', 'fcdc_global', 'fcdc_buffer', 'fcdc_mysql'})
   isdir_or_mkdir(fileparts(filename));
 end
 
-%% write the data
+% write the data
 switch dataformat
-  
+
   case {'vmr' 'vmp'}
-    %% write to brainvoyager format
+    % write to brainvoyager format
     vmpversion = ft_getopt(varargin, 'vmpversion', 2);
     write_brainvoyager(filename, dat, dataformat, vmpversion);
-    
+
   case {'analyze_old'}
-    %% write to Analyze format, using old code and functions from Darren Webbers toolbox
+    % write to Analyze format, using old code and functions from Darren Webbers toolbox
     avw = avw_hdr_make;
-    
+
     % specify the image data and dimensions
     avw.hdr.dime.dim(2:4) = size(dat);
     avw.img = dat;
-    
+
     % orientation 0 means transverse unflipped (axial, radiological)
     % X direction first,  progressing from patient right to left,
     % Y direction second, progressing from patient posterior to anterior,
     % Z direction third,  progressing from patient inferior to superior.
     avw.hdr.hist.orient = 0;
-    
+
     % specify voxel size
     avw.hdr.dime.pixdim(2:4) = [1 1 1];
     % FIXME, this currently does not work due to all flipping and permuting
@@ -181,7 +182,7 @@ switch dataformat
     % resy = y(2)-y(1);
     % resz = z(2)-z(1);
     % avw.hdr.dime.pixdim(2:4) = [resy resx resz];
-    
+
     % specify the data type
     switch class(dat)
       case 'logical'
@@ -205,27 +206,27 @@ switch dataformat
       otherwise
         ft_error('unsupported datatype "%s" to write to analyze_old format', class(dat));
     end
-    
+
     % write the header and image data
     avw_img_write(avw, filename, [], 'ieee-le');
-    
+
   case {'analyze_img' 'analyze_hdr' 'analyze' 'nifti_img' 'nifti_spm'}
     %% analyze or nifti data, using SPM
[…]
