### 2023-04-01 492b8ae
ENH - allow data to be written to vtk format files
