+++ fileio/private/write_vtk.m
@@ -1,50 +1,158 @@
-function write_vtk(fn, pnt, tri)
+function write_vtk(fn, pos, tri, val)
 
-% WRITE_VTK writes a triangulation to a VTK (Visualisation ToolKit) format file.
+% WRITE_VTK writes a mesh to a VTK (Visualisation ToolKit) format file.
 % Supported are triangles, tetraheders and hexaheders.
 %
 % Use as
-%   write_vtk(filename, pnt, tri)
+%   write_vtk(filename, pos, tri, val)
+%   write_vtk(filename, pos, tet, val)
+%   write_vtk(filename, pos, hex, val)
+% where pos describes the vertex positions and tri/tet/hex describe the connectivity
+% of the surface or volume elements. 
+% 
+% The optional val argument can be used to write scalar or vector values for
+% each vertex or element.
 %
 % See also READ_VTK, WRITE_PLY
 
-% Copyright (C) 2002, Robert Oostenveld
+% Copyright (C) 2002-2023, Robert Oostenveld
+%
+% This file is part of FieldTrip, see http://www.fieldtriptoolbox.org
+% for the documentation and details.
+%
+%    FieldTrip is free software: you can redistribute it and/or modify
+%    it under the terms of the GNU General Public License as published by
+%    the Free Software Foundation, either version 3 of the License, or
+%    (at your option) any later version.
+%
+%    FieldTrip is distributed in the hope that it will be useful,
+%    but WITHOUT ANY WARRANTY; without even the implied warranty of
+%    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+%    GNU General Public License for more details.
+%
+%    You should have received a copy of the GNU General Public License
+%    along with FieldTrip. If not, see <http://www.gnu.org/licenses/>.
 %
 % $Id$
 
-fid = fopen_or_error(fn, 'wt');
+if nargin<4
+  val = [];
+end
 
-npnt = size(pnt,1);
-ntri = size(tri,1);
+npos = size(pos,1);
+ntri = size(tri,1);   % this can be triangles, tetraheders or hexaheders
+
+if ~isempty(val)
+  assert(size(val,1)==npos || size(val,1)==ntri, 'the values are inconsistent with the mesh dimensions');
+end
+
+% see https://kitware.github.io/vtk-examples/site/Cxx/GeometricObjects/LinearCellDemo/
+VTK_VERTEX           = 1;
+VTK_POLY_VERTEX      = 2;
+VTK_LINE             = 3;
+VTK_POLY_LINE        = 4;
+VTK_TRIANGLE         = 5;
+VTK_TRIANGLE_STRIP   = 6;
+VTK_POLYGON          = 7;
+VTK_PIXEL            = 8;
+VTK_QUAD             = 9;
+VTK_TETRA            = 10;
+VTK_VOXEL            = 11;
+VTK_HEXAHEDRON       = 12;
+VTK_WEDGE            = 13;
+VTK_PYRAMID          = 14;
+VTK_PENTAGONAL_PRISM = 15;
+VTK_HEXAGONAL_PRISM  = 16;
+
+fid = fopen_or_error(fn, 'wt');
 
 % write the header
[…]
