### 2023-04-02 b2c00d4
ENH - add/synchronize some shared private functions
