### 2023-04-02 f429225
ENH - allow detailed specification of font for vertex or face indices
