### 2023-04-02 784410e
ENH - improve plotting of intersecting meshes, also support tetrahedral and hexahedral
