### 2023-04-03 8fb3dfe
ENH - add support for seg3d_mat
