### 2023-04-03 cbd2259
updated defaults for FEM meshes
