### 2023-04-03 0d26d45
changed default queryrange from 3 to 1
- do not redraw if user clicks outside of the axes
- fixed bug in transparency of functional data in absence of mask
