### 2023-04-03 0beabe6
ENH - improvements for basedoncentroids and movetocentroids
This resulted from a detailed evaluation using the updated headmodel_eeg_fem tutorial.

- speed up the movetocentroids processing using knnsearch
- improve the inside detection when the input is a labeled mesh
- better defaults for FEM meshes
