+++ ft_prepare_sourcemodel.m
@@ -12,16 +12,16 @@ function [sourcemodel, cfg] = ft_prepare_sourcemodel(cfg)
 % The different approaches for constructing a source model are
 %   cfg.method = 'basedongrid'        regular 3D grid with explicit specification
 %                'basedonpos'         regular 3D grid with specification of the resolution
-%                'basedonshape'       surface mesh based on inward shifted head surface from external file
 %                'basedonmri'         regular 3D grid, based on segmented MRI, restricted to gray matter
 %                'basedonmni'         regular 3D grid, based on a warped template grid, based on the MNI brain
 %                'basedoncortex'      cortical sheet from external software such as Caret or FreeSurfer, can also be two separate hemispheres
 %                'basedonresolution'  regular 3D grid with specification of the resolution
+%                'basedonshape'       surface mesh based on inward shifted head surface from an external file
 %                'basedonvol'         surface mesh based on inward shifted brain surface from volume conductor
 %                'basedonfile'        the sourcemodel should be read from file
 %                'basedoncentroids'   irregular 3D grid based on volumetric mesh
-% The default for cfg.method is to determine the approach automatically, based on
-% the configuration options that you specify.
+% The default method is determined automatically based on the configuration options
+% that you specify.
 %
 % BASEDONRESOLUTION - uses an explicitly specified grid, or with the desired
 % resolution, according to the following configuration options:
@@ -65,7 +65,7 @@ function [sourcemodel, cfg] = ft_prepare_sourcemodel(cfg)
 %   cfg.headshape       = string, should be a *.fif file
 %
 % BASEDONCENTROIDS - places sources on the centroids of a volumetric mesh
-%   cfg.headmodel       = volumetric mesh
+%   cfg.headmodel       = tetrahedral or hexahedral mesh
 %   cfg.headmodel.type  = 'simbio';
 %
 % Other configuration options include
@@ -108,7 +108,7 @@ function [sourcemodel, cfg] = ft_prepare_sourcemodel(cfg)
 % See also FT_PREPARE_LEADFIELD, FT_PREPARE_HEADMODEL, FT_SOURCEANALYSIS,
 % FT_DIPOLEFITTING, FT_MEGREALIGN
 
-% Copyright (C) 2004-2021, Robert Oostenveld
+% Copyright (C) 2004-2023, Robert Oostenveld
 %
 % This file is part of FieldTrip, see http://www.fieldtriptoolbox.org
 % for the documentation and details.
@@ -174,6 +174,7 @@ cfg.sourcemodel       = ft_getopt(cfg, 'sourcemodel');
 cfg.unit              = ft_getopt(cfg, 'unit');
 cfg.method            = ft_getopt(cfg, 'method'); % the default is to do automatic detection further down
 cfg.movetocentroids   = ft_getopt(cfg, 'movetocentroids', 'no');
+cfg.feedback          = ft_getopt(cfg, 'feedback', 'text');
 
 % this option was deprecated on 12 Aug 2020
 if isfield(cfg, 'warpmni')
@@ -242,50 +243,55 @@ switch cfg.method
     fprintf('reading sourcemodel from file\n');
     cfg.tight       = ft_getopt(cfg, 'tight',   'no');
[…]
