### 2023-04-03 2a2645b
ENH - more efficient handling of epoched fif-files
