### 2023-04-03 e3495e3
ENH - some code reorganization
