+++ external/mne/fiff_read_epochs.m
@@ -71,10 +71,10 @@ for k = 1:ep.nent
         case FIFF.FIFF_MNE_BASELINE_MAX
             tag = fiff_read_tag(fid,pos);
             bmax = tag.data;
-        case FIFF.FIFFB_MNE_EPOCHS_SELECTION
+        case FIFF.FIFF_MNE_EPOCHS_SELECTION
             tag = fiff_read_tag(fid,pos);
             selection = tag.data;
-        case FIFF.FIFFB_MNE_EPOCHS_DROP_LOG
+        case FIFF.FIFF_MNE_EPOCHS_DROP_LOG
             tag = fiff_read_tag(fid,pos);
             drop_log = tag.data;
     end
