### 2023-04-05 9ebeb92
FIX - reimplementation of coilaccuracy handling, to get the refs right
