+++ fileio/private/ctf2grad.m
@@ -69,11 +69,12 @@ elec = [];
 % ADC16   is 18
 % V0      is 15
 
-if ~isempty(coilaccuracy)
+if isfield(hdr, 'res4') && ~isempty(coilaccuracy)
   %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-  % use the coil definitions from the MNE coil_def.dat file
-  % these allow for varying accuracy which is specified by
-  % coilaccuracy = 0, 1 or 2
+  % the header was read using the CTF p-files, i.e. readCTFds
+  %
+  % use in additionthe coil definitions from the MNE coil_def.dat file
+  % these allow for varying accuracy which is specified by coilaccuracy = 0, 1 or 2
   %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
   
   ft_hastoolbox('mne', 1);
@@ -82,70 +83,91 @@ if ~isempty(coilaccuracy)
   
   k = 1;
   for i=1:length(hdr.res4.senres)
-    switch hdr.res4.senres(i).sensorTypeIndex
+    
+    thissens = hdr.res4.senres(i);
+    if thissens.numCoils==2
+      % line between the coils in dewar space, which  is needed to determine the type if it's a reference gradiometer
+      delta    = thissens.pos0*[-1;1]; 
+    end
+
+    switch thissens.sensorTypeIndex
       case 5 % 5001
         thisdef = def([def.id]==5001 & [def.accuracy]==coilaccuracy);
       case 0 % 5002
         thisdef = def([def.id]==5002 & [def.accuracy]==coilaccuracy);
-      case 1 % 5003
-        thisdef = def([def.id]==5003 & [def.accuracy]==coilaccuracy);
+      case 1 % 5003 or 5004
+        % this is a reference gradiometer
+        
+        % determine whether the gradient is 'off' diagonal, comparing the
+        % (cosine) of the angle between the line connecting the coils and
+        % the orientation of the first coil        % 
+        if abs((delta./norm(delta))'*thissens.ori0(:,1))<1e-3
+          % it's an off diagonal channel
+          thisdef = def([def.id]==5004 & [def.accuracy]==coilaccuracy);
+        else
+          thisdef = def([def.id]==5003 & [def.accuracy]==coilaccuracy);
+        end
       otherwise
         % do not add this as sensor to the gradiometer definition
         thisdef = [];
     end % case
     
     if ~isempty(thisdef)
-      % the sensors (i.e. the combination of coils that comprises a channel) is
-      % defined at [0 0 0] and with the direction [0 0 1]
-      pos0 = [0 0 0];
-      ori0 = [0 0 1];
-      
       if dewar
-        % convert from cm to m
-        pos2 = hdr.res4.senres(i).pos0(:,1)' / 100;
-        if hdr.res4.senres(i).numCoils==2
-          % determine the direction from the position of the two coils
-          ori2 = (hdr.res4.senres(i).pos0(:,2) - hdr.res4.senres(i).pos0(:,1))';
-        else
-          % determine the direction from the orientation of the coil
-          ori2 = hdr.res4.senres(i).ori0(:,1);
-        end
-      else
-        % convert from cm to m
-        pos2 = hdr.res4.senres(i).pos(:,1)' / 100;
-        if hdr.res4.senres(i).numCoils==2
-          % determine the direction from the position of the two coils
[…]
