### 2023-04-05 78c8f56
ENH - some improvements
