### 2023-04-06 4e71b01
ENH - allow for custom coil_def.dat file, cosmetic cleanup mne2grad
