### 2023-04-06 28c724a
ENH - some improvements for grad arrays for non elekta systems
