### 2023-04-07 40e16d0
ENH - improve display of logical scalars
