### 2023-04-11 ebef266
add detection of ascii vtk format
