### 2023-04-11 92e8b01
FIX - the coordsyslabel_x/y/z was not removed properly when zoomed-in
