### 2023-04-11 fe2e2fa
ENH - plot each of the fields in the input struct mri
it used to only plot the anatomy, now it loops over all fields that match dim
