+++ plotting/ft_plot_ortho.m
@@ -56,18 +56,27 @@ function [hx, hy, hz] = ft_plot_ortho(dat, varargin)
 %
 % $Id$
 
-if isstruct(dat) && isfield(dat, 'anatomy') && isfield(dat, 'transform')
-  % the input is an MRI structure, call this function recursively
-  varargin = ft_setopt(varargin, 'transform', dat.transform);
-  if isfield(dat, 'coordsys')
-    varargin = ft_setopt(varargin, 'coordsys', dat.coordsys);
+if isstruct(dat)
+  if isfield(dat, 'transform') && isfield(dat, 'dim')
+    % the input is an MRI structure, call this function recursively
+    varargin = ft_setopt(varargin, 'transform', dat.transform);
+    if isfield(dat, 'coordsys')
+      varargin = ft_setopt(varargin, 'coordsys', dat.coordsys);
+    end
+    if isfield(dat, 'unit')
+      varargin = ft_setopt(varargin, 'unit', dat.unit);
+    end
+    fn = fieldnames(dat);
+    for i=1:numel(fn)
+      if isequal(size(dat.(fn{i})), dat.dim)
+        ft_info('plotting %s', fn{i});
+        [hx, hy, hz] = ft_plot_ortho(dat.(fn{i}), varargin{:});
+      end % if 
+    end % for 
+    return
+  else
+    ft_error('unsupported input structure');
   end
-  if isfield(dat, 'unit')
-    varargin = ft_setopt(varargin, 'unit', dat.unit);
-  end
-  dat = dat.anatomy;
-  [hx, hy, hz] = ft_plot_ortho(dat, varargin{:});
-  return
 end
 
 % parse first input argument(s), it is either
