### 2023-04-12 ffaeee2
FIX #1834 - added gradiometer templates
- these are all expressed in meter, without balancing, with lowest resolution (single integration point per coil), and have been manually aligned with the head surface of Subject01
