+++ plotting/private/setviewpoint.m
@@ -45,7 +45,7 @@ switch lower(coordsys)
       case 'posterior'
         view(ax, [0 -1 0]);
     end % switch viewpoint
-  case {'als' 'ctf' '4d' 'bti' 'eeglab'}
+  case {'als' 'ctf' '4d' 'bti' 'yokogawa' 'eeglab'}
     switch viewpoint
       case 'superior'
         view(ax, [-90 90]); % not exactly the same as [0 0 1], this causes the nose pointing up
