### 2023-04-12 998dea9
ENH - exclude the area in front of the face
it now detects triangles with edges that are too long; this prevents the from of MEG helmets to be closed/interpolated
