+++ plotting/ft_plot_topo3d.m
@@ -1,7 +1,7 @@
 function ft_plot_topo3d(pos, val, varargin)
 
 % FT_PLOT_TOPO3D visualizes a 3D topographic representation of the electric potential
-% or magnetic field distribution at the sensor locations
+% or magnetic field distribution at the sensor locations.
 %
 % Use as
 %   ft_plot_topo3d(pos, val, ...)
@@ -9,15 +9,16 @@ function ft_plot_topo3d(pos, val, varargin)
 % given as Nx1 vector.
 %
 % Optional input arguments should be specified in key-value pairs and can include
-%   'contourstyle' = string, 'none', 'black', 'color' (default = 'none')
-%   'isolines'     = vector with values at which to draw isocontours, or 'auto' (default = 'auto')
-%   'facealpha'    = scalar, between 0 and 1 (default = 1)
-%   'refine'       = scalar, number of refinement steps for the triangulation, to get a smoother interpolation (default = 0)
-%   'unit'         = string, 'm', 'cm' or 'mm' (default = 'cm')
-%   'coordsys'     = string, assume the data to be in the specified coordinate system (default = 'unknown')
-%   'axes'         = boolean, whether to plot the axes of the 3D coordinate system (default = false)
+%   'contourstyle'  = string, 'none', 'black', 'color' (default = 'none')
+%   'isolines'      = vector with values at which to draw isocontours, or 'auto' (default = 'auto')
+%   'facealpha'     = scalar, between 0 and 1 (default = 1)
+%   'refine'        = scalar, number of refinement steps for the triangulation, to get a smoother interpolation (default = 0)
+%   'neighbourdist' = number, maximum distance between neighbouring sensors (default is automatic)
+%   'unit'          = string, 'm', 'cm' or 'mm' (default = 'cm')
+%   'coordsys'      = string, assume the data to be in the specified coordinate system (default = 'unknown')
+%   'axes'          = boolean, whether to plot the axes of the 3D coordinate system (default = false)
 %
-% See also FT_PLOT_SENS, FT_PLOT_HEADMODEL, FT_PLOT_HEADSHAPE, FT_PLOT_TOPO,
+% See also FT_PLOT_TOPO, FT_PLOT_SENS, FT_PLOT_MESH, FT_PLOT_HEADSHAPE,
 % FT_TOPOPLOTER, FT_TOPOPLOTTFR
 
 % Copyright (C) 2009-2023, Robert Oostenveld
@@ -42,13 +43,14 @@ function ft_plot_topo3d(pos, val, varargin)
 
 % get the optional input arguments
 contourstyle  = ft_getopt(varargin, 'contourstyle', 'none');
-nrefine       = ft_getopt(varargin, 'refine', 0);
+refine_       = ft_getopt(varargin, 'refine', 0);           % do not confuse with the REFINE function
+neighbourdist = ft_getopt(varargin, 'neighbourdist');
 isolines      = ft_getopt(varargin, 'isolines', 'auto');
 topostyle     = ft_getopt(varargin, 'topostyle', 'color');  % FIXME what is the purpose of this option?
 facealpha     = ft_getopt(varargin, 'facealpha', 1);
 unit          = ft_getopt(varargin, 'unit', 'cm');
 coordsys      = ft_getopt(varargin, 'coordsys');
-axes_         = ft_getopt(varargin, 'axes',       false); % do not confuse with built-in function
+axes_         = ft_getopt(varargin, 'axes', false);         % do not confuse with built-in function
 
[…]
