### 2023-04-12 21b01e0
ENH - more robust chanori determination for CTF reference grads
