### 2023-04-14 4d9c649
ENH - various improvements to coil handling
