### 2023-04-14 e7e1d4d
FIX - the order of desired datatypes matters, put 'comp' first
- this avoids the comp information from being removed
- this was detected as regression error in test_ft_sourceanalysis.m
