### 2023-04-14 44b19e4
ENH - align automatic grid nicely around [0 0 0]
- it used to be the case that the grid started at `minpos`, which is a rather random number. It is better to have it centered on, or aligned to the origin
