### 2023-04-14 f9241ee
ENH - lower-level code udpates and new function to write mne epochs
