+++ external/mne/fiff_write_epochs.m
@@ -1,0 +1,211 @@
+function fiff_write_epochs(name,data)
+%
+% function fiff_write_epochs(name,data)
+%
+% name     filename
+% data     the data structure returned from fiff_write_evoked
+%
+%
+
+% This function has been adjusted by Jan-Mathijs Schoffelen, from the 
+% fiff_write_evoked function:
+%
+%   Author : Matti Hamalainen, MGH Martinos Center
+%   License : BSD 3-clause
+%
+% Revision 1.14  2009/03/31 01:12:30  msh
+% Improved ID handling
+%
+% Revision 1.13  2009/03/30 11:37:37  msh
+% Added copying of measurement info blocks from the original like in mne_browse_raw
+%
+% Revision 1.12  2008/03/13 19:18:07  msh
+% Read and write FIFF_MEAS_DATE from/to FIFFB_MEAS_INFO as appropriate
+%
+% Revision 1.11  2006/10/04 20:12:37  msh
+% Added fiff_read_evoked_all
+% Modified fiff_pick_channels_evoked to handle multiple data sets
+% Added bad channel writing to fiff_write_evoked
+%
+% Revision 1.10  2006/09/26 23:20:23  msh
+% Moved block ID and parent block id to the meas block.
+%
+% Revision 1.9  2006/06/14 05:32:43  msh
+% Added missing FIFF_NAVE to each aspect written.
+%
+% Revision 1.8  2006/05/29 16:57:32  msh
+% Make sure that scan numbers are correct.
+%
+% Revision 1.7  2006/04/23 15:29:40  msh
+% Added MGH to the copyright
+%
+% Revision 1.6  2006/04/18 20:44:46  msh
+% Added reading of forward solution.
+% Use length instead of size when appropriate
+%
+% Revision 1.5  2006/04/14 15:49:49  msh
+% Improved the channel selection code and added ch_names to measurement info.
+%
+% Revision 1.4  2006/04/14 11:03:57  msh
+% Changed fiff_write_id write a given id.
+% Added parent id writing.
+%
+% Revision 1.3  2006/04/13 23:09:46  msh
+% Further streamlining of the coordinate transformations.
+%
+% Revision 1.2  2006/04/12 13:13:51  msh
+% Added fiff_find_evoked.m
+% Use aspect_kind field name instead of aspect_type
+%
+% Revision 1.1  2006/04/12 10:29:02  msh
+% Made evoked data writing compatible with the structures returned in reading.
+%
+%
+me='MNE:mne_write_epochs';
+if nargin ~= 2
+    error(me,'File name and data required as an arguments');
+end
+%
+FIFF = fiff_define_constants();
+%
+%  Create the file and save the essentials
+%
+fid = fiff_start_file(name);
+fiff_start_block(fid,FIFF.FIFFB_MEAS);
+fiff_write_id(fid,FIFF.FIFF_BLOCK_ID);
+if ~isempty(data.info.meas_id)
+    fiff_write_id(fid,FIFF.FIFF_PARENT_BLOCK_ID,data.info.meas_id);
+end
+%
+%    Measurement info
+%
+fiff_start_block(fid,FIFF.FIFFB_MEAS_INFO);
+%
+%    Blocks from the original
+%
[…]
