### 2023-04-14 ca34f9b
ENH - updates to improve handling of epoched fif-files
