### 2023-04-17 adca51b
ENH - added OPM magnetometers as known types
