### 2023-04-17 1f27d61
ENH - support for more efficient matrix writing and epochs support
