+++ external/mne/fiff_write_double_complex_matrix.m
@@ -23,7 +23,12 @@ function fiff_write_double_complex_matrix(fid,kind,mat)
 me='MNE:fiff_write_double_complex_matrix';
 
 if nargin ~= 3
-        error(me,'Incorrect number of arguments');
+   error(me,'Incorrect number of arguments');
+end
+
+ndim = ndims(mat);
+if ndim < 2 || ndim > 3
+   error(me,'Input should be a two-dimensional or three-dimensional matrix');
 end
 
 FIFFT_COMPLEX_DOUBLE = 21;
@@ -31,7 +36,7 @@ FIFFT_MATRIX  = bitshift(1,30);
 FIFFT_MATRIX_COMPLEX_DOUBLE = bitor(FIFFT_COMPLEX_DOUBLE,FIFFT_MATRIX);
 FIFFV_NEXT_SEQ=0;
 
-datasize = 2*8*numel(mat) + 4*3;
+datasize = 2*8*numel(mat) + 4*(ndim+1);
 
 count = fwrite(fid,int32(kind),'int32');
 if count ~= 1
@@ -49,26 +54,45 @@ count = fwrite(fid,int32(FIFFV_NEXT_SEQ),'int32');
 if count ~= 1
    error(me,'write failed');
 end
-nrow = size(mat,1);
-ncol = size(mat,2);
-for j = 1:nrow
-   for k = 1:ncol
-      count = fwrite(fid,real(mat(j,k)),'double');
-      if count ~= 1
-          error(me,'write failed');
-      end
-      count = fwrite(fid,imag(mat(j,k)),'double');
-      if count ~= 1
-          error(me,'write failed');
-      end
-   end
-end
-dims(1) = size(mat,2);
-dims(2) = size(mat,1);
-dims(3) = 2;
-count = fwrite(fid,int32(dims),'int32');
-if count ~= 3
+% nrow = size(mat,1);
+% ncol = size(mat,2);
+% for j = 1:nrow
+%    for k = 1:ncol
+%       count = fwrite(fid,real(mat(j,k)),'double');
+%       if count ~= 1
+%           error(me,'write failed');
+%       end
+%       count = fwrite(fid,imag(mat(j,k)),'double');
+%       if count ~= 1
+%           error(me,'write failed');
+%       end
+%    end
+% end
+
+% the matrix is written last dimension first, and then each element's real part followed by its imaginary part
+mat = permute(mat,ndim:-1:1);
+mat = [real(mat(:)) imag(mat(:))]';
+count = fwrite(fid,double(mat(:)),'double');
+if count ~= numel(mat)
     error(me,'write failed');
+end
+if ndim==2
+    dims(1) = size(mat,2);
+    dims(2) = size(mat,1);
+    dims(3) = 2;
+        count = fwrite(fid,int32(dims),'int32');
+    if count ~= 3
+        error(me,'write failed');
+    end
+elseif ndim==3
+    dims(1) = size(mat,3);
+    dims(2) = size(mat,2);
+    dims(3) = size(mat,1);
+    dims(4) = 3;
+    count = fwrite(fid,int32(dims),'int32');
+    if count ~= 4
+        error(me,'write failed');
+    end
 end
 
 return;
