### 2023-04-17 dcca25d
ENH - improved support for epoch data
