+++ fieldtrip2fiff.m
@@ -134,36 +134,40 @@ if israw
     event = [];
   end
   if ~isempty(event)
-    eve = convertevent(event);
-    mne_write_events(eventfile, eve);
+    [eventlist, mappings] = convertevent(event);
+    fiff_write_events(eventfile, eventlist, mappings);
     ft_info('Writing events to %s\n', eventfile)
   end
   
 elseif isepch
   
+  nsmp   = numel(data.time);
+  ntrl   = size(data.trial,1);    
   if isfield(data, 'trialinfo')
     ft_warning('Using the first column of the trialinfo field as event values');
-    events = [data.sampleinfo(:,1) zeros(size(data.trial,1),1) data.trialinfo(:,1)];
-    
-    vals = unique(data.trialinfo(:,1));
-    vals = [(1:numel(vals))' vals]';
-    eventid = sprintf('event%d:%d,',vals(:));
-    eventid = eventid(1:end-1); % remove the last comma
+    trg = data.trialinfo(:,1);
+  else
+    trg = ones(ntrl,1);
   end
+  events = [((0:nsmp:(nsmp*(ntrl-1)))+1)' zeros(ntrl,1) trg];
+  vals = unique(trg);
+  vals = [(1:numel(vals))' vals]';
+  eventid = sprintf('event%d: %d;',vals(:));
+  eventid = eventid(1:end-1); % remove the last comma
 
-  epochs.epoch = data.trial; 
+  epochs.epoch = data.trial;
   epochs.tmin  = data.time(1).*info.sfreq;
   epochs.tmax  = data.time(end).*info.sfreq;
-  epochs.baseline = [nan nan];
+  epochs.baseline  = [];
   epochs.selection = (1:size(data.trial,1))'-1; % seems 0-based
-  epochs.drop_log  = ' ';
+  epochs.drop_log  = ['[',repmat('[], ',[1 size(data.trial,1)-1]), '[]]'];
   epochs.events    = events;
   epochs.event_id  = eventid;
   epochs.comment   = ' ';
 
   fiffdata.info  = info;
   fiffdata.epoch = epochs;
-  fiff_write_epochs(fifffile, fiffdata);
+  fiff_write_epochs(fifffile, fiffdata, dtype);
 
 elseif istlck
   evoked.aspect_kind = 100;
@@ -296,64 +300,55 @@ else
 end
 
 
-function eve = convertevent(event)
-% tentative code, with lots of assumption
+function [eventlist, mappings] = convertevent(event)
 
-%CTF should use backpanel trigger
-backpanel = strcmp({event.type}, 'backpanel trigger');
-if any(backpanel)
-  fprintf('Writing the value of the backpanel trigger into the event file\n')
-  trigger = [event(backpanel).value];
-  
-  eve = zeros(numel(trigger), 3);
-  eve(:,1) = [event(backpanel).sample];
-  eve(:,3) = [event(backpanel).value];
-  return
-end
+ev_type  = {event.type};
+ev_utype = unique(ev_type);
+ft_info('Detected %d events, of %d different event types\n', numel(ev_type), numel(ev_utype));
 
-% use ev_type and ev_value
-ev_type = unique({event.type});
-% convert to cell of strings
-if any(cellfun(@isnumeric, {event.value}))
-  event_value = cellfun(@num2str, {event.value}, 'uni', false);
-else
-  event_value = {event.value};
-end
-ev_value = unique(event_value);
-
-eve = zeros(numel(event), 3);
-
-for i1 = 1:numel(ev_type)
-  for i2 = 1:numel(ev_value)
-    i_type = strcmp({event.type}, ev_type{i1});
[…]
