+++ external/dmlt/external/gpstuff/SuiteSparse/CHOLMOD/Tcov/ctest.c
@@ -1,303 +1,0 @@
-/* ========================================================================== */
-/* === Tcov/ctest =========================================================== */
-/* ========================================================================== */
-
-/* -----------------------------------------------------------------------------
- * CHOLMOD/Tcov Module.  Copyright (C) 2005-2006, Timothy A. Davis
- * The CHOLMOD/Tcov Module is licensed under Version 2.0 of the GNU
- * General Public License.  See gpl.txt for a text of the license.
- * CHOLMOD is also available under other licenses; contact authors for details.
- * http://www.cise.ufl.edu/research/sparse
- * -------------------------------------------------------------------------- */
-
-/* Test for colamd v2.4 */
-
-#include "cm.h"
-#include "colamd.h"
-
-
-/* ========================================================================== */
-/* === ctest ================================================================ */
-/* ========================================================================== */
-
-void ctest (cholmod_sparse *A)
-{
-    double knobs [COLAMD_KNOBS], knobs2 [COLAMD_KNOBS] ;
-    Int *P, *Cp, *Ci, *Si, *Sp ;
-    cholmod_sparse *C, *A2, *B, *S, *BT ;
-    Int nrow, ncol, alen, ok, stats [COLAMD_STATS], i, p, trial ;
-    size_t s ;
-
-    /* ---------------------------------------------------------------------- */
-    /* get inputs */
-    /* ---------------------------------------------------------------------- */
-
-    printf ("\nCOLAMD test\n") ;
-
-    if (A == NULL)
-    {
-	return ;
-    }
-
-    if (A->stype)
-    {
-	A2 = CHOLMOD(copy) (A, 0, 0, cm) ;
-	B = A2 ;
-    }
-    else
-    {
-	A2 = NULL ;
-	B = A ;
-    }
-
-    nrow = B->nrow ;
-    ncol = B->ncol ;
-    S = NULL ;
-
-    /* ---------------------------------------------------------------------- */
-    /* allocate workspace colamd */
-    /* ---------------------------------------------------------------------- */
-
-    P = CHOLMOD(malloc) (nrow+1, sizeof (Int), cm) ;
-
-    COLAMD_set_defaults (knobs) ;
-    COLAMD_set_defaults (knobs2) ;
-    COLAMD_set_defaults (NULL) ;
-    COLAMD_report (NULL) ;
-    SYMAMD_report (NULL) ;
-
-    alen = COLAMD_recommended (B->nzmax, ncol, nrow) ;
-    C = CHOLMOD(allocate_sparse) (ncol, nrow, alen, TRUE, TRUE, 0,
-	    CHOLMOD_PATTERN, cm) ;
-    Cp = C->p ;
-    Ci = C->i ;
-
-    /* ---------------------------------------------------------------------- */
-    /* order with colamd */
[…]
