+++ external/dmlt/external/gpstuff/gp/gp_trcov.m
@@ -1,171 +1,0 @@
-function [K, C] = gp_trcov(gp, x1, predcf)
-%GP_TRCOV  Evaluate training covariance matrix (gp_cov + noise covariance). 
-%
-%  Description
-%    K = GP_TRCOV(GP, TX, PREDCF) takes in Gaussian process GP and
-%    matrix TX that contains training input vectors to GP. Returns
-%    (noiseless) covariance matrix K for latent values, which is
-%    formed as a sum of the covariance matrices from covariance
-%    functions in gp.cf array. Every element ij of K contains
-%    covariance between inputs i and j in TX. PREDCF is an array
-%    specifying the indexes of covariance functions, which are used
-%    for forming the matrix. If not given, the matrix is formed
-%    with all functions.
-%
-%    [K, C] = GP_TRCOV(GP, TX, PREDCF) returns also the (noisy)
-%    covariance matrix C for observations y, which is sum of K and
-%    diagonal term, for example, from Gaussian noise.
-%
-%  See also
-%    GP_SET, GPCF_*
-
-% Copyright (c) 2006-2010 Jarno Vanhatalo
-% Copyright (c) 2010 Tuomas Nikoskinen
-
-% This software is distributed under the GNU General Public 
-% License (version 3 or later); please refer to the file 
-% License.txt, included with the software, for details.
-
-% no covariance functions?
-if length(gp.cf)==0 || (nargin>2 && ~isempty(predcf) && predcf(1)==0)
-  K=[];
-  C=[];
-  if nargout>1 && isfield(gp.lik.fh,'trcov')
-    C=sparse(0);
-    % Add Gaussian noise to the covariance
-    C = C + gp.lik.fh.trcov(gp.lik, x1);
-    if ~isempty(gp.jitterSigma2)
-      C=C+gp.jitterSigma2;
-    end
-  end
-  return
-end
-
-if ~(isfield(gp,'derivobs') && gp.derivobs)
-  
-  switch gp.type
-    case {'FULL' 'FIC' 'PIC' 'PIC_BLOCK' 'CS+FIC' 'VAR' 'DTC' 'SOR'}
-      [n,m]=size(x1);
-      n1 = n+1;
-      ncf = length(gp.cf);
-
-      % Evaluate the covariance without noise
-      K = sparse(0);
-      if nargin < 3 || isempty(predcf)
-        predcf = 1:ncf;
-      end
-      for i=1:length(predcf)
-        gpcf = gp.cf{predcf(i)};
-        K = K + gpcf.fh.trcov(gpcf, x1);
-      end
-
-      if ~isempty(gp.jitterSigma2)
-        if issparse(K)
-          K = K + sparse(1:n,1:n,gp.jitterSigma2,n,n);
-        else
-          K(1:n1:end)=K(1:n1:end) + gp.jitterSigma2;
-        end
-      end
-      if nargout>1
-        C=K;
-        if isfield(gp.lik.fh,'trcov')
-          % Add Gaussian noise to the covariance
-          C = C + gp.lik.fh.trcov(gp.lik, x1);
-        end
-      end
-
-    case 'SSGP'
-      [n,m]=size(x1);
-      n1 = n+1;
-      ncf = length(gp.cf);
-
-      % Evaluate the covariance without noise
-      gpcf = gp.cf{1};
-      K = gpcf.fh.trcov(gpcf, x1);
-
-      if nargout > 1 
[…]
