+++ external/dmlt/external/gpstuff/SuiteSparse/UMFPACK/Source/umf_mem_alloc_element.h
@@ -1,17 +1,0 @@
-/* -------------------------------------------------------------------------- */
-/* UMFPACK Copyright (c) Timothy A. Davis, CISE,                              */
-/* Univ. of Florida.  All Rights Reserved.  See ../Doc/License for License.   */
-/* web: http://www.cise.ufl.edu/research/sparse/umfpack                       */
-/* -------------------------------------------------------------------------- */
-
-GLOBAL Int UMF_mem_alloc_element
-(
-    NumericType *Numeric,
-    Int nrows,
-    Int ncols,
-    Int **Rows,
-    Int **Cols,
-    Entry **C,
-    Int *size,
-    Element **epout
-) ;

[…]
