+++ fieldtrip2fiff.m
@@ -9,14 +9,24 @@ function fieldtrip2fiff(filename, data, varargin)
 % as obtained from FT_PREPROCESSING, or a timelock structure obtained from
 % FT_TIMELOCKANALYSIS.
 %
+% Additional options can be specified as key-value pairs:
+%   precision = string ('single'/'double'), determines the precision with which
+%                 the numeric data is written to file, default is the class of the
+%                 numeric data
+% 
 % If the data comes from preprocessing and has only one trial, then it writes the
-% data into raw continuous format. If present in the data, the original header
-% is reused (also removing the non-used channels). Otherwise, the function 
-% attempts to create the header, which might or might not be correct (e.g. with
-% respect to the scaling and the sensor locations. If the data contains events in 
-% the cfg structure, it writes the events in the MNE format (three columns) into 
-% a file based on "filename", ending with "-eve.fif"
-%
+% data into raw continuous format. 
+% 
+% If present in the data, the original header is reused (also removing the non-used channels).
+% Otherwise, the function attempts to create the header, which might or might not be correct
+% (e.g. with respect to the scaling and the sensor locations. 
+% 
+% If the data contains events in the cfg structure, it writes the events in the MNE format
+% (three columns) into a file based on "filename", ending with "-eve.fif". If only a subset
+% of the events is to be written to the event file, one can use the additional key-value pair:
+%   eventtype = string (or cell array of string) with the eventtypes to be
+%                 written to file, default 'all'
+% 
 % See also FT_DATATYPE_RAW, FT_DATATYPE_TIMELOCK
 
 % Copyright (C) 2012-2013, Jan-Mathijs Schoffelen, Gio Piantoni
@@ -68,10 +78,19 @@ if israw
   iscomplex = ~isreal(data.trial{1});
 else
   fsample = 1./mean(diff(data.time));
-  dtype   = class(data.trial);
-  iscomplex = ~isreal(data.trial);
+  if isfield(data, 'trial')
+    dtype   = class(data.trial);
+    iscomplex = ~isreal(data.trial);
+  elseif isfield(data, 'avg')
+    dtype = class(data.avg);
+    iscomplex = ~isreal(data.avg);
+  end
 end
 precision = ft_getopt(varargin, 'precision', dtype);
+if ~ismember(precision, {'single' 'double'})
+  % FIXME consider casting non single/double data to the supported class
+  ft_error('only single or double precision data is supported');
+end
 
 % Create a fiff-header, or take information from the original header if possible
 if isfield(data, 'hdr') && isfield(data.hdr, 'orig') && isfield(data.hdr.orig, 'meas_id')
@@ -134,13 +153,27 @@ if israw
     event = [];
   end
   if ~isempty(event)
-    [eventlist, mappings] = convertevent(event);
+    eventtype = ft_getopt(varargin, 'eventtype', 'all');
+    [eventlist, mappings] = convertevent(event, eventtype);
     fiff_write_events(eventfile, eventlist, mappings);
     ft_info('Writing events to %s\n', eventfile)
   end
   
-elseif isepch
+elseif istlck
+  evoked.aspect_kind = 100;
+  evoked.is_smsh     = 0;
+  evoked.nave        = max(data.dof(:));
+  evoked.first       = round(data.time(1)*info.sfreq);
[…]
