### 2023-04-18 25cbd57
Update ft_redefinetrial.m
Clarification on the instructions for cfg.offset.
