### 2023-04-18 483f8ee
Expand ft_artifact_zvalue.m doc
I added cfg.artfctdef.zvalue.method and cfg.artfctdef.zvalue.ntrial. I'm not sure what cfg.artfctdef.zvalue.method = 'trialdemean' does, so I only mentioned it as an option and didn't explain what it does.
