### 2023-04-19 8f23e2c
ENH - fix for writing correct dimension for complex matrices
