### 2023-04-19 e9263af
ENH - some further updates, still a bit of a moving target
