### 2023-04-19 350d995
ENH - some path updates in test files that i came across
