+++ test/test_tutorial_natmeg_timefrequency.m
@@ -12,7 +12,7 @@ function test_tutorial_natmeg_timefrequency
 clear all
 close all
 
-cd(dccnpath('/home/common/matlab/fieldtrip/data/ftp/workshop/natmeg'));
+cd(dccnpath('/home/common/matlab/fieldtrip/data/ftp/workshop/natmeg2014'));
 
 cfg = [];
 cfg.dataset = 'oddball1_mc_downsampled.fif';
