+++ test/failed_bug3089.m
@@ -4,7 +4,7 @@ function test_bug3089
 % WALLTIME 00:20:00
 % DEPENDENCY ft_dipolefitting ft_compute_leadfield ft_apply_transform
 
-dataset = dccnpath('/home/common/matlab/fieldtrip/data/ftp/tutorial/natmeg/oddball1_mc_downsampled.fif');
+dataset = dccnpath('/home/common/matlab/fieldtrip/data/ftp/tutorial/natmeg2014/oddball1_mc_downsampled.fif');
 datadir = dccnpath('/home/common/matlab/fieldtrip/data/test/bug3089');
 addpath(datadir); % for the trialfuns
 
