### 2023-04-19 4a3226d
Ft2fiff (#2212)
* ENH - added option to input deltat

* ENH - improve code to reflect sensor locations + orientations

* ENH - allow writing raw files in double precision

* ENH - some updates

* ENH - update for double precision support, PR submitted to MNE folks as well

* ENH - synchronized external folder to the 20230322 version of the github repo

* ENH - more efficient handling of epoched fif-files

* ENH - some code reorganization

* FIX - some errors, this has probably never worked as intended

* FIX - reimplementation of coilaccuracy handling, to get the refs right

* ENH - some impro
[…]
