+++ fileio/private/mne2grad.m
@@ -1,4 +1,4 @@
-function [grad, elec] = mne2grad(hdr, dewar, coilaccuracy)
+function [grad, elec] = mne2grad(hdr, dewar, coilaccuracy, coildeffile)
 
 % MNE2GRAD converts a header from a fif file that was read using the MNE toolbox into
 % a gradiometer structure that can be understood by the FieldTrip low-level forward
@@ -9,6 +9,7 @@ function [grad, elec] = mne2grad(hdr, dewar, coilaccuracy)
 % where
 %   dewar        = boolean, whether to return it in dewar or head coordinates (default = false, i.e. head coordinates)
 %   coilaccuracy = empty or a number (default = [])
+%   coildeffile  = empty or a filename of a valid coil_def.dat file
 %
 % See also CTF2GRAD, BTI2GRAD
 
@@ -61,6 +62,12 @@ if nargin<3 || isempty(coilaccuracy)
   coilaccuracy = [];
 end
 
+if nargin<4 || isempty(coildeffile)
+  % if coilaccuracy is not empty, it will use the default coil_def.dat
+  % file, which is in external/mne
+  coildeffile = [];
+end
+
 % orig = fiff_read_meas_info(filename);
 % orig = fiff_setup_read_raw(filename);
 
@@ -107,18 +114,20 @@ if ~isempty(coilaccuracy)
   %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
   
   ft_hastoolbox('mne', 1);
-  [ftver, ftpath] = ft_version;
-  def = mne_load_coil_def(fullfile(ftpath, 'external', 'mne', 'coil_def.dat'));
-  
-  k = 1;
+  if isempty(coildeffile)
+    [ftver, ftpath] = ft_version;
+    coildeffile     = fullfile(ftpath, 'external', 'mne', 'coil_def.dat');
+  end
+  def = mne_load_coil_def(coildeffile);
   
   grad.chanpos = nan(length(orig.chs),3);
   grad.chanori = nan(length(orig.chs),3);
-  grad.label = cell(length(orig.chs),1);
+  grad.label   = cell(length(orig.chs),1);
   grad.tra      = nan(length(orig.chs),0);
   grad.chantype = repmat({'unknown'}, size(grad.label));
   grad.chanunit = repmat({'unknown'}, size(grad.label));
   
+  k = 1; % counter for the coils
   for i=1:length(orig.chs)
     thisdef = def([def.id]==orig.chs(i).coil_type & [def.accuracy]==coilaccuracy);
     if isempty(thisdef)
@@ -148,75 +157,55 @@ if ~isempty(coilaccuracy)
   grad.label = grad.label(:);
   grad.unit  = 'm'; % the coil_def.dat file is in meter
   
-  % all the chs.kinds and chs.coil_types are obtained from the MNE manual, p.210-211
-  for sel=find([orig.chs.kind]==1 & [orig.chs.coil_type]==2)' % planar gradiometers
-    grad.chantype(sel) = {'megplanar'}; %Neuromag-122 planar gradiometer
-  end
-  for sel=find([orig.chs.kind]==1 & [orig.chs.coil_type]==3012)' %planar gradiometers
-    grad.chantype(sel) = {'megplanar'}; %Type T1 planar grad
-  end
-  for sel=find([orig.chs.kind]==1 & [orig.chs.coil_type]==3013)' %planar gradiometers
-    grad.chantype(sel) = {'megplanar'}; %Type T2 planar grad
-  end
[…]
