### 2023-04-19 a99959a
Update ft_artifact_zvalue.m
Removed a duplicate in the doc, corrected a mistake I added to the doc, and hopefully made the code more robust when a table trl is used, as well as a test for making sure cfg.artfctdef.zvalue.method is one of the possible options. Still not sure what trialdemean is doing...
