### 2023-04-19 ff2abba
make ACCEL consistent with documentation
