### 2023-04-19 4e18168
introduce channel count for motion channels
