### 2023-04-20 87265d6
FIX - dashboard broke due to name change in FIFF constants
