### 2023-04-20 32a184b
FIX - dashboard broken due to wrong path
