### 2023-04-21 00d6250
document the units of the inverse operation for eLORETA
- added check that the user did not specify leadfield options in case the input already contains the leadfield
- added check that the user did not specify a regularization parameter in case the input already contains the filter
