### 2023-04-22 85ae3e1
switch true and false to lower case
see: https://github.com/bids-standard/bids-examples/commit/fe7468cc832d20542e6d7cfd00380acb2e80cf52
