### 2023-04-24 dc8b87f
add checks in hasfilter
- added check that the user did not specify leadfield options in case the input already contains the leadfield
- added check that the user did not specify inverse solution parameters in case the input already contains the filter
