### 2023-04-24 fa77f37
add empty-checks in sam
doubt: is the check on LF backprojection redundant now?
