### 2023-04-24 37e679a
add empty-checks in lcmv
