### 2023-04-26 88ea43c
ENH - overhaul of function for speed up
