### 2023-04-26 1ea6917
small tweaks
- Changed from double to single precision
- Modified number of samples
- grad.type hardcoded to 'megmag'
- added tra matrix to grad structure
