### 2023-04-26 c4954ec
create channels.tsv before json files
So we can use channels.tsv to calculate the ChannelCount for the json files.
On first sight, this should not create any problems, but would be good to check with other data.
