### 2023-04-26 d1db3bd
count channels in channels_tsv instead of hdr
to check: headlock channels (need_meg_json) also from channels.tsv?
