+++ data2bids.m
@@ -1206,16 +1206,16 @@ end % if need_mri_json
 %% need_meg_json
 if need_meg_json
     meg_json.SamplingFrequency          = hdr.Fs;
-    meg_json.MEGChannelCount            = sum(strcmpi(hdr.chantype, 'megmag') | strcmpi(hdr.chantype, 'meggrad') | strcmpi(hdr.chantype, 'megplanar') | strcmpi(hdr.chantype, 'megaxial'));
-    meg_json.MEGREFChannelCount         = sum(strcmpi(hdr.chantype, 'refmag') | strcmpi(hdr.chantype, 'refgrad') | strcmpi(hdr.chantype, 'refplanar') | strcmpi(hdr.chantype, 'ref'));
-    meg_json.EEGChannelCount            = sum(strcmpi(hdr.chantype, 'eeg'));
-    meg_json.ECOGChannelCount           = sum(strcmpi(hdr.chantype, 'ecog'));
-    meg_json.SEEGChannelCount           = sum(strcmpi(hdr.chantype, 'seeg') | strcmpi(hdr.chantype, 'dbs'));
-    meg_json.EOGChannelCount            = sum(strcmpi(hdr.chantype, 'eog'));
-    meg_json.ECGChannelCount            = sum(strcmpi(hdr.chantype, 'ecg'));
-    meg_json.EMGChannelCount            = sum(strcmpi(hdr.chantype, 'emg'));
-    meg_json.MiscChannelCount           = sum(strcmpi(hdr.chantype, 'misc') | strcmpi(hdr.chantype, 'unknown'));
-    meg_json.TriggerChannelCount        = sum(contains(lower(hdr.chantype), 'trigger'));
+    meg_json.MEGChannelCount            = sum(strcmpi(channels_tsv.type, 'megmag') | strcmpi(channels_tsv.type, 'meggrad') | strcmpi(channels_tsv.type, 'megplanar') | strcmpi(channels_tsv.type, 'megaxial'));
+    meg_json.MEGREFChannelCount         = sum(strcmpi(channels_tsv.type, 'refmag') | strcmpi(channels_tsv.type, 'refgrad') | strcmpi(channels_tsv.type, 'refplanar') | strcmpi(channels_tsv.type, 'ref'));
+    meg_json.EEGChannelCount            = sum(strcmpi(channels_tsv.type, 'eeg'));
+    meg_json.ECOGChannelCount           = sum(strcmpi(channels_tsv.type, 'ecog'));
+    meg_json.SEEGChannelCount           = sum(strcmpi(channels_tsv.type, 'seeg') | strcmpi(channels_tsv.type, 'dbs'));
+    meg_json.EOGChannelCount            = sum(strcmpi(channels_tsv.type, 'eog'));
+    meg_json.ECGChannelCount            = sum(strcmpi(channels_tsv.type, 'ecg'));
+    meg_json.EMGChannelCount            = sum(strcmpi(channels_tsv.type, 'emg'));
+    meg_json.MiscChannelCount           = sum(strcmpi(channels_tsv.type, 'misc') | strcmpi(channels_tsv.type, 'unknown'));
+    meg_json.TriggerChannelCount        = sum(contains(lower(channels_tsv.type), 'trigger') | strcmpi(channels_tsv.type, 'trig'));
     meg_json.RecordingDuration          = (hdr.nTrials*hdr.nSamples)/hdr.Fs;
 
     if hdr.nTrials>1
@@ -1246,12 +1246,12 @@ end % if need_meg_json
 %% need_eeg_json
 if need_eeg_json
     eeg_json.SamplingFrequency          = hdr.Fs;
[…]
