### 2023-04-26 42bd0da
nirs channels are specified by the user
e.g. NIRSCWOPTICALDENSITY
