+++ data2bids.m
@@ -1415,7 +1415,7 @@ if need_motion_json
     motion_json.ORNTChannelCount      = sum(strcmpi(channels_tsv.type, 'ORNT'));
     motion_json.VELChannelCount       = sum(strcmpi(channels_tsv.type, 'VEL'));
     motion_json.ANGVELChannelCount    = sum(strcmpi(channels_tsv.type, 'ANGVEL'));
-    motion_json.ACCChannelCount       = sum(strcmpi(channels_tsv.type, 'ACC'));
+    motion_json.ACCELChannelCount       = sum(strcmpi(channels_tsv.type, 'ACCEL'));
     motion_json.ANGACCChannelCount    = sum(strcmpi(channels_tsv.type, 'ANGACC'));
     motion_json.MAGNChannelCount      = sum(strcmpi(channels_tsv.type, 'MAGN'));
     motion_json.JNTANGChannelCount    = sum(strcmpi(channels_tsv.type, 'JNTANG'));
