### 2023-05-01 a36a385
FIX #2240 - rename test files and add Precision field to the JSON
