+++ fileio/private/opm_fil.m
@@ -41,8 +41,8 @@ needdat = (nargin==5);
 assert(isfile(filename), sprintf('file "%s" not found', filename));
 
 % Remove the _meg part of the filename
-if strcmp(f(end-3:end),'_meg')
-    f = f(1:end-4);
+if endsWith(f, '_meg')
+  f = f(1:end-4);
 end
 
 % Get the BIDS compliant files
@@ -52,28 +52,32 @@ datafile      = fullfile(p,[f '_meg.bin']);
 headerfile    = fullfile(p,[f '_meg.json']);
 positionsfile = fullfile(p,[f '_positions.tsv']);
 
-precision = 'single';
-
-switch precision
-  case 'single'
-    samplesize = 4;
-  case 'double'
-    samplesize = 8;
-end
-
-
 if needhdr
   %% read the header
   try
-      fid = fopen(headerfile, 'rt');
-      header = jsondecode(fread(fid, [1 inf], 'char=>char'));
-      fclose(fid);
+    fid = fopen(headerfile, 'rt');
+    header = jsondecode(fread(fid, [1 inf], 'char=>char'));
+    fclose(fid);
   catch
-      ft_warning('Cannot open header');
+    ft_warning('Cannot open header');
   end
-  
+
+  % determine the precision from the json header
+  if isfield(header, 'Precision')
+    precision = header.Precision;
+  else
+    precision = 'single';
+  end
+
+  switch precision
+    case 'single'
+      samplesize = 4;
+    case 'double'
+      samplesize = 8;
+  end
+
   channels  = readtable(channelsfile, 'Delimiter', 'tab', 'FileType', 'text');
-  
+
   % this one is optional
   if exist(coordsysfile, 'file')
     fid = fopen(coordsysfile, 'rt');
@@ -82,16 +86,16 @@ if needhdr
   else
     coordsys = [];
   end
-  
+
   % this one is optional
   if exist(positionsfile, 'file')
     positions = readtable(positionsfile, 'Delimiter', 'tab', 'FileType', 'text');
   else
     positions = [];
   end
-  
+
   d = dir(datafile);
-  
+
   hdr.label       = channels.name;
   hdr.nChans      = size(channels, 1);
   hdr.nSamples    = (d.bytes/(size(channels, 1) * samplesize))-1;
@@ -100,17 +104,17 @@ if needhdr
   hdr.Fs          = header.SamplingFrequency;
   hdr.chantype    = channels.type;
   try
-      hdr.chanunit    = channels.unit;
+    hdr.chanunit    = channels.unit;
   catch
-      hdr.chanunit    = repmat({'unknown'}, size(hdr.label));
+    hdr.chanunit    = repmat({'unknown'}, size(hdr.label));
   end
-  
+
   % keep the original header details
   hdr.orig.header     = header;
   hdr.orig.coordsys   = coordsys;
   hdr.orig.positions  = positions;
   hdr.orig.channels   = channels;
-  
+
   if ~isempty(positions)
     % FIXME construct a grad structure
     hdr.grad.label = positions.name;
@@ -123,30 +127,46 @@ if needhdr
       hdr.grad.coordsys = coordsys.MEGCoordinateSystem;
     end
   end
-  
+
   % return the header details
   varargout = {hdr};
-  
+
 elseif needdat
   %% read the data, note that it is big endian
+
+  % determine the precision from the json header
+  % see https://github.com/fieldtrip/fieldtrip/issues/2240
+  if isfield(hdr.orig.header, 'Precision')
+    precision = hdr.orig.header.Precision;
+  else
+    precision = 'single';
+  end
+
+  switch precision
+    case 'single'
+      samplesize = 4;
+    case 'double'
[…]
