### 2023-05-02 2a3e6a3
DOC - this fixes #2235 by improving the data2bids help
It also includes a test script that did not reproduce the error.
