### 2023-05-03 7e3c059
Updated support for QZFM_UCL data
#2130

- So we are now thinking of including HelmetManufacturer and HelmetManufacturersModelName in the json file. The latter would contain a text field with something like: 'slotnumber-56_location-wholehead_fit-custom'

- I have used the grad.type = 'QZFM_UCL' to differentiate ourselves from other facilities using QZFM sensors

All seems to work fine with existing test data.

Issue: I can't seem to get ft_channelselection working without providing a third input argument (i.e. it's not picking up the senstype automatically for some reason)
