### 2023-05-04 dde2ac2
ENH - improvements to fiff-file handling
