+++ test/test_fieldtrip2fiff.m
@@ -1,57 +1,145 @@
-function test_fieldtrip2fiff
+function test_fieldtrip2fiff(mnedatadir)
 
+%%
 savedir = tempdir;
 
-% test a variety of MEG files
+%%
+% Use the mne-testing-data if path is specified
+if nargin
+  pw_dir = pwd;
+  cd(mnedatadir);
+
+%   datasets = {
+%     fullfile(mnedatadir, 'CTF', 'catch-alp-good-f.ds')
+%     fullfile(mnedatadir, 'CTF', 'somMDYO-18av.ds')
+%     fullfile(mnedatadir, 'CTF', 'testdata_ctf.ds')
+%     fullfile(mnedatadir, 'CTF', 'testdata_ctf_mc.ds')
+%     fullfile(mnedatadir, 'CTF', 'testdata_ctf_pseudocontinuous.ds')
+%     fullfile(mnedatadir, 'CTF', 'testdata_ctf_short.ds')
+%     fullfile(mnedatadir, 'CTF', 'testdata_ctf_short_discontinuous.ds')
+%     fullfile(mnedatadir, 'BTi', '4Dsim',   'c,rfDC')
+%     fullfile(mnedatadir, 'BTi', 'erm_HFH', 'c,rfDC')
+%     fullfile(mnedatadir, 'KIT', 'ArtificalSignalData_RICOH_1khz.con')
+%     fullfile(mnedatadir, 'KIT', 'ArtificalSignalData_Yokogawa_1khz.con')
+%     fullfile(mnedatadir, 'KIT', '010409_Motor_task_coregist-export_tiny_1s.con')
+%     fullfile(mnedatadir, 'KIT', 'data_berlin.con')
+%     fullfile(mnedatadir, 'MEG', 'sample', 'sample_audvis_trunc-ave.fif')
+%     fullfile(mnedatadir, 'MEG', 'sample', 'sample_audvis-ave.fif')
+%     };
+
+  % the idea is here to test the status quo for reading and writing, hence
+  % operate in first instance only on the fif-files
+  d = dir('*/*.fif');
+  d = cat(1, d, dir('*/*/*.fif'));
+  d = cat(1, d, dir('*/*/*/*.fif'));
+  
+  datasets = cell(numel(d),1);
+  for k = 1:numel(d)
+    datasets{k} = fullfile(d(k).folder, d(k).name);
+  end
+
+  ok = false(numel(datasets),1);
+  for k = 1:numel(datasets)
+    cfg = [];
+    cfg.dataset = datasets{k};
+    try
+      data = ft_preprocessing(cfg);
+      ok(k,1) = true;
+    catch
+      ok(k,1) = false;
+    end
+  end
+
+
+  cd(pw_dir);
+end
+
 datadir = dccnpath('/home/common/matlab/fieldtrip/data/test/original/meg');
 
-cfg = [];
+% Use the 'Subject01.ds' dataset for basic testing of the writing and reading
+cfg         = [];
 cfg.dataset = fullfile(datadir, 'ctf151', 'Subject01.ds');
-hdr = ft_read_header(cfg.dataset);
+hdr         = ft_read_header(cfg.dataset);
 
-cfg.trl = [(0:1).*hdr.nSamples+1;(1:2).*hdr.nSamples;-300.*ones(1,2)]';
+% Define 2 epochs worth of data
+cfg.trl     = [(0:1).*hdr.nSamples+1;(1:2).*hdr.nSamples;-300.*ones(1,2)]';
 cfg.channel = 'MEG';
-data = ft_preprocessing(cfg);
+data        = ft_preprocessing(cfg);
+
+% Benchmark trial for the original data
 t1 = data.trial{1};
[…]
