### 2023-05-06 7d7ee5a
ENH - fix for correct representation of coordinate systems for ctf-like coordsys
