+++ utilities/ft_affinecoordinates.m
@@ -42,7 +42,6 @@ function [transform] = ft_affinecoordinates(original, target, varargin)
 
 % get the options
 requireorigin = ft_getopt(varargin, 'requireorigin', true);
-computeall    = ft_getopt(varargin, 'computeall', false);
 
 % ensure these are in lower case
 original = lower(original);
@@ -65,31 +64,6 @@ generic = {
 % the specific ones specify the axes and also the origin
 specific = {'ctf', 'bti', 'fourd', 'yokogawa', 'eeglab', 'neuromag', 'itab', 'acpc', 'spm', 'mni', 'fsaverage', 'tal', 'scanras', 'scanlps', 'dicom'};
 
-% pre allocate all exhaustive combinations
-coordsys = [specific generic];
-n        = numel(coordsys);
-ngeneric = numel(generic);
-nspecific = numel(specific);
-n1       = repmat(1:n,    n, 1);
-n2       = repmat((1:n)', 1, n);
-
-from = coordsys(n1(:))';
-to   = coordsys(n2(:))';
-T    = nan*zeros(4, 4, numel(to));
-
-% numeric indices of the corresponding from/to lists into the static
-% coordsys list, speeds up comparisons in the below.
-from_idx = n1(:);
-to_idx   = n2(:);
-
-% do this once, and use it to check for the existence of the transform, to
-% be able to break out of the exhaustive for-loop below
-original_idx = strcmp(coordsys, original);
-target_idx   = strcmp(coordsys, target);
-
-% make the combined and the inverse transformations where possible
-implemented = zeros(length(coordsys)); % this is only for debugging
-
 %--------------------------------------------------------------------------
 % These approximate alignments are based on transformation matrices that were
 % determined by clicking on the CTF fiducial locations in the canonical T1 template
@@ -98,331 +72,202 @@ implemented = zeros(length(coordsys)); % this is only for debugging
 % All of the transformation matrices here are expressed in millimeter.
 
 % this is based on the ear canals, see ALIGN_CTF2ACPC
-fromthis = 'acpc';
-tothis   = 'ctf';
-M = [
+acpc2ctf = [
    0.0000  0.9987  0.0517  34.7467
   -1.0000  0.0000  0.0000   0.0000
    0.0000 -0.0517  0.9987  52.2749
    0.0000  0.0000  0.0000   1.0000
   ];
-[T, implemented] = transform_add(M, fromthis, tothis, T, from, to, implemented);
 
 % this is based on the ear canals, see ALIGN_NEUROMAG2ACPC
-fromthis = 'acpc';
-tothis   = 'neuromag';
-M = [
+acpc2neuromag = [
   1.0000  0.0000  0.0000   0.0000
   0.0000  0.9987  0.0517  34.7467
   0.0000 -0.0517  0.9987  52.2749
   0.0000  0.0000  0.0000   1.0000
   ];
-[T, implemented] = transform_add(M, fromthis, tothis, T, from, to, implemented);
 
 % see http://freesurfer.net/fswiki/CoordinateSystems
-fromthis = 'fsaverage';
-tothis   = 'mni';
-M = [
+fsaverage2mni = [
    0.9975   -0.0073    0.0176   -0.0429
    0.0146    1.0009   -0.0024    1.5496
   -0.0130   -0.0093    0.9971    1.1840
    0.0000    0.0000    0.0000    1.0000
   ];
-[T, implemented] = transform_add(M, fromthis, tothis, T, from, to, implemented);
 
 % this is a 90 degree rotation around the z-axis
-fromthis = 'ctf';
-tothis   = 'neuromag';
-M = [
+ctf2neuromag = [
   0.0000   -1.0000    0.0000    0.0000
[…]
