+++ test/test_ft_affinecoordinates.m
@@ -1,323 +1,0 @@
-function test_ft_affinecoordinates
-
-% WALLTIME 01:00:00
-% MEM 4gb
-% DEPENDENCY ft_affinecoordinates
-
-generic = {
-  'als'; 'ali'; 'ars'; 'ari';...
-  'pls'; 'pli'; 'prs'; 'pri';...
-  'las'; 'lai'; 'ras'; 'rai';...
-  'lps'; 'lpi'; 'rps'; 'rpi';...
-  'asl'; 'ail'; 'asr'; 'air';...
-  'psl'; 'pil'; 'psr'; 'pir';...
-  'sal'; 'ial'; 'sar'; 'iar';...
-  'spl'; 'ipl'; 'spr'; 'ipr';...
-  'sla'; 'ila'; 'sra'; 'ira';...
-  'slp'; 'ilp'; 'srp'; 'irp';...
-  'lsa'; 'lia'; 'rsa'; 'ria';...
-  'lsp'; 'lip'; 'rsp'; 'rip'}';
-
-% the specific ones specify the axes and also the origin
-specific = {'ctf', 'bti', 'fourd', 'yokogawa', 'eeglab', 'neuromag', 'itab', 'acpc', 'spm', 'mni', 'fsaverage', 'tal', 'scanras', 'scanlps', 'dicom'};
-
-coordsys = [generic specific];
-ncoordsys = numel(coordsys);
-
-f = randi(ncoordsys, 100, 1);
-t = randi(ncoordsys, 100, 1);
-
-T1 = zeros(4,4,100);
-T2 = zeros(4,4,100);
-Tt1 = zeros(100,1);
-TT2 = zeros(100,1);
-for k = 1:numel(f)
-  from = coordsys{f(k)};
-  to = coordsys{t(k)};
-  fprintf('%s -> %s\n', from, to);
-  fprintf('k=%d, new, ', k);
-  try
-  tic;T1(:,:,k) = ft_affinecoordinates(from,to); Tt1(k,1) = toc;
-  fprintf('time elapsed: %2.2g s\n', Tt1(k));
-  catch
-    fprintf('failed\n');
-  end
-  fprintf('k=%d, orig, ', k);
-  try
-  tic;T2(:,:,k) = ft_affinecoordinates_orig(from,to); Tt2(k,1) = toc;
-  fprintf('time elapsed: %2.2g s\n', Tt2(k));
-  catch
-    fprintf('failed\n');
-  assert(isequal(T1(:,:,k),T2(:,:,k)));
-  end
-end
-
-
-function [transform] = ft_affinecoordinates_orig(original, target, varargin)
-
-% this is the original implementation from Robert
-
-% these are the 48 generic axis orientation triplets, these specify the axes but no origin
-%   a = anterior
-%   p = posterior
-%   l = left
-%   r = right
-%   s = superior
-%   i = inferior
-
-% get the options
-requireorigin = ft_getopt(varargin, 'requireorigin', true);
-
-% ensure these are in lower case
-original = lower(original);
-target = lower(target);
-
-generic = {
-  'als'; 'ali'; 'ars'; 'ari';...
-  'pls'; 'pli'; 'prs'; 'pri';...
-  'las'; 'lai'; 'ras'; 'rai';...
-  'lps'; 'lpi'; 'rps'; 'rpi';...
-  'asl'; 'ail'; 'asr'; 'air';...
-  'psl'; 'pil'; 'psr'; 'pir';...
-  'sal'; 'ial'; 'sar'; 'iar';...
-  'spl'; 'ipl'; 'spr'; 'ipr';...
-  'sla'; 'ila'; 'sra'; 'ira';...
-  'slp'; 'ilp'; 'srp'; 'irp';...
[…]
