### 2023-05-11 638f9b4
FIX typo, stcmpi should be strcmpi
