### 2023-05-11 291db7a
removed EXG, it is not needed and also not being discussed
