### 2023-05-11 bf8b8a1
ENH - adde GeneratedBy with FieldTrip as default
- some other small cleanups
