+++ data2bids.m
@@ -185,7 +185,7 @@ function [cfg] = data2bids(cfg, varargin)
 % officially supported data types.
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-% Copyright (C) 2018-2020, Robert Oostenveld
+% Copyright (C) 2018-2023, Robert Oostenveld
 %
 % This file is part of FieldTrip, see http://www.fieldtriptoolbox.org
 % for the documentation and details.
@@ -216,8 +216,8 @@ ft_preamble init
 ft_preamble debug
 
 if ft_abort
-    % do not continue function execution in case the outputfile is present and the user indicated to keep it
-    return
+  % do not continue function execution in case the outputfile is present and the user indicated to keep it
+  return
 end
 
 % ensure backward compatibility
@@ -278,21 +278,21 @@ cfg.datatype  = ft_getopt(cfg, 'datatype');
 fn = {'sub', 'ses', 'task', 'tracksys', 'acq', 'ce', 'rec', 'dir', 'run', 'mod', 'echo', 'proc', 'space', 'desc'};
 
 for i=1:numel(fn)
-    if ischar(cfg.(fn{i})) && any(cfg.(fn{i})=='-')
-        ft_error('the field cfg.%s cannot contain a "-"', fn{i});
-    end
+  if ischar(cfg.(fn{i})) && any(cfg.(fn{i})=='-')
+    ft_error('the field cfg.%s cannot contain a "-"', fn{i});
+  end
 end
 
 if isempty(cfg.datatype)
-    modality = {'meg', 'eeg', 'ieeg', 'emg', 'motion', 'audio', 'video', 'eyetracker', 'physio', 'stim', 'motion', 'nirs'};
-    for i=1:numel(modality)
-        if isfield(cfg, modality{i}) && ~isempty(cfg.(modality{i}))
-            % the user specified modality-specific options, assume that the datatype matches
-            cfg.datatype = modality{i};
-            ft_notice('assuming that the datatype is %s', cfg.datatype);
-            continue
-        end
-    end % for each modality
+  modality = {'meg', 'eeg', 'ieeg', 'emg', 'motion', 'audio', 'video', 'eyetracker', 'physio', 'stim', 'motion', 'nirs'};
+  for i=1:numel(modality)
+    if isfield(cfg, modality{i}) && ~isempty(cfg.(modality{i}))
+      % the user specified modality-specific options, assume that the datatype matches
+      cfg.datatype = modality{i};
+      ft_notice('assuming that the datatype is %s', cfg.datatype);
+      continue
+    end
+  end % for each modality
 end
 
 cfg.dicomfile     = ft_getopt(cfg, 'dicomfile');                      % get header details from the specified DICOM files
@@ -333,6 +333,14 @@ cfg.dataset_description.Funding             = ft_getopt(cfg.dataset_description,
 cfg.dataset_description.EthicsApprovals     = ft_getopt(cfg.dataset_description, 'EthicsApprovals'       ); % OPTIONAL. List of ethics committee approvals of the research protocols and/or protocol identifiers.
 cfg.dataset_description.ReferencesAndLinks  = ft_getopt(cfg.dataset_description, 'ReferencesAndLinks'    ); % OPTIONAL. List of references to publication that contain information on the dataset, or links.
 cfg.dataset_description.DatasetDOI          = ft_getopt(cfg.dataset_description, 'DatasetDOI'            ); % OPTIONAL. The Document Object Identifier of the dataset (not the corresponding paper).
+
+% this is a structure, and in the json file an object
[…]
