### 2023-05-11 cefecef
ENH - remove xxxChannelCount in case it has a count of zero
