### 2023-05-11 efdda00
to be placed in a separate file
which will be standardized later, see: https://github.com/bids-standard/bids-specification/issues/1488#issuecomment-1544027528
