### 2023-05-11 3b7603a
ANGVEL becomes GYRO
see https://github.com/bids-standard/bids-specification/issues/1436#issuecomment-1527155987
