### 2023-05-11 1e6f5e1
add column for reference_frame
see: https://github.com/bids-standard/bids-specification/issues/1488#issuecomment-1544027528
