### 2023-05-12 d39293e
ft2mne (#2246)
* ENH - represent channel coordinates (at least for CTF) consistent w.r.t. MNE-python's convention
