+++ fieldtrip2fiff.m
@@ -12,7 +12,13 @@ function fieldtrip2fiff(filename, data, varargin)
 % Additional options can be specified as key-value pairs:
 %   precision = string ('single'/'double'), determines the precision with which
 %                 the numeric data is written to file, default is the class of the
-%                 numeric data
+%                 numeric data.
+%   coordsys  = string ('native'/'neuromag'), determines the coordinate system in which
+%                 the MEG sensors are written (default: 'neuromag'). This option does not 
+%                 have an effect on EEG electrodes or fNIRS optodes. If coordsys = 'neuromag'
+%                 the MEG sensors are expressed in (approximate) neuromag coordinates, which
+%                 may facilitate downstream handling of the fif-files in other software, e.g.
+%                 MNE-python. This is according to the official definitions of the fif-file format
 % 
 % If the data comes from preprocessing and has only one trial, then it writes the
 % data into raw continuous format. 
@@ -52,6 +58,9 @@ function fieldtrip2fiff(filename, data, varargin)
 
 % this ensures that the path is correct and that the ft_defaults global variable is available
 ft_defaults
+
+coordsys  = ft_getopt(varargin, 'coordsys', 'neuromag');
+headshape = ft_getopt(varargin, 'headshape', []);
 
 % ensure that the filename has the correct extension
 [pathstr, name, ext] = fileparts(filename);
@@ -108,7 +117,10 @@ else
   info.highpass        = NaN;
   info.lowpass         = NaN;
   
-  % these are not strictly necessary, but the inverse functions in MNE works better if this matrix is present
+  % these are not strictly necessary, for basic functionality, i.e. time series manipulation. Thinks work
+  % best if an attempt is made to represent MEG-sensors according to the MNE conventions, i.e. sensors in 
+  % (neuromag) device coordinates, and the appropriate additional transformations specified. Fieldtrip2fiff 
+  % makes an attempt to do this, if 'coordsys' = 'neuromag' (handled below)
   info.dev_head_t.from  = 1;
   info.dev_head_t.to    = 4;
   info.dev_head_t.trans = eye(4); % this is of course not correct, but the exact transformation depends on the system
@@ -123,6 +135,75 @@ else
   info.isepoched    = double(isepch);
   info.iscontinuous = double(israw);
   info.sfreq        = fsample;
+end
+
+if isfield(data, 'grad')
+  % express in m to be sure
+  data.grad = ft_convert_units(data.grad, 'm');
+
+  % data contains a gradiometer structure, for which the coordsys is relevant
+  if ~isfield(data.grad, 'coordsys')
+    data.grad = ft_determine_coordsys(data.grad);
+  end
+  if isequal(coordsys, 'neuromag') || ~isequal(data.grad.coordsys, 'neuromag')
+    origcoordsys = data.grad.coordsys;
+    ft_info('Converting the gradiometer description into approximate neuromag coordinates. This optimizes the odds that the data can be seamlessly used in MNE-python.');
+    T         = ft_affinecoordinates(data.grad.coordsys, 'neuromag');
[…]
