### 2023-05-15 998c973
Update ft_topoplotIC.m
Extend doc
