### 2023-05-17 3132aad
FIX - the change in channel handling in #2238 broke data2bids for mri and other modalities that don't need a channels.tsv file
