+++ data2bids.m
@@ -1118,6 +1118,7 @@ coordsystem_settings = keepfields(cfg.coordsystem, fn);
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %% construct the content for the json and tsv files
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
 %% need_channels_tsv
 if need_channels_tsv
 
@@ -1175,230 +1176,232 @@ if need_channels_tsv
   % channel types in BIDS must be in upper case
   channels_tsv.type = upper(channels_tsv.type);
 
-  %% need_mri_json
-  if need_mri_json
-    % start with the information from the DICOM header
-    mri_json = keepfields(dcm, [fieldnames(mri_settings); fieldnames(generic_settings)]);
-    % merge the information specified by the user with that from the data
-    % in case fields appear in both, the first input overrules the second
-    mri_json = mergestruct(mri_settings,     mri_json, false);
-    mri_json = mergestruct(generic_settings, mri_json, false);
-  end % if need_mri_json
+end % if need_channels_tsv
 
-  %% need_meg_json
-  if need_meg_json
-    meg_json.SamplingFrequency          = hdr.Fs;
-    meg_json.MEGChannelCount            = sum(strcmpi(channels_tsv.type, 'megmag') | strcmpi(channels_tsv.type, 'meggrad') | strcmpi(channels_tsv.type, 'megplanar') | strcmpi(channels_tsv.type, 'megaxial'));
-    meg_json.MEGREFChannelCount         = sum(strcmpi(channels_tsv.type, 'refmag') | strcmpi(channels_tsv.type, 'refgrad') | strcmpi(channels_tsv.type, 'refplanar') | strcmpi(channels_tsv.type, 'ref'));
-    meg_json.EEGChannelCount            = sum(strcmpi(channels_tsv.type, 'eeg'));
-    meg_json.ECOGChannelCount           = sum(strcmpi(channels_tsv.type, 'ecog'));
-    meg_json.SEEGChannelCount           = sum(strcmpi(channels_tsv.type, 'seeg') | strcmpi(channels_tsv.type, 'dbs'));
-    meg_json.EOGChannelCount            = sum(strcmpi(channels_tsv.type, 'eog'));
-    meg_json.ECGChannelCount            = sum(strcmpi(channels_tsv.type, 'ecg'));
-    meg_json.EMGChannelCount            = sum(strcmpi(channels_tsv.type, 'emg'));
-    meg_json.MiscChannelCount           = sum(strcmpi(channels_tsv.type, 'misc') | strcmpi(channels_tsv.type, 'unknown'));
-    meg_json.TriggerChannelCount        = sum(contains(lower(channels_tsv.type), 'trigger') | strcmpi(channels_tsv.type, 'trig'));
-    meg_json.RecordingDuration          = (hdr.nTrials*hdr.nSamples)/hdr.Fs;
+%% need_mri_json
+if need_mri_json
+  % start with the information from the DICOM header
+  mri_json = keepfields(dcm, [fieldnames(mri_settings); fieldnames(generic_settings)]);
+  % merge the information specified by the user with that from the data
+  % in case fields appear in both, the first input overrules the second
+  mri_json = mergestruct(mri_settings,     mri_json, false);
+  mri_json = mergestruct(generic_settings, mri_json, false);
+end % if need_mri_json
 
-    if hdr.nTrials>1
-      meg_json.EpochLength              = hdr.nSamples/hdr.Fs;
-    end
[…]
