### 2023-05-17 d30993a
FIX - dashboard failed due to some flawed logic and ill-defined dewar coordsys
