### 2023-05-31 23d74b1
improve support for handling of tables in the input cfg
