### 2023-06-01 505bf89
FIX - reading of events from openvibe_mat
