### 2023-06-02 8c1bdb6
FIX - the location of the test data changed with https://github.com/fieldtrip/website/commit/ff0874aa2760fb22a9cead5b130816add9206e74
