+++ test/test_shared_virtual_channels.m
@@ -1,13 +1,13 @@
-function failed_shared_virtual_channels
+function test_shared_virtual_channels
 
 % MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_timelockanalysis ft_sourceanalysis ft_channelselection ft_databrowser
 
-load(dccnpath('/home/common/matlab/fieldtrip/data/ftp/tutorial/beamformer_extended/hdm.mat'));
-load(dccnpath('/home/common/matlab/fieldtrip/data/ftp/tutorial/beamformer_extended/source_diff.mat'));
-load(dccnpath('/home/common/matlab/fieldtrip/data/ftp/tutorial/beamformer_extended/source_coh_lft.mat'));
-load(dccnpath('/home/common/matlab/fieldtrip/data/ftp/tutorial/beamformer_extended/data_cmb.mat'));
+load(dccnpath('/home/common/matlab/fieldtrip/data/ftp/tutorial/beamformingextended/hdm.mat'));
+load(dccnpath('/home/common/matlab/fieldtrip/data/ftp/tutorial/beamformingextended/source_diff.mat'));
+load(dccnpath('/home/common/matlab/fieldtrip/data/ftp/tutorial/beamformingextended/source_coh_lft.mat'));
+load(dccnpath('/home/common/matlab/fieldtrip/data/ftp/tutorial/beamformingextended/data_cmb.mat'));
 
 [maxval, maxcohindx] = max(source_coh_lft.avg.coh);
 source_coh_lft.pos(maxcohindx, :)
