### 2023-06-05 126c9aa
FIX - issue with point magnetometer data
