### 2023-06-07 36687c4
Update ft_singleplotER.m
Clarify in the doc when baselinetype is relevant.
