### 2023-06-12 30dbbb4
fix typo ft_singleplotER.m
Fixed an accidental typo I added to the code.
