### 2023-06-12 c81899f
fix typo ft_singleplotER.m
Another one...
