### 2023-06-16 #5566 @editorialbot
[REVIEW]: Reducing the efforts to create reproducible analysis code with FieldTrip
**Submitting author:** @matsvanes (<a href="http://orcid.org/0000-0002-7133-509X">Mats W.J. van Es</a>)
**Repository:** https://github.com/fieldtrip/fieldtrip
**Branch with paper.md** (empty if default branch): JOSS
**Version:** 20231220
