### 2023-06-23 #5566 @crsl4
Hi @matsvanes, in response to @gflofst 's comment, do you mind sharing your author's contributions to FieldTrip? Thanks!
