### 2023-06-28 2896148
add axis handle in order to support subplot functionality; add frequency axis log scaling
