### 2023-06-28 c799a3a
Update ft_icabrowser.m
typo fix, as discussed
