### 2023-06-28 d63e543
add axis handle in order to support subplot functionality; add freque… (#2264)
* add axis handle in order to support subplot functionality; add frequency axis log scaling

* Update ft_icabrowser.m

* Update ft_icabrowser.m

typo fix, as discussed

---------

Co-authored-by: Jan-Mathijs Schoffelen <1517611+schoffelen@users.noreply.github.com>
