### 2023-07-02 a5a24a6
FIX - do not refer to exg, that has been removed as dataformat
