### 2023-07-02 9610bc3
ENH - only use megmag and meggrad as default channels
