### 2023-07-02 c59bdcd
ENH - support 'meggrad' also for CTF
