### 2023-07-04 #5566 @matsvanes
@crsl4 Thanks for reaching out. I can see how my contributions are not clear from FieldTrip's github directly. I hope the following clears things up.
Unfortunately, much of the work submitted here has somewhat dispersed origins. My contributions were mostly in design, testing, and documentation, and haven't ended up in the main branch of the FieldTrip toolbox. See below for a trail of my involvement in the various stages.

**Design**
The toolbox addition was inspired by [this](https://github.com/fieldtrip/fieldtrip/compare/master...matsvanes:fieldtrip:diary) early implementation.

**Development and testing**
Mostly through comments on issues and pulls on the FieldTrip Github, for example: [#1617](https://github.com/fieldtrip/fieldtrip/issues/1617), [#1362](https://github.com/fieldtrip/fieldtrip/pull/1362), [#1330](https://github.com/fieldtrip/fieldtrip/pull/1330), [#1317](https://github.com/fieldtrip/fieldtrip/pull/1317), [#973](https://github.com/fieldtrip/fieldtrip/pull/973)

**Documentation**
 A trail of developing the tests and documentation is found [here](https://github.com/matsvanes/reproducescript/tree/master), and their implementation on the FieldTrip website [here](https://github.com/fieldtrip/website/commit/6b73fcd87a6f57b21d70053e14639fe38a9cdd34).
