### 2023-07-04 1aa4306
Update resample.m
Fix out crash trying to resample Hanson/Wakeman data from 1100 Hz to 100 Hz when signal processing toolbox absent
