### 2023-07-05 #5566 @gflofst
@matsvanes JOSS will have to judge if this is sufficient. I don't have adequate guidance except to see if "significant" effort was involved on the submitter's part. If you have any design notes or otherwise that can be added to this thread, that would make it more easily pass the bar. I appreciate tests and documentation rather than just code. I would certainly count those. Similar software is hard to judge attributions without something explicit.
