### 2023-07-06 #5566 @matsvanes
Thank you for the suggestion @gflofst. Of importance would definitely be the two links above under Documentation. In case it's relevant, we also have a [pre-print](https://www.biorxiv.org/content/10.1101/2021.02.05.429886v1.abstract) available.
