### 2023-07-11 09ddde5
Update DEPENDENCY list of test functions by adding a new field in the test description for data. The data will be split into three categories NODATA/PUBLICDATA/PRIVATEDATA
