+++ test/test_ft_spike_psth.m
@@ -3,6 +3,7 @@ function test_ft_spike_psth()
 % MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_spike_psth
+% NODATA
 
 % fixed number of spikes, equal length trials, one LFP channel, gamma distribution psth
 % one channel with spikes at known positions
