+++ test/test_data2bids.m
@@ -3,6 +3,7 @@ function test_data2bids
 % WALLTIME 00:20:00
 % MEM 5gb
 % DEPENDENCY data2bids
+% PRIVATEDATA
 
 %% Example with only behavioural data
 cd(dccnpath('/home/common/matlab/fieldtrip/data/test/data2bids/test1'))
