+++ test/test_issue1329.m
@@ -3,6 +3,7 @@ function test_issue1329
 % WALLTIME 00:60:00
 % MEM 2gb
 % DEPENDENCY ft_convert_coordsys
+% NODATA
 
 %%
 
