+++ test/test_bug1483.m
@@ -3,6 +3,7 @@ function test_bug1483
 % MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_selectdata ft_selectdata_new ft_selectdata_old
+% NODATA
 
 %% first confirm the reported bug, i.e. the incapability of ft_selectdata_new
 % to select 'rpt'
