+++ test/test_bug1129.m
@@ -3,6 +3,7 @@ function test_bug1129
 % MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_preprocessing ft_preproc_bandpassfilter ft_preproc_bandstopfilter ft_preproc_lowpassfilter ft_preproc_highpassfilter
+% PRIVATEDATA
 
 % change filter order for 'fir' (fir1) filter type, rather than 25 by
 % default, instead to be based on Fs, low-frequency, and data-length
