+++ test/test_ft_checkdata.m
@@ -3,6 +3,7 @@ function test_ft_checkdata
 % MEM 2gb
 % WALLTIME 00:20:00
 % DEPENDENCY ft_checkdata
+% NODATA
 
 %% converting raw data to timelock data
 
