+++ contrib/spike/private/getdimord.m
@@ -1,836 +1,3 @@
-function dimord = getdimord(data, field, varargin)
-
-% GETDIMORD determine the dimensions and order of a data field in a FieldTrip
-% structure.
-%
-% Use as
-%   dimord = getdimord(data, field)
-%
-% See also GETDIMSIZ, GETDATFIELD, FIXDIMORD
-
-% Copyright (C) 2014-2022, Robert Oostenveld
-%
-% This file is part of FieldTrip, see http://www.fieldtriptoolbox.org
-% for the documentation and details.
-%
-%    FieldTrip is free software: you can redistribute it and/or modify
-%    it under the terms of the GNU General Public License as published by
-%    the Free Software Foundation, either version 3 of the License, or
-%    (at your option) any later version.
-%
-%    FieldTrip is distributed in the hope that it will be useful,
-%    but WITHOUT ANY WARRANTY; without even the implied warranty of
-%    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-%    GNU General Public License for more details.
-%
-%    You should have received a copy of the GNU General Public License
-%    along with FieldTrip. If not, see <http://www.gnu.org/licenses/>.
-%
-% $Id$
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% Please note that this function is called from many other FT functions. To avoid
-% unwanted recursion, you should avoid (where possible) calling other FT functions
-% inside this one.
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-if ~isfield(data, field) && isfield(data, 'avg') && isfield(data.avg, field)
-  field = ['avg.' field];
-elseif ~isfield(data, field) && isfield(data, 'trial') && isfield(data.trial, field)
-  field = ['trial.' field];
-elseif ~isfield(data, field)
-  ft_error('field "%s" not present in data', field);
-end
-
-if strncmp(field, 'avg.', 4)
-  prefix = '';
-  field = field(5:end); % strip the avg
-  data.(field) = data.avg.(field); % copy the avg into the main structure
-  data = rmfield(data, 'avg');
-elseif strncmp(field, 'trial.', 6)
-  prefix = '(rpt)_';
-  field = field(7:end); % strip the trial
-  data.(field) = data.trial(1).(field); % copy the first trial into the main structure
-  data = rmfield(data, 'trial');
-else
-  prefix = '';
-end
-
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% ATTEMPT 1: the specific dimord is simply present
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-if isfield(data, [field 'dimord'])
-  dimord = data.([field 'dimord']);
-  return
-end
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% if not present, we need some additional information about the data strucure
[…]
