### 2023-07-12 c15e2f3
Update DEPENDENCY list of test function: % DATA no/public/private
