+++ test/test_bug1397.m
@@ -3,6 +3,7 @@ function test_bug1397
 % MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_preprocessing ft_appenddata
+% DATA public
 
 % the following code was obtained from http://www.fieldtriptoolbox.org/tutorial/coherence
 % on Wed Mar 28 15:36:40 CEST 2012
