+++ test/test_notification.m
@@ -3,6 +3,7 @@ function test_notification
 % WALLTIME 00:10:00
 % MEM 1gb
 % DEPENDENCY ft_debug ft_info ft_notice ft_warning ft_error ft_notification
+% DATA no
 
 subfunction(@ft_debug)
 subfunction(@ft_info)
