+++ test/inspect_bug2721.m
@@ -3,6 +3,7 @@ function inspect_bug2721
 % WALLTIME 00:10:00
 % MEM 1gb
 % DEPENDENCY ft_multiplotTFR
+% DATA private
 
 load(dccnpath('/home/common/matlab/fieldtrip/data/test/bug2721.mat'));
 
