+++ test/test_bug1166.m
@@ -3,6 +3,7 @@ function test_bug1166
 % MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_prepare_headmodel ft_headmodel_bem_asa 
+% DATA no
 
 % This function tests that the inputs for the headmodel functions are
 % read-in correctly
