+++ test/test_ft_preproc_standardize.m
@@ -3,6 +3,7 @@ function tests = test_ft_preproc_standardize
 % MEM 1gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_preproc_standardize
+% DATA no
 
 if nargout
   % assume that this is called by RUNTESTS
