### 2023-07-12 4c88d71
Manually FIX test_example_entropy_analysis20220113.m
