### 2023-07-12 05e6beb
Manually FIX inspect_menu_viewpoint.m
