+++ test/test_example_bids20220113.m
@@ -1,8 +1,8 @@
 function test_example_bids
 
-% DATA no
 % MEM 4gb
 % WALLTIME 00:10:00
+% DATA no
 
 %
 %% BIDS - the brain imaging data structure
@@ -11,7 +11,6 @@ function test_example_bids
 %
 % FieldTrip can read data from all MEG, EEG, iEEG and MRI file formats that are supported in BIDS. Furthermore, it includes the **[data2bids](https://github.com/fieldtrip/fieldtrip/blob/release/data2bids.m)** function to convert data to the BIDS structure. This function can convert data formats (if needed), and writes the accompanying "sidecar files" that contain metadata about channels, coordinates, events, etcetera to the .json and .tsv format.
 %
-% DATA no
 % We recommend that you convert your raw data to BIDS *prior* to implementing your analyses scripts. That will make it much easier to share the scripts, the raw and the derived data.
 %
 % The BIDS project overview is presented on <http://bids.neuroimaging.io>. You can read the full specification on <https://bids-specification.readthedocs.io>. Please note that BIDS is not only relevant for FieldTrip, but that it also been embraced by the SPM, EEGLAB, MNE-Python and BrainStorm developers, and of course by the MRI and fMRI community, for example on <https://openneuro.org>.
