+++ test/test_example_bids_meg20220113.m
@@ -1,8 +1,8 @@
 function test_example_bids_meg
 
-% DATA no
 % MEM 4gb
 % WALLTIME 00:10:00
+% DATA no
 
 %
 %% Converting an example MEG dataset for sharing in BIDS
@@ -26,7 +26,6 @@ function test_example_bids_meg
 %
 % The data is anonimized using the `go_anonymiseDs.m` function from [this example page](/faq/how_can_i_anonymize_a_ctf_dataset/#using-matlab). This reads in the data with a low-level CTF function, scrubs potentially identifying  information, and writes it with a low-level CTF function. Some of the auxiliary files that can optionally be present (for example when the data was opened with the CTF DataEditor application) and the headlocalizer data will be ignored upon reading, and hence also not appear in the copy that is written.
 %
-% DATA no
 % The anonimized version of the CTF data has a short descriptive name. Using the **[data2bids](https://github.com/fieldtrip/fieldtrip/blob/release/data2bids.m)** we copy and rename the CTF data so that it has the right directory structure and file names, and we add the metadata.
 %
 %
