### 2023-07-12 f7da6f2
Manually FIX  test_example_coherence_snr20220113.m
