+++ test/test_example_samplesize20220113.m
@@ -1,8 +1,8 @@
 function test_example_samplesize
 
-% DATA no
 % MEM 8gb
 % WALLTIME 01:30:00
+% DATA no
 
 %
 %% Using simulations to estimate the sample size for cluster-based permutation test
@@ -13,16 +13,13 @@ function test_example_samplesize
 %
 % Here, we demonstrate two easy-to-use MATLAB functions that use simulations to estimate the sample size for cluster-based permutation tests. These functions can be used for EEG/MEG research involving **t-test** between two conditions, **one-way ANOVA** with three or more conditions, or **2×N interactions**. The experiment can be of a with-subjects, between-subjects, or mixed design. The functions were written for, and first used in [Wang and Zhang (2021)](https://doi.org/10.1111/psyp.13775). Please cite this paper where appropriate.
 %
-% DATA no
 % From this [OSF project](https://osf.io/rmqhc/files/), you can download the functions and the corresponding demo files. The MATLAB functions are stored in the _functions_ folder.
 %
-% DATA no
 %*   **sampleSize_erp.m**  is for ERP data
 %*   **sampleSize_timefreq.m**  is for time-frequency data
 %
 % Two demo files demonstrating how to use the functions are in the _demo_ folder:
 %
-% DATA no
 %*   **demo_erp.m**
 %*   **demo_timefreq.m**
 %
@@ -59,12 +56,10 @@ unzip('https://files.osf.io/v1/resources/rmqhc/providers/osfstorage/60347b1e88ea
 addpath(t);
 % JM added: one of the downloaded functions requires the financial
 % toolbox's corr2cov function. This is such a no-brainer that it's created
-% DATA no
 % here on the fly
 fid=fopen(fullfile(t,'corr2cov.m'),'w');
 fprintf(fid,'function out = corr2cov(sd,cor)\nout=diag(sd)*cor*diag(sd);\n');
 fclose(fid);
-% DATA no
 
 % settings
 p_heads_h1  = 1;     % chance to land on head under H1, aka, the effect size
@@ -171,18 +166,14 @@ title(['Required sample size is ' num2str(n_sample)])
 %
 % Then a cluster-based permutation test is performed on the simulated dataset to test whether there is a significant difference between the two conditions. We run the simulations for 1000 times, and the power is calculated as the proportion of the number of times that the null hypothesis is rejected. These simulations can subsequently be repeated for an incrementally increasing sample size, starting from 10, increasing in steps of 1, until the power reached the desired threshold (e.g., 0.8). The above processing is implemented in the two aforementioned MATLAB functions `sampleSize_erp.m` and `sampleSize_timefreq.m`.
 %
-% DATA no
 % Next, we will demonstrate how to use the `sampleSize_timefreq.m` function; the usage of `sampleSize_erp.m` is very similar. See also the two demo files.
 %
-% DATA no
 % First, you need to have a time-frequency dataset that is generated by the **[ft_freqanalysis](http://github.com/fieldtrip/fieldtrip/blob/release/ft_freqanalysis.m)** function. You can also use the _exempleData_timefreq.mat_ dataset which can be downloaded with the functions from [OSF](https://osf.io/rmqhc). This dataset is used only for retrieving the data structure for simulating time-frequency data. The `sampleSize_timefreq.m` function has two parts of parameters, which are respectively stored in |cfg| and `stat_cfg`. The structure |cfg| stores the configurations for simulating the data, and `stat_cfg` stores the configurations for the cluster-based permutation tests.
 %
-% DATA no
 %%
 close all;
 addpath('F:\SampleSize\functions')
 load('exampleData_timefreq.mat');   % load a time-freq data obtained from the ft_freqanalysis function,
-% DATA no
                                     % to retrieve the fieldtrip data structure
                                     
[…]
