### 2023-07-12 f8fc15a
Manually FIX test_example_nirs_fingertapping20220113.m
