+++ test/test_example_ft_realtime_hilbert20220113.m
@@ -1,8 +1,8 @@
 function test_example_ft_realtime_hilbert
 
-% DATA no
 % MEM 4gb
 % WALLTIME 00:10:00
+% DATA no
 
 %
 %% Realtime neurofeedback application based on Hilbert phase estimation
@@ -26,21 +26,18 @@ function test_example_ft_realtime_hilbert
 %
 % Now it will start the function realtime_baseline that for 2 minutes will record the subject baseline and store the results for realtime_hilbert.
 % Basically there are 12 channels: 4 for EEG (I used F3-P3-F4-P4), 4 for the EOG, 4 for the EMG (front and neck). In this script, those are selected from the forty originally recorded channels you can find in the variable 'lab'.
-% DATA no
 % The algorithm will care to take away both EOG and EMG artifacts.
 %
 %% # MATLAB code for ft_realtime_hilbert
 %
 function ft_realtime_hilbert()
 
-% DATA no
 % FT_REALTIME_HILBERT is a neurofeedback application based on Hilbert phase estimation.
 %
 % Use as
 %   ft_realtime_hilbert()
 % with the following configuration options that are coded inside the function
 %   cfg.channel    = cell-array, see FT_CHANNELSELECTION (default = 'gui')
-% DATA no
 %   cfg.foilim     = [Flow Fhigh] (default = [0 120])
 %   cfg.blocksize  = number, size of the blocks/chuncks that are processed (default = 1 second)
 %   cfg.bufferdata = whether to start on the 'first or 'last' data that is available (default = 'first')
@@ -58,7 +55,6 @@ function ft_realtime_hilbert()
 %
 % To stop the realtime function, you have to press Ctrl-C
 
-% DATA no
 % Take off the warning message to avoid problems with ATAN2 and hilbert.
 warning off all;
 
@@ -411,14 +407,12 @@ end % while true
 %
 function ft_realtime_baseline()
 
-% DATA no
 % FT_REALTIME_BASELINE is a neurofeedback application based on Hilbert phase estimation.
 %
 % Use as
 %   ft_realtime_baseline()
 % with the following configuration options that are coded inside the function
 %   cfg.channel    = cell-array, see FT_CHANNELSELECTION (default = 'gui')
-% DATA no
 %   cfg.foilim     = [Flow Fhigh] (default = [0 120])
 %   cfg.blocksize  = number, size of the blocks/chuncks that are processed (default = 1 second)
 %   cfg.bufferdata = whether to start on the 'first or 'last' data that is available (default = 'first')
@@ -436,7 +430,6 @@ function ft_realtime_baseline()
 %
 % To stop the realtime function, you have to press Ctrl-C
 
-% DATA no
 %Take off the warning message to avoid problems with ATAN2 and hilbert.
 warning off all;
 
