### 2023-07-12 c15d054
Manually FIX test_example_compute_leadfield20220113.m
