### 2023-07-12 397cf1a
Manually FIX trialfun_affcog.m
