### 2023-07-12 cd582b5
Manually FIX test_example_stratify20220113.m
