### 2023-07-12 ad75c7c
Manually FIX test_example_compute_forward_simulated_data20220113.m
