### 2023-07-12 c6a15d0
Manually FIX test_example_combined_eeg_and_meg_source_reconstruction20220113.m
