+++ test/test_example_crossfreq20220113.m
@@ -1,8 +1,8 @@
 function test_example_crossfreq
 
-% DATA no
 % MEM 4gb
 % WALLTIME 00:10:00
+% DATA no
 
 %
 %% Cross-frequency analysis
@@ -12,7 +12,6 @@ function test_example_crossfreq
 %
 % With the **[ft_freqsimulation](https://github.com/fieldtrip/fieldtrip/blob/release/ft_freqsimulation.m)** function you can generate simulated data in FieldTrip format which the different types of cross-frequency interactions. The different methods are:
 %
-% DATA no
 %* [phalow_amphigh (is phase to power in Jensen and Colgin)](/example/crossfreq/phalow_amphigh)
 %* [amplow_amphigh (is power to power in Jensen and Colgin](/example/crossfreq/amplow_amphigh)
 %* [phalow_freqhigh (is phase to frequency in Jensen and Colgin)](/example/crossfreq/phalow_freqhigh)
