### 2023-07-12 605bd57
Manually FIX test_example_determine_the_filter_characteristics20220113.m
