+++ test/test_example_determine_the_filter_characteristics20220113.m
@@ -1,15 +1,14 @@
 function test_example_determine_the_filter_characteristics
 
-% DATA no
 % MEM 4gb
 % WALLTIME 00:10:00
+% DATA no
 
 %
 %% Determine the filter characteristics
 %
 % The following script demonstrates how you can determine the filter characteristics. Better ways than those described here are available, but this will give you an idea on how to approach the filter details. This snippet of code generates a one second piece of data with a delta function in it (i.e. a spike). This signal is passed though the **[ft_preproc_lowpassfilter](https://github.com/fieldtrip/fieldtrip/blob/release/preproc/ft_preproc_lowpassfilter.m)** function (located in fieldtrip/preproc) and the result is plotted in the time domain (left panel) and frequency domain (right panel).
 %
-% DATA no
 % generate the signal and the corresponding time axis
 s = zeros(1,1000); s(500) = 1; s = s - mean(s);
 t = 0.001:0.001:1;
