### 2023-07-12 538514f
Manually FIX test_example_glm_trials20220113.m
