### 2023-07-12 285d81d
revert trialfun_affcog.m
this is not a test script itself
