### 2023-07-12 5e92cc2
Revert trialfun_stimon.m
this is not a test script itself
