+++ test/trialfun_stimon.m
@@ -1,6 +1,5 @@
 function trl = trialfun_stimon(cfg)
- 
-% DATA no
+
 hdr   = ft_read_header(cfg.dataset);
 event = ft_read_event(cfg.dataset);
 correctresponse  = 10041;
