### 2023-07-12 82ac9ab
Revert trialfun_stimon_samples.m
this is not a test script itself
