### 2023-07-12 23d1083
Update DEPENDENCY list of test functions by adding a new field in the test description for data. (#2278)
* Update DEPENDENCY list of test function with DATA no/public/private
* Manually FIX test_example_entropy_analysis20220113.m
* Manually FIX inspect_menu_viewpoint.m
* Manually FIX test_example_bids20220113.m
* Manually FIX test_example_bids_meg20220113.m
* Manually FIX  test_example_coherence_snr20220113.m
* Manually FIX test_example_neighbours20220113.m
* Manually FIX test_example_samplesize20220113.m
* Manually FIX test_example_nirs_fingertapping20220113.m
* Manually FIX test_example_ft_r
[…]
