+++ test/test_bug1262.m
@@ -3,6 +3,7 @@ function test_bug1262
 % MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_read_header ft_read_data ft_read_event
+% DATA private
 
 % Quick and dirty sanity check for reading of fcdc_buffer_offline data.
 
