+++ test/test_example_neighbours20220113.m
@@ -2,6 +2,7 @@ function test_example_neighbours
 
 % MEM 4gb
 % WALLTIME 00:10:00
+% DATA no
 
 %
 %% Defining electrodes as neighbours for cluster-level statistics
