+++ test/test_bug2397.m
@@ -3,6 +3,7 @@ function test_bug2397
 % MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_prepare_mesh
+% DATA private
 
 % this pertains to the OpenMEEG pipeline, see http://bugzilla.fieldtriptoolbox.org/show_bug.cgi?id=2397
 
