+++ test/test_ft_specest_hilbert.m
@@ -3,6 +3,7 @@ function tests = test_ft_specest_hilbert
 % MEM 1gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_specest_hilbert
+% DATA no
 
 if nargout
   % assume that this is called by RUNTESTS
