+++ test/test_bug1357.m
@@ -3,6 +3,7 @@ function test_bug1357
 % MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_multiplotTFR ft_singleplotTFR ft_singleplotTFR ft_singleplotTFR
+% DATA no
 
 cfg = [];
 cfg.layout = 'CTF151.lay';
