+++ test/test_tutorial_natmeg2014_statistics.m
@@ -3,6 +3,7 @@ function test_tutorial_natmeg2014_statistics
 % WALLTIME 00:20:00
 % MEM 4gb
 % DEPENDENCY
+% DATA public
 
 % this script executes the MATLAB content from
 % http://www.fieldtriptoolbox.org/tutorial/natmeg2014/statistics
