+++ test/test_warning_once.m
@@ -3,6 +3,7 @@ function test_warning_once
 % MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY
+% DATA no
 
 warning1 = 'hululu';
 warning2 = 'aloah hey';
