+++ test/test_bug3359.m
@@ -3,6 +3,7 @@ function test_bug3359
 % MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_connectivityanalysis ft_topoplotER 
+% DATA private
 
 % first create some data
 %--------------------------------------------------------
