+++ test/test_bug1043.m
@@ -3,6 +3,7 @@ function test_bug1043
 % MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_megplanar ft_apply_montage yokogawa2grad channelposition ft_apply_transform
+% DATA private
 
 % the bug consists of ft_apply_montage not adequately dealing with
 % sensor descriptions that contain coilori/pos chanori/pos.
