### 2023-07-13 c8f085d
Updated alternative2 to work also on Windows
