### 2023-07-14 54af97c
improve GeneratedBy, see #2277
- change default for GeneratedBy
- update cfg fields that should be a list-of-strings
- update cfg fields that should be a list-of-objects
