### 2023-07-15 bdd6469
FIX - merging the GeneratedBy failed due to it not being represented as an array of objects. This commit adds the fix_dataset_description function which is called both on the new and the extsting structure.
