### 2023-07-15 a63bc0a
This commit consists of two changes: 1. dccnpath is updated to automatically download files and folders to the local copy, if they are not already downloaded. To download a whole folder, dccnpath uses the function utilities/private/recursiveDownload.m, 2. test_dccnpath is created to test the old and the newly added functionality
