### 2023-07-18 c8b3a6c
New features added to the dccnpath() and test_dccnpath was updated: 1. when ft_default.dccnpath is not specified the data are automatically downloaded to the tempdir 2. dccnpath() always outputs a path, even for alternative1 3. warning for private data and automatic errors given by webread for typos in the path.
