+++ utilities/dccnpath.m
@@ -40,6 +40,7 @@ function filename = dccnpath(filename)
 
 global ft_default
 
+
 % alternative0 allows to test when in the DCCN cluster
 if ispc
   alternative0 = strrep(filename,'/home','H:');
@@ -49,6 +50,14 @@ else
   alternative0 = strrep(alternative0,'\','/');
 end
 
+if exist(alternative0, 'file') || exist(alternative0, 'dir')
+    warning('using default DCCN path %s', alternative0);
+    filename = alternative0;
+    return;
+end
+
+
+
 % alternative1 allows to test with local files in the present working directory
 % this is often convenient when initially setting up a new test script while the data is not yet uploaded
 [p, f, x] = fileparts(alternative0);
@@ -57,11 +66,38 @@ alternative1 = [f x];
 if strcmp(alternative1, 'test')
   % this should not be used when the filename is only "test", since that is too generic
   alternative1 = '';
+  warning('You are currently in the test directory, change directory so dccnpath can function properly')
 elseif endsWith(pwd, alternative1)
   % exist(alternative1, 'file') also returns 7 in case the present working directory name matches alternative1
   % this is not the match we want
   alternative1 = '';
 end
+
+if exist(alternative1, 'file') || exist(alternative1, 'dir') 
+    if ~isempty(x) % if alternative1 is a file
+      filenamepath = which(alternative1); % output the path that alternative1 is located at
+      warning('using present working directory %s', filenamepath);
+      filename = filenamepath;
+      return;
+    else  % if alternative1 is a folder
+        searchPath = path; % Get the MATLAB search path
+        directories = strsplit(searchPath, pathsep); % Split the search path into individual directories
+        
+        folderPath = '';
+        for i = 1:numel(directories)
+            currentFolder = directories{i};
+            if exist(fullfile(currentFolder, alternative1), 'dir') == 7
+                folderPath = fullfile(currentFolder, alternative1);
+                break;
+            end
+        end
+        warning('using present working directory %s', folderPath);
+        filename = folderPath;
+        return;
+    end 
+end
+
+
 
 % alternative2 allows for the user to specify the path in the global ft_default variable
 % see https://github.com/fieldtrip/fieldtrip/issues/1998
@@ -71,49 +107,60 @@ if isfield(ft_default, 'dccnpath')
     else
         alternative2 = strrep(alternative0, 'H:\common\matlab\fieldtrip', ft_default.dccnpath);
     end
-else
-  alternative2 = '';
-end
-
-if exist(alternative1, 'file')
-  warning('using present working directory %s', alternative1);
-  filename = alternative1;
-elseif isfield(ft_default, 'dccnpath')
-  warning('using local copy %s ', alternative2);
-
-  % Check if filename exists. If not, download it (Todo: after permission)    
-  if ~exist(alternative2, 'file')
- 
-     % Public data are downloaded from https://download.fieldtriptoolbox.org
-     if contains(alternative2, 'ftp')
-        % Find the right path to the HTTPS download server
-        pattern = 'ftp(.*)';                      
-        datadir = regexp(alternative2, pattern, 'tokens', 'once');          
-        weblocation = strcat("https://download.fieldtriptoolbox.org", datadir);
-        weblocation = strrep(weblocation, '\', '/');
-        
-        urlContent = webread(weblocation, weboptions('ContentType', 'text')); % Runs a bit slow for large files
-    
-       % Check if the URL corresponds to a folder or file
-       if contains(urlContent, '<html') % Folder (HTML content) 
-              recursiveDownload(weblocation, alternative2);
-       else % File
-
-              % Create the necessary directories if they do not exist
[…]
