+++ test/test_dccnpath.m
@@ -5,98 +5,112 @@ function test_dccnpath
 % DEPENDENCY
 % DATA public private
 
-%% Alternative0
+%% Alternative0: Finds the right path for users that have access to the DCCN intranet. It will replace '/home' by 'H:' and will replace forward by backward slashes.
 
-global ft_default;
+% % This can only run by people that have access to the DCCN intranet
+% filename=dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf'); 
+%     if ispc
+%         if ~strcmp(filename, 'H:\common\matlab\fieldtrip\data\ftp\test\edf\testAlphaIR20170321-0.edf')
+%             error('Alternative0 does not work');
+%         end
+%     else
+%         if ~strcmp(filename, '/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf')
+%             error('Alternative0 does not work');
+%         end
+%     end
 
-filename=dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf');
-if ispc
-    if ~strcmp(filename, 'H:\common\matlab\fieldtrip\data\ftp\test\edf\testAlphaIR20170321-0.edf')
-        error('Alternative0 does not work');
-    end
-else
-    if ~strcmp(filename, '/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf')
-        error('Alternative0 does not work');
-    end
+% This can work for everyone
+filename=dccnpath(tempdir);
+if ~strcmp(filename, tempdir)
+     error('dccnpath() does not find the right DCCN path');
 end
 
+%% Alternative1: Allows to test with local files in the present working directory
 
-%% Alternative1
-
-ft_defaults
+ft_defaults % '/home/common/matlab/fieldtrip/data/ftp/test/fileio' now belongs to the present working directory
 filename=dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/fileio');
-if ~strcmp(filename, 'fileio')
-        error('Alternative0 does not work');
+if ~contains(filename, 'fileio')
+        error('dccnpath() does not find the right working directory');
 end
 
 
-%% Alternative2
-
-% Create the right paths inside tempdir in order for alternative2 to work
-global ft_default;
-ft_default.dccnpath = tempdir;
-ft_default.dccnpath = ft_default.dccnpath(1:end-1); % we do not want the temppdir to end with a '/' or '\'
+%% Alternative2: It downloads test data to the local computer, if test data is not already downloaded. It has 6 different tests that are listed below
 
 
-if ispc
-    path=strcat(ft_default.dccnpath,'\data\ftp');
-else
-    path=strcat(ft_default.dccnpath,'/data/ftp');
-end
-if ~isfolder(path) % make the directories data/ftp, if they do not exist in the temp path
-    mkdir(path);
-end
+%% Test1: When ft_default.dccnpath is given by the user, dccnpath() should always select alternative2 (this is not done now)
 
-
-%% Test1: When ft_default.dccnpath is given by the user, dccnpath() should select alternative2
-
-if ~isempty(ft_default.dccnpath)
-    filename=dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf');
-    if ~contains(filename, ft_default.dccnpath)
-            error('The alternative2 was not selected while it should have');
-    end
-end
-
-delete(filename) 
+% if ~isempty(ft_default.dccnpath)
[…]
