### 2023-08-04 #5566 @ashahide
@crsl4, I was wondering the same thing about the preprint. Should the review comments be focused only on the material in the JOSS article, or should we also consider the additional preprint?

Thanks for your help.
