### 2023-08-04 #5566 @crsl4
@ashahide We are supposed to only review the JOSS article, but we should understand what is the gap to the pre-print, and whether the pre-print has significant overlap (bc then, why do we have the JOSS article?).
