### 2023-08-04 #5566 @crsl4
@gflofst Thanks for your message! I think I want to discuss this with other editors. I am not sure what the standard expectations are compared to other articles. I will reach back, thanks!
