### 2023-08-07 56038ac
FIX #2084 - allow cfg.README and added template
this also adds support for cfg.LICENSE and cfg.CHANGES
