+++ data2bids.m
@@ -66,16 +66,51 @@ function [cfg] = data2bids(cfg, varargin)
 %   cfg.space                   = string
 %   cfg.desc                    = string
 %
-% When specifying the output directory in cfg.bidsroot, you can also specify
-% additional information to be added as extra columns in the participants.tsv and
-% scans.tsv files. For example:
-%   cfg.participants.age        = scalar
-%   cfg.participants.sex        = string, 'm' or 'f'
-%   cfg.scans.acq_time          = string, should be formatted according to RFC3339 as '2019-05-22T15:13:38'
-%   cfg.sessions.acq_time       = string, should be formatted according to RFC3339 as '2019-05-22T15:13:38'
-%   cfg.sessions.pathology      = string, recommended when different from healthy
+% General BIDS options that apply to all data types are
+%   cfg.InstitutionName             = string
+%   cfg.InstitutionAddress          = string
+%   cfg.InstitutionalDepartmentName = string
+%   cfg.Manufacturer                = string
+%   cfg.ManufacturersModelName      = string
+%   cfg.DeviceSerialNumber          = string
+%   cfg.SoftwareVersions            = string
+%
+% General BIDS options that apply to all functional data types are
+%   cfg.TaskName                    = string
+%   cfg.TaskDescription             = string
+%   cfg.Instructions                = string
+%   cfg.CogAtlasID                  = string
+%   cfg.CogPOID                     = string
+%
+% You can specify information to be added in the corresponding columns in the
+% participants.tsv, scans.tsv, and sessions.tsv files. For example:
+%   cfg.participants.age            = scalar
+%   cfg.participants.sex            = string, 'm' or 'f'
+%   cfg.scans.acq_time              = string, should be formatted according to RFC3339 as '2019-05-22T15:13:38'
+%   cfg.sessions.acq_time           = string, should be formatted according to RFC3339 as '2019-05-22T15:13:38'
+%   cfg.sessions.pathology          = string, recommended when different from healthy
 % In case any of these values is specified as empty (i.e. []) or as nan, it will be
 % written to the tsv file as 'n/a'.
+%
+% If you specify cfg.bidsroot, this function will also write the dataset_description.json
+% file. You can specify the following fields
+%   cfg.dataset_description.writesidecar        = string
+%   cfg.dataset_description.Name                = string
+%   cfg.dataset_description.BIDSVersion         = string
+%   cfg.dataset_description.License             = string
+%   cfg.dataset_description.Authors             = string or cell-array of strings
+%   cfg.dataset_description.Acknowledgements    = string
+%   cfg.dataset_description.HowToAcknowledge    = string
+%   cfg.dataset_description.Funding             = string or cell-array of strings
+%   cfg.dataset_description.ReferencesAndLinks  = string or cell-array of strings
+%   cfg.dataset_description.DatasetDOI          = string
+%
+% If you specify cfg.bidsroot, this function can also write some modality agnostic
+% files at the top-level of the dataset. You can specify their content here and/or
+% subsequently edit them with a text editor.
+%   cfg.README                  = string (default is a template with instructions)
+%   cfg.LICENSE                 = string (no default)
+%   cfg.CHANGES                 = string (no default)
 %
 % For anatomical and functional MRI data you can specify cfg.dicomfile to read the
 % detailed MRI scanner and sequence details from the header of that DICOM file. This
[…]
