### 2023-08-08 0cf402a
ENH - fix numerical rounding error for zero padding (#2267)
