### 2023-08-09 #5566 @matsvanes
> Hi @matsvanes , is there a list of the required Matlab toolboxes somewhere that I may be missing? I see that there is an external folder for additional non-Matlab dependencies and that some of the tests check for necessary toolboxes, but I'm unable to find a list in the documentation.

Hi @ashahide, the list of external toolboxes can be found [here](https://www.fieldtriptoolbox.org/faq/requirements/). Most of these are supplied alongside FieldTrip (i.e., in `/external`), and are added to the path when running `ft_defaults` (see [here](https://www.fieldtriptoolbox.org/faq/should_i_add_fieldtrip_with_all_subdirectories_to_my_matlab_path/) for more info on installation)
