### 2023-08-09 #5566 @matsvanes
> @crsl4, We have a question as to whether the submission meets the required significant effort bar. In the submitted materials, the author is not in the code repo at all, but was responsible for writing some of the associated papers, provided after the fact. If this is sufficient, I think we need to have the repo updated with links/copies of the additional materials to be able to check that review criteria as satisfied. I am waiting for your guidance on whether what is in the repo is sufficient, if the additional materials existing are sufficient, if the existing materials need to be added/linked in the repo to be sufficient, or if it is insufficient given the lack of code contributions.

I hope I can clarify these questions with some historic perspective/context.
This project was part of my PhD - a time when I unfortunately was rather new to git(hub). Because of this, my FieldTrip branch was messy, and I used the quick route through my collaborators (co-authors on this paper) to get the code on the master branch. As mentioned [before](https://github.com/openjournals/joss-reviews/issues/5566#issuecomment-1619999946), I have also done most of the testing and documentation of the code.
Regarding the pre-print: this is essentially one chapter out of my [PhD thesis](https://repository.ubn.ru.nl/handle/2066/227394). We were interested in submitting to JOSS because we believe its [aims](https://www.arfon.org/announcing-the-journal-of-open-source-software) align well with our paper, and we wanted to change the narrative from a tutorial-style opinion/recommendation paper to put the focus on the software implementation itself (and get credit for it). Consequently, after opening a [pre-submission enquiry](https://github.com/openjournals/joss/issues/1109), we adapted the pre-print to the JOSS format, and moved the tutorials to the website documentation: [1](https://github.com/fieldtrip/website/blob/master/example/reproducescript.md), [2](https://github.com/fieldtrip/website/blob/master/example/reproducescript_group.md), [3](https://github.com/fieldtrip/website/blob/master/example/reproducescript_andersen.md).
