### 2023-08-09 #5566 @robertoostenveld
> @matsvanes does not appear to have made any contributions to this code base. Am I looking at this wrong?

Let me as the head of the FieldTrip project and senior author on this submission confirm Mats' [comment](https://github.com/openjournals/joss-reviews/issues/5566#issuecomment-1670960283): Mats did indeed make significant contributions to the conceptual design, the implementation, the testing, and the the documentation (initially in the form of a manuscript that was rejected elsewhere, later also on the [website](https://www.fieldtriptoolbox.org/tag/reproducescript/)). He fully deserves credit for this as first author.

My own involvement was mainly to supervise the process, to ensure consistency and future maintainability, and to help with the technical aspects of the coding and merging process. Since we are neuroscientists and neither of us are git and GitHub experts, I realize that the representation on github is not as ideal as it could have been. We continue to constantly learn how to make use of version control using git, and also how to pass those skills on to students that are working on concrete projects which involve coding (like this `reproducescript` one).
