### 2023-08-10 8f52c0c
ENH - allow for multiframe dicom reading, use SPM by default
