+++ fileio/ft_read_mri.m
@@ -30,7 +30,9 @@ function [mri] = ft_read_mri(filename, varargin)
 %   'ctf_mri'
 %   'ctf_mri4'
 %   'ctf_svl'
-%   'dicom'                      uses FreeSurfer code
+%   'dicom'                      uses SPM code, same as dicom_spm
+%   'dicom_spm'                  uses SPM code, also allows for multiframe dicoms (one volume in a file)     
+%   'dicom_fs'                   uses FreeSurfer code, which relies on MATLAB image processing toolbox
 %   'dicom_old'                  uses MATLAB image processing toolbox code
 %   'freesurfer_mgh'             uses FreeSurfer code
 %   'freesurfer_mgz'             uses FreeSurfer code
@@ -39,10 +41,10 @@ function [mri] = ft_read_mri(filename, varargin)
 %   'matlab'                     assumes a MATLAB *.mat file containing a struct
 %   'minc'                       uses SPM, this requires SPM5 or older
 %   'mrtrix_mif'                 uses mrtrix code
-%   'neuromag_fif'               uses MNE toolbox
+%   'neuromag_fif'               uses MNE toolbox code
 %   'neuromag_fif_old'           uses meg-pd toolbox
 %   'nifti'                      uses FreeSurfer code
-%   'nifti_spm'                  uses SPM
+%   'nifti_spm'                  uses SPM code
 %   'seg3d_mat'                  MATLAB file from Seg3D with a scirunnrrd structure
 %   'yokogawa_mri'
 %
@@ -461,7 +463,37 @@ switch dataformat
     hdr.coords = coords;
     hdr.dev    = dev;
 
-  case 'dicom'
+  case {'dicom' 'dicom_spm'}
+    % requires spm12
+    ft_hastoolbox('spm12', 1);
+    dicomhdr = spm_dicom_headers(filename); % read the header for a single file
+    if isfield(dicomhdr{1}, 'SharedFunctionalGroupSequence')
+      % this seems to be a multiframe image, we don't need to read in the separate headers for each of the slices
+    else
+      % this is a 'traditional' image, one file per slice
+      [f, p, e] = fileparts(filename);
+      d         = dir(fullfile(f,sprintf('*%s',e)));
+      filenames = {d.name};
+      for k = 1:numel(filenames)
+        filenames{k} = fullfile(f, filenames{k});
+      end
+      hdr1 = spm_dicom_headers(filename);  % read the header for the predefined slice
+      hdrs = spm_dicom_headers(filenames); % read the headers for all slices in the folder 
+      for k = 1:numel(hdrs)
+        seriesnumber(k,1) = hdrs{k}.SeriesNumber;
+      end
+      sel      = seriesnumber==hdr1{1}.SeriesNumber;
+      dicomhdr = hdrs(sel);
+    end
+    tdir  = tempdir;
+    fname = spm_dicom_convert(dicomhdr, 'all', 'flat', 'nii', tdir, 1);
+    mri   = ft_read_mri(fname.files{1});
+    mri.coordsys = 'scanras';
+    mri.unit     = 'mm';
+    mri.hdr      = dicomhdr;
+    delete(fname.files{1});
+    
+  case 'dicom_fs'
     % this returns a right-handed volume with the transformation matrix stored in the file headers
     % see https://github.com/fieldtrip/website/pull/444
 
