+++ test/test_ft_read_mri.m
@@ -49,3 +49,17 @@ assert(isequal(mri0.anatomy, mri2.anatomy));
 % they should all be the same
 assert(isequal(mri0.transform, mri1.transform));
 assert(isequal(mri0.transform, mri2.transform));
+
+% test the behavior of the different dicom readers
+filename = fullfile(datadir,  'dicom/19112010_JHORSCHIG.MR.FCDC_SEQUENCES_STANDARD_SEQUENCES.0002.0064.2010.11.19.12.08.01.265625.73005239.IMA');
+mri_spm = ft_read_mri(filename, 'dataformat', 'dicom_spm');
+mri_fs  = ft_read_mri(filename, 'dataformat', 'dicom_fs');
+
+[ftver, ftpath] = ft_version;
+pwdir = pwd;
+cd(fullfile(ftpath,'private'));
+mri_spm = align_ijk2xyz(mri_spm);
+mri_fs  = align_ijk2xyz(mri_fs);
+[ok,message]=isalmostequal(rmfield(mri_spm, 'hdr'), rmfield(mri_fs, 'hdr'));
+assert(ok);
+cd(pwdir);
