### 2023-08-11 cae4823
FIX - some coulance with respect to numeric difference
