### 2023-08-11 298c7c2
FIX - return dicom hdr as struct(-array) for dicom_spm dataformat
