+++ fileio/ft_read_mri.m
@@ -487,10 +487,18 @@ switch dataformat
     end
     tdir  = tempdir;
     fname = spm_dicom_convert(dicomhdr, 'all', 'flat', 'nii', tdir, 1);
+    
+    % convert  the cell(-array) dicomhdr as generated by SPM into a
+    % struct(-array)
+    for k = 1:numel(dicomhdr)
+      dicomhdr_struct(k) = dicomhdr{k};
+    end
+    clear dicomhdr;
+    
     mri   = ft_read_mri(fname.files{1});
     mri.coordsys = 'scanras';
     mri.unit     = 'mm';
-    mri.hdr      = dicomhdr;
+    mri.hdr      = dicomhdr_struct;
     delete(fname.files{1});
     
   case 'dicom_fs'
