### 2023-08-16 c8f4f98
[DATALAD RUNCMD] fix ambigous typos in utilities manually
=== Do not change lines below ===
{
 "chain": [],
 "cmd": "codespell -i3 -C4 -w utilities",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^
