### 2023-08-17 8c26cb9
[DATALAD RUNCMD] run codespell on utilities
=== Do not change lines below ===
{
 "chain": [],
 "cmd": "codespell -w utilities",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^
