+++ utilities/ft_selectdata.m
@@ -509,7 +509,7 @@ end
 switch selmode
   case 'intersect'
     if iscell(selindx)
-      % there are multiple selections in multipe vectors, the selection is in the matrices contained within the cell-array
+      % there are multiple selections in multiple vectors, the selection is in the matrices contained within the cell-array
       for j=1:numel(selindx)
         if ~isempty(selindx{j}) && all(isnan(selindx{j}))
           % no selection needs to be made
@@ -999,7 +999,7 @@ elseif numel(cfg.latency)==1
   end
   
 elseif numel(cfg.latency)==2
-  % the [min max] range can be specifed with +inf or -inf, but should
+  % the [min max] range can be specified with +inf or -inf, but should
   % at least partially overlap with the time axis of the input data
   mintime = min(alltimevec);
   maxtime = max(alltimevec);
@@ -1137,7 +1137,7 @@ if isfield(cfg, 'frequency')
     end
     
   elseif numel(cfg.frequency)==2
-    % the [min max] range can be specifed with +inf or -inf, but should
+    % the [min max] range can be specified with +inf or -inf, but should
     % at least partially overlap with the freq axis of the input data
     minfreq = min(freqaxis);
     maxfreq = max(freqaxis);
