+++ utilities/private/fixsource.m
@@ -84,7 +84,7 @@ else
 end
 
 
-% FIXME the code bellow should be repatterned
+% FIXME the code below should be repatterned
 % first get the dimord string
 % then determine the data dimensions
 % then determine the dimensions according to the dimord
