+++ README.md
@@ -1,3 +1,7 @@
+[![Platform: MATLAB](https://img.shields.io/badge/MATLAB-orange.svg?style=plastic)](https://www.mathworks.com)[![Last commit](https://img.shields.io/github/last-commit/fieldtrip/fieldtrip?style=plastic)](https://github.com/fieldtrip/fieldtrip)[![License: GPL-3.0](https://img.shields.io/badge/License-GPLv3-blue.svg?style=plastic)](https://www.gnu.org/licenses/gpl-3.0)
+
+# FieldTrip 
+
 FieldTrip is the MATLAB software toolbox for MEG and EEG analysis that is being
 developed at the Donders Institute for Brain, Cognition and Behaviour in Nijmegen,
 the Netherlands together with collaborating institutes.
@@ -6,12 +10,12 @@ The toolbox offers advanced analysis methods of MEG, EEG, and invasive
 electrophysiological data, such as time-frequency analysis, source reconstruction
 using dipoles, distributed sources and beamformers and non-parametric statistical
 testing. It supports the data formats of all major MEG systems (CTF,
-Neuromag/Elekta/Megin, BTi/4D, Yokogawa/Ricoh) and of most popular EEG systems, and
-new formats can be added easily. FieldTrip contains high-level functions that you
-can use to construct your own analysis protocols in MATLAB. Furthermore, it easily
-allows methods researchers to incorporate new methods for EEG/MEG analysis.
+Neuromag/Elekta/Megin, BTi/4D, Yokogawa/Ricoh, FieldLine) and of most popular EEG
+systems, and new formats can be added easily. FieldTrip contains high-level functions
+that you can use to construct your own analysis protocols in MATLAB. Furthermore, it
+easily allows methods researchers to incorporate new methods for EEG/MEG analysis.
 
-For more information please visit http://www.fieldtriptoolbox.org
+For more information, documentation, and tutorials, please visit http://www.fieldtriptoolbox.org
 
 The FieldTrip software is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -36,34 +40,25 @@ The functions in the FieldTrip toolbox are copyrighted by their respective autho
   - Roemer van der Meij, DCC 
   - Saskia Haegens, DCCN, FCDC
   - Vladimir Litvak, UCL
-  - and many others ...
+  - and many other contributors
 
-Copyright (C) 2008-2022, Donders Institute for Brain, Cognition and Behaviour, Radboud University, The Netherlands (DCCN, DCC, DCN)
+The authors are or were working at a range of academic institutions:
 
-Copyright (C) 2014-2018, Karolinska Institute, Stockholm, Sweden (NatMEG)
-
-Copyright (C) 2012-2018, Max Planck Institute for Psycholinguistics, The Netherlands (MPI)
-
-Copyright (C) 2008-2017, The Wellcome Trust Centre for Neuroimaging, University College London, UK (UCL)
-
-Copyright (C) 2010-2013, Swammerdam Institute for Life Sciences, University of Amsterdam (SILS)
-
-Copyright (C) 2008-2009, Centre for Cognitive Neuroimaging in Glasgow, United Kingdom (CCNi)
-
-Copyright (C) 2009-2009, Netherlands Institute for Neuroscience (NIN)
-
-Copyright (C) 2003-2008, F.C. Donders Centre, Radboud University Nijmegen, The Netherlands (FCDC)
-
-Copyright (C) 2004-2007, Nijmegen Institute for Cognition and Information, The Netherlands (NICI)
-
-Copyright (C) 2004-2005, Universitatsklinikum Hamburg-Eppendorf, Germany (UKE)
-
[…]
