### 2023-08-17 4975b7e
FIX - dccnpath(path/to/testdataFolder) was giving an error when testdataFolder did not exist. On the contrary, when I added a '/' in the end of the path, i.e., dccnpath(path/to/testdataFolder/) was not giving an error even if testdataFolder did not exist.
