### 2023-08-18 229f762
FIX - do not follow the path outline of the download server but the path outline given in ft_default.dccnpath. Based on this, test_dccnpath was updated accordingly
