+++ utilities/dccnpath.m
@@ -1,24 +1,36 @@
 function filename = dccnpath(filename)
 
-% DCCNPATH manages the filename and path for test files. It helps to locate and read
-% test file from Linux, Windows or macOS computers both inside and outside the DCCN.
-%
-% Use as
-%  filename = dccnpath(filename)
-%
-% The default location for FieldTrip and its test data is '/home/common/matlab/fieldtrip'.
-% This function will search-and-replace this string by the location that applies to
-% your computer. It will replace '/home' by 'H:' and will replace forward by backward slashes.
-%
-% In case you have a local copy, you can override the default location by
-%   global ft_default
-%   ft_default.dccnpath = '/your/copy';
-%
-% Note that most test scripts expect data located at /home/common/matlab/fieldtrip/data/ftp
-% or /home/common/matlab/fieldtrip/data/test, hence you should organize your local
-% copy of the data under /your/copy/data/ftp and /your/copy/data/test.
-%
-% Copyright (C) 2012-2022, Donders Centre for Cognitive Neuroimaging, Nijmegen, NL
+% DCCNPATH manages the filename and path for test files. It helps to locate and read test file from Linux, Windows or macOS computers both inside and outside the DCCN.
+% 
+% Use as 
+%   filename = dccnpath(filename)
+% 
+% Inputs:
+%   filename - The relative path to the test data file within the DCCN cluster.
+%              Examples:
+%              - For public test data: '/home/common/matlab/fieldtrip/data/ftp/testdata' 
+%              - For private test data: '/home/common/matlab/fieldtrip/data/test/testdata' 
+% 
+% Outputs:
+%   filename - string that corresponds to the local path that the test data
+%   are saved.
+% 
+% The default location for FieldTrip and its test data is '/home/common/matlab/fieldtrip'. This function will search-and-replace this string by the location that applies to your computer. It will replace '/home' by 'H:' and will replace forward by backward slashes. 
+% 
+% In case you have a local copy:
+% 
+% - You should override the default location by 
+%    global ft_default 
+%    ft_default.dccnpath = '/your/copy'; 
+% 
+%    If you do not define ft_default.dccnpath manually then dccnpath will
+%    automatically use: 
+%    ft_default.dccnpath = tempdir;
+% 
+% - If the test data are not downloaded inside '/your/copy', then dccnpath
+%   will automatically download them.
+% 
+% Copyright (C) 2012-2023, Donders Centre for Cognitive Neuroimaging, Nijmegen, NL
 
 % This file is part of FieldTrip, see http://www.fieldtriptoolbox.org
 % for the documentation and details.
@@ -101,12 +113,8 @@ end
 
 % alternative2 allows for the user to specify the path in the global ft_default variable
 % see https://github.com/fieldtrip/fieldtrip/issues/1998
-if isfield(ft_default, 'dccnpath')
-    if ~ispc
-        alternative2 = strrep(alternative0, '/home/common/matlab/fieldtrip', ft_default.dccnpath);
-    else
-        alternative2 = strrep(alternative0, 'H:\common\matlab\fieldtrip', ft_default.dccnpath);
-    end
+if isfield(ft_default, 'dccnpath')    
+        alternative2 = ft_default.dccnpath;      
 else % save the test data automatically in the tempdir
     ft_default.dccnpath = tempdir;
     ft_default.dccnpath = ft_default.dccnpath(1:end-1); % we do not want the temppdir to end with a '/' or '\'   
@@ -117,23 +125,22 @@ else % save the test data automatically in the tempdir
     end
 end
 
+% if folder or file doesn't exist or is empty, then download test data
+dirContents = dir(alternative2);
[…]
