### 2023-08-18 9ad98e3
FIX - test_dccnpath to work in the DCCN computer cluster
