### 2023-08-18 0b2a9a5
Update README.md
added additional badges from https://shields.io/badges
