### 2023-08-21 09377c7
FIX - present working directory output of dccnpath. Also, I run 15 tests with public data in my local laptop and I did minor changes to 3 of them to work correctly together with dccnpath
