+++ test/test_bug1254.m
@@ -2,12 +2,12 @@ function test_bug1254
 
 % MEM 2gb
 % WALLTIME 00:10:00
+% DEPENDENCY qsubcompile qsubcellfun
+% DATA public
 
 % the bug has not been fixed yet, so there is no point in automatically
 % executing this test
 return
-% DEPENDENCY qsubcompile qsubcellfun
-% DATA public
 
 dataset = dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/ctf/Subject01.ds');
 
