### 2023-08-22 9292ec4
Added a test for finding the right path in the DCCN cluster
