+++ test/test_dccnpath.m
@@ -5,25 +5,26 @@ function test_dccnpath
 % DEPENDENCY dccnpath
 % DATA public
 
-%% Alternative0: Finds the right path for users that have access to the DCCN intranet. It will replace '/home' by 'H:' and will replace forward by backward slashes.
+%% Alternative0: Finds the right path for users that have access to the DCCN computer cluster. It will replace '/home' by 'H:' and will replace forward by backward slashes.
 
-% % This can only run by people that have access to the DCCN intranet
-% filename=dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf'); 
-%     if ispc
-%         if ~strcmp(filename, 'H:\common\matlab\fieldtrip\data\ftp\test\edf\testAlphaIR20170321-0.edf')
-%             error('Alternative0 does not work');
-%         end
-%     else
-%         if ~strcmp(filename, '/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf')
-%             error('Alternative0 does not work');
-%         end
-%     end
-
-% This can work for everyone
-filename=dccnpath(tempdir);
-if ~strcmp(filename, tempdir)
-     error('dccnpath() does not find the right DCCN path');
+try
+filename=dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf'); 
+    if ispc
+        if ~strcmp(filename, 'H:\common\matlab\fieldtrip\data\ftp\test\edf\testAlphaIR20170321-0.edf')
+            error('Alternative0 does not work');
+        end
+    else
+        if ~strcmp(filename, '/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf')
+            error('Alternative0 does not work');
+        end
+    end
+catch
+    hostname=gethostname();
+    if startsWith(hostname, 'DCCN')
+        error('dccnpath() does not find the right DCCN path')
+    end
 end
+
 
 %% Alternative1: Allows to test with local files in the present working directory
 
@@ -35,7 +36,6 @@ end
 
 
 %% Alternative2: It downloads test data to the local computer, if test data is not already downloaded. It has 6 different tests that are listed below
-
 
 %% Test1: When ft_default.dccnpath is given by the user, dccnpath() should always select alternative2 (this is not done now)
 
@@ -56,28 +56,28 @@ mkdir(testdata);
 global ft_default
 ft_default.dccnpath=testdata;
 
-filename=dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/ctf/Subject01.ds/BadChannels');
+filename=dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/ctf/Subject01.ds/ClassFile.cls');
 if ~exist(filename, 'file')
     error('File is not downloaded from the HTTPS FieldTrip server');
 end
 
 
 %% Test3: File exists in the local copy and it doesn't get downloaded
-ft_default.dccnpath=fullfile(testdata,'BadChannels');
+ft_default.dccnpath=fullfile(testdata,'ClassFile.cls');
 
-filename=dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/ctf/Subject01.ds/BadChannels'); %'/home/common/matlab/fieldtrip/data/ftp/test/ctf/Subject01.ds/BadChannels' has already been downloaded by test2
[…]
