### 2023-08-22 e18bdad
Alternative2 - Test6: this test applies to external contributors and not for users in the DCCN cluster
