### 2023-08-22 447928f
FIX - hostname might start with 'DCCN' or 'dccn'
