### 2023-08-23 3f2c1e3
improve the conversion between probabilistic and indexed representation
