### 2023-08-23 3ee1c28
restructured some items in the test script
