+++ test/test_dccnpath.m
@@ -5,117 +5,130 @@ function test_dccnpath
 % DEPENDENCY dccnpath
 % DATA public
 
+global ft_default
+ft_default.dccnpath = [];
+
+% remember the original directory
+origdir = pwd;
+
 %% Alternative0: Finds the right path for users that have access to the DCCN computer cluster. It will replace '/home' by 'H:' and will replace forward by backward slashes.
 
 try
-filename=dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf'); 
-    if ispc
-        if ~strcmp(filename, 'H:\common\matlab\fieldtrip\data\ftp\test\edf\testAlphaIR20170321-0.edf')
-            error('Alternative0 does not work');
-        end
-    else
-        if ~strcmp(filename, '/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf')
-            error('Alternative0 does not work');
-        end
+  filename = dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf');
+  if ispc
+    if ~strcmp(filename, 'H:\common\matlab\fieldtrip\data\ftp\test\edf\testAlphaIR20170321-0.edf')
+      error('Alternative0 does not work');
     end
+  else
+    if ~strcmp(filename, '/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf')
+      error('Alternative0 does not work');
+    end
+  end
 catch
-    hostname=gethostname();
-    if startsWith(hostname, 'DCCN') || startsWith(hostname, 'dccn')
-        error('dccnpath() does not find the right DCCN path')
-    end
+  hostname = gethostname();
+  if startsWith(hostname, 'DCCN') || startsWith(hostname, 'dccn')
+    error('dccnpath() does not find the right DCCN path')
+  end
 end
 
+%% Alternative1: Test with local files in the present working directory
 
-%% Alternative1: Allows to test with local files in the present working directory
+tmpdir = tempname;
+mkdir(fullfile(tmpdir, 'ftp', 'test', 'edf'));
+cd(fullfile(tmpdir, 'ftp', 'test', 'edf'));
+websave('testAlphaIR20170321-0.edf', 'https://download.fieldtriptoolbox.org/test/edf/testAlphaIR20170321-0.edf');
 
-ft_defaults % '/home/common/matlab/fieldtrip/data/ftp/test/fileio' now belongs to the present working directory
-filename=dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/fileio');
-if ~contains(filename, 'fileio')
-        error('dccnpath() does not find the right working directory');
+filename = dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf');
+if ~startsWith(filename, tmpdir)
+  error('dccnpath() does not find the right working directory');
 end
 
+% return to the original directory
+cd(origdir);
 
 %% Alternative2: It downloads test data to the local computer, if test data is not already downloaded. It has 6 different tests that are listed below
 
 %% Test1: When ft_default.dccnpath is given by the user, dccnpath() should always select alternative2 (this is not done now)
 
-% if ~isempty(ft_default.dccnpath)
-%     filename=dccnpath('/home/common/matlab/fieldtrip/data/ftp/test/edf/testAlphaIR20170321-0.edf');
-%     if ~contains(filename, ft_default.dccnpath)
-%             error('The path defined by ft_default.dccnpath was not used, while it should have');
-%     end
-% end
-% 
[…]
