### 2023-08-23 d2b9bbd
renamed to snake_case, added copyrights
