### 2023-08-23 825d546
updated the help, updated the internal logic for alternative2
