+++ utilities/dccnpath.m
@@ -1,37 +1,35 @@
 function filename = dccnpath(filename)
 
-% DCCNPATH manages the filename and path for test files. It helps to locate and read test file from Linux, Windows or macOS computers both inside and outside the DCCN.
-% 
-% Use as 
+% DCCNPATH manages the filename and path for test files. It helps to locate and read
+% test file from Linux, Windows or macOS computers both inside and outside the DCCN.
+%
+% Use as
 %   filename = dccnpath(filename)
-% 
-% Inputs:
-%   filename - The relative path to the test data file within the DCCN cluster.
-%              Examples:
-%              - For public test data: '/home/common/matlab/fieldtrip/data/ftp/testdata' 
-%              - For private test data: '/home/common/matlab/fieldtrip/data/test/testdata' 
-% 
-% Outputs:
-%   filename - string that corresponds to the local path that the test data
-%   are saved.
-% 
-% The default location for FieldTrip and its test data is '/home/common/matlab/fieldtrip'. This function will search-and-replace this string by the location that applies to your computer. It will replace '/home' by 'H:' and will replace forward by backward slashes. 
-% 
-% In case you have a local copy:
-% 
-% - You should override the default location by 
-%    global ft_default 
-%    ft_default.dccnpath = '/your/copy'; 
-% 
-%    If you do not define ft_default.dccnpath manually then dccnpath will
-%    automatically use: 
-%    ft_default.dccnpath = tempdir;
-% 
-% - If the test data are not downloaded inside '/your/copy', then dccnpath
-%   will automatically download them.
-% 
-% Copyright (C) 2012-2023, Donders Centre for Cognitive Neuroimaging, Nijmegen, NL
+% where the input filename corresponds to the test data on the DCCN cluster and the
+% output filename corresponds to the local file including the full path where the
+% test data is available.
+%
+% The test data location on the DCCN cluster is '/home/common/matlab/fieldtrip/data'
+% and the specification of the input filename MUST start with this.
+%
+% This function will search-and-replace the location on the DCCN cluster by the
+% location that applies to your computer. If needed, it will replace '/home' by 'H:'
+% and will replace forward by backward slashes.
+%
+% In case you have a local copy of the data, or if you are inside the DCCN and have
+% mounted the '/home' drive on another letter than 'H:', you should override the
+% default location using
+%    global ft_default
+%    ft_default.dccnpath = '/your/copy';
+%
+% If you do not have a local copy and do not define ft_default.dccnpath manually,
+% then dccnpath will automatically use a temporary directory and try to download the
+% data.
+%
+% See also WHICH, WEBSAVE
 
+% Copyright (C) 2012-2023, Donders Centre for Cognitive Neuroimaging, Nijmegen, NL
+%
 % This file is part of FieldTrip, see http://www.fieldtriptoolbox.org
 % for the documentation and details.
 %
@@ -52,8 +50,10 @@ function filename = dccnpath(filename)
 
 global ft_default
 
+% it must always start with this
+assert(startsWith(filename, '/home/common/matlab/fieldtrip/data'));
 
-% alternative0 allows to test when in the DCCN cluster
+% alternative0 applies when in the DCCN cluster
 if ispc
   alternative0 = strrep(filename,'/home','H:');
   alternative0 = strrep(alternative0,'/','\');
@@ -63,112 +63,123 @@ else
 end
 
 if exist(alternative0, 'file') || exist(alternative0, 'dir')
-    warning('using default DCCN path %s', alternative0);
-    filename = alternative0;
-    return;
+  ft_notice('using default DCCN path %s', alternative0);
[…]
