### 2023-08-24 dd2b111
restructured test script, conditional on inside/outside dccn - added dropin replacement for strip function on older MATLABs - updated some comments in dccnpath
