### 2023-08-24 b61b3f7
see #2300 - ensure that each test script specifies WALLTIME, MEM and DATA
