+++ test/inspect_pull1434.m
@@ -1,6 +1,9 @@
 function inspect_pull1434
 
+% WALLTIME 00:10:00
+% MEM 3gb
 % DATA no
+
 %%
 
 fsample = 1000;
