+++ test/inspect_bug2588.m
@@ -1,5 +1,7 @@
 function inspect_bug2588
 
+% WALLTIME 00:10:00
+% MEM 3gb
 % DEPENDENCY ft_databrowser
 % DATA no
 
