+++ utilities/dccnpath.m
@@ -139,8 +139,8 @@ else
   % if the file doesn't exist or the folder is empty, then download test data
   % see also UNTAR, UNZIP, GUNZIP, which can download on the fly
 
-  if contains(alternative2, 'data/test') || contains(alternative2, 'data\test')
-    error('the test data are private and can not be downloaded')
+  if contains(alternative0, 'data/test') || contains(alternative0, 'data\test')
+    error('the test data are private and can not be downloaded from https://download.fieldtriptoolbox.org')
   end
 
   % public data are downloaded from https://download.fieldtriptoolbox.org
