### 2023-08-24 3d51c50
small cleanup to test script
