+++ test/test_old_trialdef.m
@@ -5,7 +5,6 @@ function test_old_trialdef
 % DEPENDENCY
 % DATA private
 
-
 % This script tests the implementation of the new representation of trial
 % specific information in the data structure.
 % The idea is as follows:
@@ -27,9 +26,6 @@ function test_old_trialdef
 % working on the raw datatype and outputting any other datatype
 % -sampleinfo and trialinfo should be concatenated in ft_appenddata
 % -sampleinfo and trialinfo should be adjusted by ft_rejectartifact
-
-% needed for the dccnpath function, since we will change directory later on
-addpath(fileparts(mfilename('fullpath')));
 
 cd(dccnpath('/home/common/matlab/fieldtrip/data/test/original/meg/ctf275/A0132_Aud-Obj-Recognition_20051115_02.ds'));
 headerfile = 'A0132_Aud-Obj-Recognition_20051115_02.res4';
