+++ test/test_pull2050.m
@@ -1,6 +1,6 @@
 function test_pull2050
 
-% MEM 1500mb
+% MEM 2gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_read_event
 % DATA no
