+++ test/test_bug2096.m
@@ -5,9 +5,6 @@ function test_bug2096
 % DEPENDENCY ft_sourcewrite ft_read_cifti ft_write_cifti
 % DATA private
 
-% needed for the dccnpath function, since we will change directory later on
-addpath(fileparts(mfilename('fullpath')));
-
 cd(dccnpath('/home/common/matlab/fieldtrip/data/test/bug2096'));
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -128,11 +125,7 @@ ft_sourcewrite(cfg, source);
 source1 = ft_read_cifti('test_bug2096.timeseries.dtseries.nii');
 
 %%
-try
-  parcellation = ft_read_atlas(dccnpath('/home/common/matlab/fieldtrip/template/atlas/aal/ROI_MNI_V4.nii'));
-catch
-  parcellation = ft_read_atlas(fullfile(getenv('HOME'), '/matlab/fieldtrip/template/atlas/aal/ROI_MNI_V4.nii'));
-end
+parcellation = ft_read_atlas(dccnpath('/home/common/matlab/fieldtrip/template/atlas/aal/ROI_MNI_V4.nii'));
 source = ft_checkdata(parcellation, 'datatype', 'source');
 source = removefields(source, {'tissue', 'tissuelabel'});
 source.pow = randn(prod(parcellation.dim), 1);
