### 2023-08-24 bc3b02f
update to v3.0.2, also show suggested tests
