+++ .github/workflows/suggest_tests.yml
@@ -4,7 +4,7 @@ on:
   pull_request:
 
 jobs:
-  create_comment:
+  suggest-test:
     runs-on: ubuntu-latest
     steps:
       - name: Check out repository
@@ -30,8 +30,12 @@ jobs:
         run: |
           python .github/scripts/suggest_tests.py ${{ steps.changed-files.outputs.all_changed_files }} > ./comment
 
+      - name: Show suggested tests
+        run: |
+          cat ./comment
+
       - name: Create comment with suggested tests
-        uses: peter-evans/create-or-update-comment@v3
+        uses: peter-evans/create-or-update-comment@v3.0.2
         with:
           issue-number: ${{ github.event.number }} 
           body-path: ./comment
