+++ .github/workflows/suggest_tests.yml
@@ -1,17 +1,19 @@
 name: Suggest tests
           
 on:
-  pull_request:
+  pull_request_target:
+    types: [opened, synchronize, reopened]
 
 jobs:
   suggest-test:
     runs-on: ubuntu-latest
     permissions:
-      contents: read
       pull-requests: write
     steps:
       - name: Check out repository
         uses: actions/checkout@v3
+        with:
+          fetch-depth: 2
 
       - name: Setup Python
         uses: actions/setup-python@v4
@@ -40,7 +42,6 @@ jobs:
       - name: Create comment with suggested tests
         uses: peter-evans/create-or-update-comment@v3.0.2
         with:
-          repo-token: ${{ secrets.GITHUB_TOKEN }}
           issue-number: ${{ github.event.number }} 
           body-path: ./comment
           
