+++ .github/workflows/suggest_tests.yml
@@ -1,25 +1,7 @@
 name: Suggest tests
           
 on:
-  pull_request_target:
-    types:
-      - assigned
-      - unassigned
-      - labeled
-      - unlabeled
-      - opened
-      - edited
-      - closed
-      - reopened
-      - synchronize
-      - converted_to_draft
-      - ready_for_review
-      - locked
-      - unlocked 
-      - review_requested 
-      - review_request_removed 
-      - auto_merge_enabled 
-      - auto_merge_disabled
+  pull_request:
 
 jobs:
   suggest-test:
