+++ .github/workflows/suggest_tests.yml
@@ -39,8 +39,8 @@ jobs:
           cat ./comment
 
       - name: Create comment with suggested tests
-        uses: peter-evans/create-or-update-comment@v3.0.2
+        uses: thollander/actions-comment-pull-request@v2
         with:
-          issue-number: ${{ github.event.number }} 
-          body-path: ./comment
+          message: |
+            Hello world ! :wave:
           
