+++ .github/workflows/suggest_tests.yml
@@ -1,7 +1,7 @@
 name: Suggest tests
           
 on:
-  pull_request:
+  pull_request_target:
 
 jobs:
   suggest-test:
