+++ test/test_pull1377.m
@@ -1,7 +1,7 @@
 function test_pull1377
 
 % MEM 12gb
-% WALLTIME 1:30:00
+% WALLTIME 01:30:00
 % DEPENDENCY ft_prepare_sourcemodel headsurface ft_prepare_leadfield ft_freqanalysis ft_sourceanalysis
 % DATA private
 
