+++ test/test_pull1377b.m
@@ -1,7 +1,7 @@
 function test_pull1377b
 
 % MEM 6gb
-% WALLTIME 1:00:00
+% WALLTIME 01:00:00
 % DEPENDENCY ft_sourceanalysis ft_dipolefitting
 % DATA private
 
