+++ test/test_bug2874.m
@@ -1,6 +1,6 @@
 function test_bug2874
 
-% WALLTIME 00:10:100
+% WALLTIME 00:10:10
 % MEM 1gb
 % DEPENDENCY ft_sourcegrandaverage
 % DATA no
