### 2023-08-25 e633b8c
FIX - do not use dccnpath for template data, see #2298
