+++ test/test_bug1405.m
@@ -5,6 +5,9 @@ function test_bug1405
 % DEPENDENCY ft_checkdata ft_senstype
 % DATA no
 
-load(dccnpath('/home/common/matlab/fieldtrip/template/headmodel/standard_mri.mat'));
+[ftver, ftpath] = ft_version;
+templatedir  = fullfile(ftpath, 'template');
+
+load(fullfile(templatedir, 'headmodel', 'standard_mri.mat'));
 mri = ft_checkdata(mri, 'hasunit', 'yes');
 
