+++ test/test_bug3190.m
@@ -5,9 +5,12 @@ function test_bug3190
 % DEPENDENCY ft_sourceinterpolate ft_sourceparcellate ft_volumereslice
 % DATA private
 
+[ftver, ftpath] = ft_version;
+templatedir  = fullfile(ftpath, 'template');
+
 %% Lets load atlas, sourcemodel, and a connectivity structure
-atlas = ft_read_atlas(dccnpath('/home/common/matlab/fieldtrip/template/atlas/aal/ROI_MNI_V4.nii'));
-load(dccnpath('/home/common/matlab/fieldtrip/template/sourcemodel/standard_sourcemodel3d10mm.mat')); % 'sourcemodel'
+atlas = ft_read_atlas(fullfile(templatedir, 'atlas', 'aal', 'ROI_MNI_V4.nii'));
+load(fullfile(templatedir, 'sourcemodel', 'standard_sourcemodel3d10mm.mat')); % 'sourcemodel'
 load(dccnpath('/home/common/matlab/fieldtrip/data/test/bug3190.mat'));
 
 source_pco  = ft_convert_units(source_pco, 'cm');
@@ -33,7 +36,7 @@ imagesc(source_pco_parc.powcorrspctrm)
 
 %% This subsequently failed
 % see http://bugzilla.fieldtriptoolbox.org/show_bug.cgi?id=3190#c6
-mri = ft_read_mri(dccnpath('/home/common/matlab/fieldtrip/template/anatomy/single_subj_T1_1mm.nii'));
+mri = ft_read_mri(fullfile(templatedir, 'anatomy', 'single_subj_T1_1mm.nii'));
 
 cfg            = [];
 cfg.resolution = 1;
