+++ test/test_tutorial_beamformingextended.m
@@ -5,9 +5,11 @@ function test_tutorial_beamformingextended
 % DEPENDENCY ft_read_mri ft_redefinetrial ft_freqanalysis ft_volumesegment ft_appenddata ft_selectdata ft_prepare_singleshell ft_sourceanalysis ft_prepare_leadfield ft_prepare_headmodel ft_prepare_sourcemodel ft_plot_headmodel ft_plot_sens ft_plot_mesh ft_sourceinterpolate ft_sourceplot
 % DATA public
 
+[ftver, ftpath] = ft_version;
+templatedir  = fullfile(ftpath, 'template');
+
 datadir = dccnpath('/home/common/matlab/fieldtrip/data/ftp/tutorial/sensor_analysis');
 mridir = dccnpath('/home/common/matlab/fieldtrip/data/ftp/tutorial/beamformingextended');
-templatedir  = dccnpath('/home/common/matlab/fieldtrip/template/sourcemodel');
 
 load(fullfile(datadir, 'subjectK.mat'));
 
@@ -82,7 +84,7 @@ cfg.method = 'singleshell';
 hdm = ft_prepare_headmodel(cfg, segmentedmri);
 
 
-template = load(fullfile(templatedir, 'standard_sourcemodel3d8mm'));
+template = load(fullfile(templatedir, 'sourcemodel', 'standard_sourcemodel3d8mm'));
 % inverse-warp the subject specific grid to the template grid
 cfg = [];
 cfg.sourcemodel.warpmni   = 'yes';
@@ -138,7 +140,7 @@ source_diff.pos = template.sourcemodel.pos;
 source_diff.dim = template.sourcemodel.dim;
 
 % note that the exact directory is user-specific
-templatefile = dccnpath('/home/common/matlab/fieldtrip/external/spm8/templates/T1.nii');
+templatefile = fullfile(ftpath, 'external', 'spm8', 'templates', 'T1.nii');
 template_mri = ft_read_mri(templatefile);
 template_mri.coordsys = 'spm';
 
@@ -157,7 +159,7 @@ cfg.opacitymap    = 'rampup';
 ft_sourceplot(cfg,source_diff_int);
 
 cfg.method = 'ortho';
-cfg.atlas  = dccnpath('/home/common/matlab/fieldtrip/template/atlas/aal/ROI_MNI_V4.nii');
+cfg.atlas  = fullfile(templatedir, 'atlas', 'aal','ROI_MNI_V4.nii');
 ft_sourceplot(cfg,source_diff_int);
 
 cfg.method = 'surface';
@@ -198,7 +200,7 @@ source_coh_lft.pos = template.sourcemodel.pos;
 source_coh_lft.dim = template.sourcemodel.dim;
 
 % note that the exact directory is user-specific
-templatefile = dccnpath('/home/common/matlab/fieldtrip/external/spm8/templates/T1.nii');
+templatefile = fullfile(ftpath, 'external', 'spm8', 'templates', 'T1.nii');
 template_mri = ft_read_mri(templatefile);
 template_mri.coordsys = 'spm';
 
@@ -218,6 +220,6 @@ cfg.opacitylim    = [00 .15];
 cfg.maskparameter = cfg.funparameter;
 cfg.opacitymap    = 'rampup';
 
-cfg.atlas         = dccnpath('/home/common/matlab/fieldtrip/template/atlas/aal/ROI_MNI_V4.nii');
+cfg.atlas         = fullfile(templatedir, 'atlas', 'aal', 'ROI_MNI_V4.nii');
 
 ft_sourceplot(cfg, source_coh_int);
