+++ test/test_bug1029.m
@@ -8,10 +8,13 @@ function test_bug1029
 % tests the functionality of the forward models after changing the names of
 % the headmodels
 
+[ftver, ftpath] = ft_version;
+templatedir  = fullfile(ftpath, 'template');
+
 %%%%%%%%%%%%%%%%
 % BEM_ASA
 %%%%%%%%%%%%%%%%
-filename = dccnpath('/home/common/matlab/fieldtrip/template/headmodel/skin/standard_skin_1222.vol');
+filename = fullfile(templatedir, 'headmodel', 'skin', 'standard_skin_1222.vol');
 vol = ft_read_headmodel(filename);
 
 vol = ft_headmodel_asa(filename);
