+++ test/test_bug1832.m
@@ -9,7 +9,10 @@ function test_bug1832
 % 1) make test script that checks on the units in these objects (with assert)
 % 2) also go through the rest of the example script and check on units
 
-template = ft_read_mri(dccnpath('/home/common/matlab/fieldtrip/external/spm8/templates/T1.nii'));
+[ftver, ftpath] = ft_version;
+templatedir  = fullfile(ftpath, 'template');
+
+template = ft_read_mri(fullfile(ftpath, 'external', 'spm8', 'templates', 'T1.nii'));
 template.coordsys = 'spm'; % so that FieldTrip knows how to interpret the coordinate system
 
 % segment the template brain and construct a volume conduction model (i.e. head model): this is needed

[…]
