+++ test/test_bug1665.m
@@ -10,13 +10,13 @@ function test_bug1665
 % MRI contains NaNs
 
 % get the FieldTrip version and path
-[v, p] = ft_version;
+[ftver, ftpath] = ft_version;
 
 ft_hastoolbox('spm8',1,0);
 load(dccnpath('/home/common/matlab/fieldtrip/data/test/bug1665/segmentedS2.mat'));
 struct_reslice = ft_checkdata(struct_reslice, 'datatype', 'volume');
 
-cd(dccnpath(fullfile(p, 'utilities', 'private')));
+cd(fullfile(ftpath, 'utilities', 'private'));
 output = align_ctf2acpc(struct_reslice);
 
 struct_reslice.coordsys = 'itab';
