### 2023-08-25 298d0f5
FIX - more assorted issues with dccnpath related to the update in #2298
