+++ test/test_old_ft_multiplotER.m
@@ -10,7 +10,7 @@ function test_old_ft_multiplotER
 % different input datatypes. no other functionality is tested.
 % the script has been written in order to test a clean up of the code
 
-filename = dccnpath('/home/common/matlab/fieldtrip/data/test/latest/raw/eeg/preproc_neuroscan16');
+filename = dccnpath('/home/common/matlab/fieldtrip/data/test/latest/raw/eeg/preproc_neuroscan16.mat');
 load(filename);
 
 %there's an unresolved issue with duplicate labels 'FREE'
