+++ test/test_bug2879.m
@@ -5,7 +5,7 @@ function test_bug2879
 % DEPENDENCY ft_sourcestatistics
 % DATA private
 
-filename = dccnpath('/home/common/matlab/fieldtrip/data/test/bug2879');
+filename = dccnpath('/home/common/matlab/fieldtrip/data/test/bug2879.mat');
 load(filename)
 
 stat = ft_sourcestatistics(cfg,dat1{:},dat2{:});
