### 2023-08-27 c736c03
FIX - test script needed a small update following the DCCNPATH changes
