### 2023-08-27 #5566 @crsl4
Thanks @robertoostenveld ! I talked with other JOSS editors, and it seems that @matsvanes satisfies the criteria for first authorship.
