### 2023-08-27 #5566 @crsl4
One thing we need to determine is the existing overlap between the pre-print and the JOSS article. @matsvanes Is the pre-print to be published in another journal? If the pre-print will be published elsewhere, then we need to make sure that there is no overlap with the JOSS manuscript. That is, it is ok to use the software in a dataset for another publication, but not for the JOSS manuscript (or the introduction of the software) to be a subset of the pre-print to be published elsewhere.

@ashahide @gflofst Can you add this as part of your review? Thanks!
