### 2023-08-27 65f2fec
ENH - document the existing "inventorise" option, related to #2294 This also adds support for selecting on no/public/private data, but does not yet remove the filtering based on dccnpath.
