+++ utilities/private/ft_test_run.m
@@ -27,7 +27,7 @@ command = varargin{1};
 assert(isequal(command, 'run') || isequal(command, 'inventorize'));
 varargin = varargin(2:end);
 
-optbeg = find(ismember(varargin, {'dependency', 'dccnpath', 'maxmem', 'maxwalltime', 'upload', 'sort', 'returnerror'}));
+optbeg = find(ismember(varargin(1:2:end), {'dependency', 'dccnpath', 'maxmem', 'maxwalltime', 'data', 'upload', 'sort', 'returnerror'}));
 if ~isempty(optbeg)
   optarg   = varargin(optbeg:end);
   varargin = varargin(1:optbeg-1);
@@ -43,6 +43,7 @@ dependency  = ft_getopt(optarg, 'dependency', {});
 hasdccnpath = ft_getopt(optarg, 'dccnpath');  % default is handled below
 maxmem      = ft_getopt(optarg, 'maxmem', inf);
 maxwalltime = ft_getopt(optarg, 'maxwalltime', inf);
+data        = ft_getopt(optarg, 'data', {'no', 'public', 'private'});
 upload      = ft_getopt(optarg, 'upload', 'yes'); % this will be set to 'no' in case FieldTrip version is not clean
 sortarg     = ft_getopt(optarg, 'sort', 'alphabetical');
 returnerror = ft_getopt(optarg, 'returnerror', 'no');
@@ -50,6 +51,11 @@ returnerror = ft_getopt(optarg, 'returnerror', 'no');
 if ischar(dependency)
   % this should be a cell-array
   dependency = {dependency};
+end
+
+if ischar(data)
+  % this should be a cell-array
+  data = {data};
 end
 
 if isempty(hasdccnpath)
@@ -95,6 +101,7 @@ dep  = true(size(filelist));
 file = false(size(filelist)); % by default ignore file reads
 mem  = zeros(size(filelist));
 tim  = zeros(size(filelist));
+dat  = repmat({'unknown'}, size(filelist)); % not to be confused with data
 
 for i=1:numel(filelist)
   fid = fopen(filelist{i}, 'rt');
@@ -136,45 +143,58 @@ for i=1:numel(filelist)
       s = s + length('% MEM'); % strip this part
       mem(i) = str2mem(line{k}(s:e));
     end
+
+    [s, e] = regexp(line{k}, '% DATA.*', 'once', 'start', 'end');
+    if ~isempty(s)
+      s = s + length('% DATA'); % strip this part
+      dat{i} = line{k}(s:e);
+    end
   end % for each line
   
 end % for each function/file
 
 fprintf('%3d scripts are excluded due to the dependencies\n',                  sum(~dep));
-fprintf('%3d scripts are excluded due to loading files from the DCCN path\n',  sum(file));
 fprintf('%3d scripts are excluded due to the requirements for memory\n',       sum(mem>maxmem));
-fprintf('%3d scripts are excluded due to the requirements for walltime \n',    sum(tim>maxwalltime));
+fprintf('%3d scripts are excluded due to the requirements for walltime\n',     sum(tim>maxwalltime));
+fprintf('%3d scripts are excluded due to the requirements for data\n',         sum(~contains(dat, data)));
+fprintf('%3d scripts are excluded due to loading files from the DCCN path\n',  sum(file));
 
 % make selection of test scripts, remove scripts that exceed walltime or memory
-sel                  = true(size(filelist));
-sel(tim>maxwalltime) = false;
-sel(mem>maxmem)      = false;
-sel(dep==false)      = false;
-sel(file==true)      = false;
+sel                       = true(size(filelist));
+sel(tim>maxwalltime)      = false;
+sel(mem>maxmem)           = false;
+sel(~contains(dat, data)) = false;
[…]
