### 2023-08-28 cf47761
FIX - added .mat to the input filename to not confuse dccnpath
