### 2023-08-28 f41a2f7
ENH - update the mem and walltime, it is now added by the dashboard script See also #2300
