+++ test/test_ft_math.m
@@ -1,6 +1,6 @@
 function test_ft_math
 
-% MEM 2gb
+% MEM 1gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_math
 % DATA no
