+++ test/test_ft_compute_leadfield.m
@@ -1,6 +1,6 @@
 function test_ft_compute_leadfield
 
-% MEM 2gb
+% MEM 1gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_compute_leadfield eeg_halfspace_medium_leadfield eeg_halfspace_monopole eeg_leadfieldb eeg_slab_monopole inf_medium_leadfield leadfield_fns leadfield_simbio magnetic_dipole meg_forward meg_leadfield1 ft_prepare_vol_sens
 % DATA private
