+++ test/test_bug1150.m
@@ -1,6 +1,6 @@
 function test_bug1150
 
-% MEM 2gb
+% MEM 1gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_sourcestatistics
 % DATA private
