+++ test/test_bug1210.m
@@ -1,6 +1,6 @@
 function test_bug1210
 
-% MEM 2gb
+% MEM 1gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_datatype_sens
 % DATA private
