+++ test/test_bug3029.m
@@ -1,6 +1,6 @@
 function test_bug3029
 
-% MEM 5gb
+% MEM 4gb
 % WALLTIME 00:20:00
 % DEPENDENCY ft_sourceanalysis prepare_freq_matrices
 % DATA private
