+++ test/test_issue2169.m
@@ -1,7 +1,7 @@
 function test_issue2169
 
 % WALLTIME 00:10:00
-% MEM 2gb
+% MEM 1gb
 % DEPENDENCY ft_apply_montage
 % DATA no
 
