+++ test/test_ft_heartrate.m
@@ -1,7 +1,7 @@
 function test_ft_heartrate
 
 
-% MEM 2gb
+% MEM 1gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_heartrate
 % DATA no
