+++ test/test_ft_timelockanalysis_new.m
@@ -3,7 +3,7 @@ function test_ft_timelockanalysis_new(datainfo,writeflag)
 % MEM 1gb
 % WALLTIME 00:10:00
 % DEPENDENCY ft_timelockanalysis_new ft_timelockanalysis ref_datasets
-% DATA public
+% DATA private
 
 % the optional writeflag determines whether the output should be saved to
 % disk
