+++ test/test_ft_timelockanalysis.m
@@ -3,7 +3,7 @@ function test_ft_timelockanalysis(datainfo, writeflag, version)
 % MEM 1gb
 % WALLTIME 00:10:00
 % DEPENDENCY
-% DATA no
+% DATA private
 
 % ft_timelockanalysis ref_datasets
 
