+++ test/test_ft_preprocessing.m
@@ -3,7 +3,7 @@ function test_ft_preprocessing(datainfo, writeflag, version)
 % MEM 1gb
 % WALLTIME 00:20:00
 % DEPENDENCY ft_preprocessing ref_datasets
-% DATA no
+% DATA private
 
 % writeflag determines whether the output should be saved to disk
 % version determines the output directory
