+++ test/test_bug2220.m
@@ -3,7 +3,7 @@ function test_bug2220(datainfo)
 % MEM 1gb
 % WALLTIME 00:20:00
 % DEPENDENCY ft_preprocessing ft_preproc_padding preproc
-% DATA no
+% DATA private
 
 
 %load('C:\Users\jorhor\Downloads\bugdatafile.mat')
