### 2023-08-29 #5566 @matsvanes
We have no intentions to publish the pre-print in another journal. After releasing the pre-print we changed course and decided to focus more on the software for publication. For this reason, we moved most of the content of the pre-print (e.g., the examples) to the [FieldTrip Website](https://www.fieldtriptoolbox.org/example/reproducescript/). If the software were to be published at JOSS, the pre-print would thus become obsolete, just like it would when a pre-print is finally published in a peer-reviewed journal.
