### 2023-08-29 a531816
ENH - allow either prestim/poststim or duration/offset as trialdef options
