+++ trialfun/ft_trialfun_bids.m
@@ -4,17 +4,21 @@ function [trl, allevents] = ft_trialfun_bids(cfg)
 % the basis of the BIDS "events.tsv" file. This function should in general not be
 % called directly, it will be called by FT_DEFINETRIAL.
 %
-% Use this function by calling 
+% Use this function by calling
 %   [cfg] = ft_definetrial(cfg)
 % where the configuration structure should contain
 %   cfg.dataset   = string with the filename
 %   cfg.trialdef  = structure with the details of trial definition, see below
 %   cfg.trialfun  = 'ft_trialfun_bids'
 %
-% The trialdef structure should contain the following specifications
-%   cfg.trialdef.prestim    = latency in seconds (required)
-%   cfg.trialdef.poststim   = latency in seconds (required)
-% and you can specify your selection of events as
+% The trialdef structure should either contain the following
+%   cfg.trialdef.prestim    = latency in seconds
+%   cfg.trialdef.poststim   = latency in seconds
+% or the duration and offset relative to the event of interest
+%   cfg.trialdef.duration    = latency in seconds
+%   cfg.trialdef.offset      = latency in seconds
+%
+% You can specify your selection of events as
 %   cfg.trialdef.columnname = columnvalue
 % where the column name and value have to match those present in the events.tsv file.
 %
@@ -27,7 +31,7 @@ function [trl, allevents] = ft_trialfun_bids(cfg)
 %
 % See also FT_DEFINETRIAL, FT_TRIALFUN_GENERAL
 
-% Copyright (C) 2021, Robert Oostenveld
+% Copyright (C) 2021-2023, Robert Oostenveld
 %
 % This file is part of FieldTrip, see http://www.fieldtriptoolbox.org
 % for the documentation and details.
@@ -75,10 +79,6 @@ else
   events = ft_read_tsv(eventsfile);
 end
 
-% these cannot be used as column headings, as they are used as normal cfg options
-assert(~ismember('prestim', events.Properties.VariableNames));
-assert(~ismember('poststim', events.Properties.VariableNames));
-
 % make a selection of the rows
 sel = true(size(events,1),1);
 
@@ -93,23 +93,64 @@ end
 
 % remember and return all events
 allevents = events;
-% continue with the selection of events
+
+% continue defining trials with the selection of events
 events = events(sel,:);
 
-if ismember('sample', events.Properties.VariableNames)
-  % use the sample number, it is assumed that sample 1 corresponds to the first data sample (and not sample 0)
-  begsample = round(events.sample - cfg.trialdef.prestim  * hdr.Fs);
-  endsample = round(events.sample + cfg.trialdef.poststim * hdr.Fs);
-  offset    = round(-cfg.trialdef.prestim*hdr.Fs) * ones(size(begsample));
+% FIXME it would also be possible to construct trials if the events contained prestim and poststim
+% FIXME it would also be possible to construct trials if the events contained duration and offset
+% FIXME in those cases the user-specified cfg.trialdef options could overrule the ones from the events
+
+if isfield(cfg.trialdef, 'prestim') && isfield(cfg.trialdef, 'poststim')
+  % these are mutually exclusive
+  assert(~isfield(cfg.trialdef, 'duration'));
+  assert(~isfield(cfg.trialdef, 'offset'));
+
+  if ismember('sample', events.Properties.VariableNames)
+    % use the sample number, it is assumed that sample 1 corresponds to the first data sample (and not sample 0)
+    begsample = round(events.sample - cfg.trialdef.prestim  * hdr.Fs);
[…]
