### 2023-08-30 e3b0d2f
dump the github action context to debug the conditional execution
