### 2023-08-30 84e37b6
ENH - improve event representation and extractability from fif-files
