+++ fieldtrip2fiff.m
@@ -119,7 +119,7 @@ else
   info.highpass        = NaN;
   info.lowpass         = NaN;
   
-  % these are not strictly necessary, for basic functionality, i.e. time series manipulation. Thinks work
+  % these are not strictly necessary, for basic functionality, i.e. time series manipulation. Things work
   % best if an attempt is made to represent MEG-sensors according to the MNE conventions, i.e. sensors in 
   % (neuromag) device coordinates, and the appropriate additional transformations specified. Fieldtrip2fiff 
   % makes an attempt to do this, if 'coordsys' = 'neuromag' (handled below)
@@ -236,18 +236,25 @@ if israw
   fiff_write_int(outfid, FIFF.FIFF_FIRST_SAMPLE, round(data.time{1}(1)*fsample));
   fiff_write_int(outfid, FIFF.FIFF_DATA_SKIP, 0);
   fiff_write_raw_buffer(outfid, data.trial{1}, cals, dtype);
-  fiff_finish_writing_raw(outfid);
   
   % write events, if specified or present in the structure provenance
-  if isempty(event) && isfield(data, 'cfg')
-    event = ft_findcfg(data.cfg, 'event');
-  end
   if ~isempty(event)
     eventtype = ft_getopt(varargin, 'eventtype', 'all');
-    [eventlist, mappings] = convertevent(event, eventtype);
-    fiff_write_events(eventfile, eventlist, mappings);
-    ft_info('Writing events to %s\n', eventfile)
+    if isequal(eventtype, 'all')
+      eventtype = {event.type};
+    else
+      type      = {event.type};
+      sel       = match_str(unique(type), eventtype);
+      eventtype = type(sel);
+    end
+
+    if ~isempty(eventtype)
+      ft_info('Writing event matrix to %s\n', fifffile);
+      [eventlist, mappings] = convertevent(event, eventtype);
+      fiff_write_events(outfid, eventlist, mappings)
+    end
   end
+  fiff_finish_writing_raw(outfid);
   
 elseif istlck
   evoked.aspect_kind = 100;
@@ -270,6 +277,7 @@ elseif isepch
     ft_warning('Using the first column of the trialinfo field as event values');
     trg = data.trialinfo(:,1);
   else
+    ft_warning('Marking each epoch boundary with a single event value (1)')
     trg = ones(ntrl,1);
   end
   events = [((0:nsmp:(nsmp*(ntrl-1))))' zeros(ntrl,1) trg];
@@ -338,11 +346,25 @@ else
     coilunit  = grad2coilunit(data.grad, FIFF);
   end
 
+  % FIXME: experimental attempt to assign digital trigger channels correctly
+  % these are so far only from 4d or ctf systems
+  [i_labstim, i_stim] = match_str(data.label, {'STIM' 'UPPT001' 'TRIGGER'}');
+  stype(i_labstim) = 2;
+  indx(i_labstim)  = i_stim;
+
+  % FIXME: experimental attempt to assign digital response channels correctly
+  % these are so far only from 4d or ctf systems
+  [i_labresp, i_resp] = match_str(data.label, {'UPPT002' 'RESPONSE'}');
+  stype(i_labresp) = 3;
+  indx(i_labresp)  = i_resp;
+
   scanno  = num2cell((1:nchan)');
   ch_name = data.label(:);
   range   = num2cell(ones(nchan,1));
   cal     = num2cell(ones(nchan,1));
 
+  cnt_resp = 0;
+  cnt_stim = 0;
   cnt_grad = 0;
   cnt_elec = 0;
   cnt_else = 0;
@@ -350,6 +372,34 @@ else
   for k = 1:nchan
   
[…]
