### 2023-08-30 603cf89
ENH - some updates for smoother chantype detection in some cases
