+++ fieldtrip2fiff.m
@@ -60,9 +60,10 @@ function fieldtrip2fiff(filename, data, varargin)
 % this ensures that the path is correct and that the ft_defaults global variable is available
 ft_defaults
 
-coordsys  = ft_getopt(varargin, 'coordsys', 'neuromag');
+coordsys  = ft_getopt(varargin, 'coordsys',  'neuromag');
 headshape = ft_getopt(varargin, 'headshape', []);
-event     = ft_getopt(varargin, 'event', []);
+event     = ft_getopt(varargin, 'event',     []);
+hdr       = ft_getopt(varargin, 'hdr',       []);
 
 % ensure that the filename has the correct extension
 [pathstr, name, ext] = fileparts(filename);
@@ -70,11 +71,23 @@ if ~strcmp(ext, '.fif')
   ft_error('If the filename is specified with a file extension, this should read .fif');
 end
 fifffile  = fullfile(pathstr ,[name '.fif']);
-eventfile = fullfile(pathstr ,[name '-eve.fif']);
 
 % ensure the mne-toolbox to be on the path
 ft_hastoolbox('mne', 1);
 FIFF = fiff_define_constants; % some constants are not defined in the MATLAB function
+
+% check the input data for the presence of a hdr before passing it to
+% ft_checkdata (because the hdr might be scrubbed there)
+if isempty(hdr) && ~isfield(data, 'hdr')
+  ft_info('The input data does not include header information');
+elseif isempty(hdr) && isfield(data, 'hdr')
+  ft_info('Using the header information from the input data');
+  hdr = data.hdr;
+elseif ~isempty(hdr) && ~isfield(data, 'hdr')
+  ft_info('Using the header information supplied as separate input');
+elseif ~isempty(hdr) && isfield(data, 'hdr')
+  ft_error('You should either supply a separate header, or a data argument with a hdr field, not both');
+end
 
 % check if the input data is valid for this function
 data   = ft_checkdata(data, 'datatype', {'raw', 'timelock'}, 'hassampleinfo', 'yes', 'feedback', 'yes');
@@ -82,6 +95,8 @@ istlck = ft_datatype(data, 'timelock') && isfield(data, 'avg');
 israw  = ft_datatype(data, 'raw') && numel(data.trial)==1;
 isepch = ft_datatype(data, 'timelock') || (ft_datatype(data, 'raw') && numel(data.trial)>1);
 if isepch
+  % this step ensures that all trials have a common time axis, and that variable length trials can 
+  % be handled (injecting the shorter trials with NaNs)
   data = ft_checkdata(data, 'datatype', 'timelock', 'feedback', 'yes');
 end
 if israw
@@ -105,9 +120,9 @@ if ~ismember(precision, {'single' 'double'})
 end
 
 % Create a fiff-header, or take information from the original header if possible
-if isfield(data, 'hdr') && isfield(data.hdr, 'orig') && isfield(data.hdr.orig, 'meas_id')
+if ~isempty(hdr) && isfield(hdr, 'orig') && isfield(hdr.orig, 'meas_id')
   ft_notice('The data contains header information that seems to be derived from a FIFF file,\nre-using header information, channel locations may be read \nfrom .grad and .elec in the data');
-  info = data.hdr.orig;
+  info = hdr.orig;
 else
[…]
