### 2023-08-30 68dc574
ENH - some fixes and extra tests
