### 2023-08-30 5c1c26f
ENH - implemented writing raw and timelock data to CTF dataset
