### 2023-09-01 cc4d435
ENH - some updates + documentation
