+++ fieldtrip2fiff.m
@@ -7,32 +7,39 @@ function fieldtrip2fiff(filename, data, varargin)
 %   fieldtrip2fiff(filename, data)
 % where filename is the name of the output file, and data is a raw data structure
 % as obtained from FT_PREPROCESSING, or a timelock structure obtained from
-% FT_TIMELOCKANALYSIS.
+% FT_TIMELOCKANALYSIS. If the input data is a raw data structure with a single
+% trial, a continuous fif-file will be written. If the input data contains multiple
+% trials, either in a timelock or raw format, and epoched fif-file will be written. 
+% If trials have different time axes, nans will be added to pad the trials to equal
+% length and time axis. If the input data contains an average across trials, an evoked
+% fif-file will be written.
 %
 % Additional options can be specified as key-value pairs:
-%   precision = string ('single'/'double'), determines the precision with which
-%               the numeric data is written to file, default is the class of the
-%               numeric data.
+%   precision = string ('single'/'double'), determines the precision with which the
+%               numeric data is written to file, default is the class of the data.
 %   coordsys  = string ('native'/'neuromag'), determines the coordinate system in which
 %               the MEG sensors are written (default = 'neuromag'). In case of 
-%               'neuromag' the MEG sensors are expressed in (approximate) neuroma^
+%               'neuromag' the MEG sensors are expressed in (approximate) neuromag
 %               coordinates, which may facilitate downstream handling of the fif-files
 %               in other software such as MNE-python. This is according to the
 %               official fif-file format definition. This option does not have an
 %               effect on EEG electrodes or fNIRS optodes.
-%   event     = structure as obtained from FT_READ_EVENT
+%   event     = structure as obtained from FT_READ_EVENT, note that the sampling in the
+%               event structure should be the same as the sampling of the data structure,
+%               i.e. the values in data.sampleinfo should be in line with event.sample, and
+%               the sampling rate should be the same. No check will be performed. Also, the 
+%               events will only be written to file if the input data is of type raw with
+%               a single trial.               
 %   eventtype = string or cell array of string with the event types to be
-%               written to file (default is all)
-% 
-% If the data comes from FT_PREPROCESSING and has only one trial, then it writes the
-% data into raw continuous format. 
+%               written to the continuous fif-file (default is all)
+%   hdr       = structure as obtained from FT_READ_HEADER
 % 
 % If present in the data, the original header is reused (also removing the non-used channels).
 % Otherwise, the function attempts to create the header, which might or might not be correct
-% (e.g. with respect to the scaling and the sensor locations. 
+% (e.g. with respect to the scaling and the sensor locations). 
 % 
-% The events are written in MNE format (three columns) into a file based on "filename", ending
-% with "-eve.fif".
+% The events are written in MNE format (three columns) into the continuous
+% fif-file, with a mapping string that allows for a richer interpretation of the events.
 % 
 % See also FT_DATATYPE_RAW, FT_DATATYPE_TIMELOCK
 
@@ -94,6 +101,9 @@ data   = ft_checkdata(data, 'datatype', {'raw', 'timelock'}, 'hassampleinfo', 'y
 istlck = ft_datatype(data, 'timelock') && isfield(data, 'avg');
 israw  = ft_datatype(data, 'raw') && numel(data.trial)==1;
 isepch = ft_datatype(data, 'timelock') || (ft_datatype(data, 'raw') && numel(data.trial)>1);
+if ~israw && ~isempty(event)
+  ft_error('events are only used for the writing of continuous files');
+end
 if isepch
   % this step ensures that all trials have a common time axis, and that variable length trials can 
[…]
