### 2023-09-01 9199afe
improve representation of events and triggers following fieldtrip2fiff (#2313)
* ENH - improve event representation and extractability from fif-files, made
some updates for smoother chantype detection in some cases, included some extra tests, updated docstring
