### 2023-09-01 9d24027
Update ft_crossfrequencyanalysis.m (#2285)
* Update ft_crossfrequencyanalysis.m use nanmean (don't use the 'omitnan' option for mean (yet -> see the assosciated discussion on github)

---------

Co-authored-by: Jan-Mathijs Schoffelen <1517611+schoffelen@users.noreply.github.com>
