### 2023-09-02 4ba8cbe
ENH - Updated ft_test help documentation to include find dependency, update_dependency and untested_functions and moved these functions inside utilities/private
