+++ utilities/ft_test.m
@@ -1,16 +1,38 @@
 function [result] = ft_test(varargin)
 
-% FT_TEST performs selected FieldTrip test scripts or reports on previous test
+% FT_TEST performs selected FieldTrip test scripts, finds and updates the dependencies of test scripts, finds which high-level FieldTrip functions are not tested, or reports on previous test
 % results from the dashboard database.
 %
 % Use as
 %   ft_test inventorize     ...
 %   ft_test run             ...
-%   ft_test moxunit_run     ...
-%   ft_test report          ...
-%   ft_test compare         ...
-%
-% ========= Running simple tests scripts =========
+%   ft_test find_dependency ...
+%   ft_test update_dependency ...
+%   ft_test untested_functions ...
+%   ft_test moxunit_run     ... % This is obsolete
+%   ft_test report          ... % This is obsolete
+%   ft_test compare         ... % This is obsolete
+% 
+% ========= INVENTORIZE =========
+% 
+% To list the tests based on their dependencies, you would do 
+%   ft_test inventorize
+% to list all test functions, or
+%   ft_test inventorize data no
+% to list test functions that don't need any external data to run.
+%  
+% 
+% Additional optional arguments are specified as key-value pairs and can include
+%   dependency       = string or cell-array of strings
+%   upload           = string, upload test results to the dashboard, can be 'yes' or 'no' (default = 'yes')
+%   dccnpath         = string, allow files to be read from the DCCN path, can be 'yes' or 'no' (default is automatic)
+%   maxmem           = number (in bytes) or string such as 10GB
+%   maxwalltime      = number (in seconds) or string such as HH:MM:SS
+%   data             = string or cell-array of strings with 'no', 'public' and/or 'private'
+%   sort             = string, can be 'alphabetical', 'walltime', 'mem' or 'random' (default = 'alphabetical')
+%   returnerror      = string, whether give an error upon detecting a failed script, can be 'immediate', 'final', 'no' (default = 'no')
+% 
+% ========= RUN =========
 %
 % To execute a test and submit the results to the dashboard database, you would do
 %   ft_test run
@@ -30,8 +52,37 @@ function [result] = ft_test(varargin)
 %   data             = string or cell-array of strings with 'no', 'public' and/or 'private'
 %   sort             = string, can be 'alphabetical', 'walltime', 'mem' or 'random' (default = 'alphabetical')
 %   returnerror      = string, whether give an error upon detecting a failed script, can be 'immediate', 'final', 'no' (default = 'no')
+% 
+% ========= FIND_DEPENDENCY =========
+% 
+% To find on what functions test scripts depend on, you would do
+%   ft_test find_dependency test_bug46
+% to find on what functions test_bug46 depends on.
+% 
+% It outputs:
+%   inlist   = Nx1 cell-array, descibes the rows and lists the test scripts
+%   outlist  = 1xM cell-array, describes the columns and lists the
+%   dependencies
+%   depmat   = NxM dependency matrix, see below
+% 
+% The dependency matrix contains the following values:
+%  - 0 if there is no dependency
+%  - 2 for a direct dependency
+% 
+% 
+% ========= UPDATE_DEPENDENCY =========
+% 
+% To update the DEPENDENCY header items of test scripts, you would do:
+%   ft_test update_dependency test_bug46
+% to update the DEPENDENCY header items of test_bug46.
+% 
+% ========= UNTESTED_FUNCTIONS =========
+% 
+% To find FieldTrip high-level functions not tested by any test scripts,
+% you would do
+%  ft_test untested_functions
 %
-% ========= Running MOxUnit tests =========
+% ========= MOXUNIT_RUN =========
 %
[…]
