### 2023-09-03 8eaeb4c
Refactored ft_test Function
- Updated the ft_test code to serve as a wrapper function.
- It now facilitates the execution of find_dependency, update_dependency, and untested_functions.
