+++ utilities/private/ft_test_report.m
@@ -1,6 +1,8 @@
 function [result] = ft_test_report(varargin)
 
-% FT_TEST_REPORT
+% FT_TEST_REPORT documentation is included inside ft_test documentation.
+% 
+% See also FT_TEST
 
 % Copyright (C) 2017, Robert Oostenveld
 %
