### 2023-09-05 80e875a
some cleanups to the help and whitespace
