### 2023-09-05 9e05c40
added support for writelines function for MATLAB versions older than matlab2022a
