### 2023-09-06 72fe9d9
Update writelines.m
