### 2023-09-07 4705fdb
Improved the output that is printed in the command window for ft_test_find_dependency. For example it is printed as 'The test scripts: test_bug46 test_bug1232 test_bug1891 test_bug1212 depend on: ft_connectivityanalysis ft_selectdata ft_connectivity_wpli ft_layoutplot '
