+++ utilities/private/ft_test_find_dependency.m
@@ -34,9 +34,10 @@ else
     inlist=varargin(2:end);
 end
 
+% Remove path and file extension from 'inlist'
 for i=1:length(inlist)  
   [p, f, x] = fileparts(inlist{i});
-  inlist{i} = f; % Remove path and file extension
+  inlist{i} = f; 
 
   try
     fprintf('processing "%s"\n', inlist{i});
@@ -77,9 +78,8 @@ for i=1:length(s)
   end
 end 
 
+% Remove path and file extension from 'outlist'
 for i=1:length(outlist)  
   [p, f, x] = fileparts(outlist{i});
-  outlist{i} = f; % Remove path and file extension
+  outlist{i} = f; 
 end
-
-% outlist = string(outlist);
