### 2023-09-11 28f493d
ENH - improvement for consistency in event(_id) representation between epoched and continuous files (#2321)
