### 2023-09-12 134f8e2
ft2fiff (#2322)
* FIX - write event_id with semicolon as separator and make interpretation of event_id more robust
