### 2023-09-13 fb610de
return 'trial' events with values for CTF
