+++ fileio/ft_read_event.m
@@ -102,7 +102,7 @@ end
 if iscell(filename)
   % use recursion to read the events from multiple files
   ft_warning('concatenating events from %d files', numel(filename));
-  
+
   hdr = ft_getopt(varargin, 'header');
   if isempty(hdr)
     % read the combined and individual file headers
@@ -112,7 +112,7 @@ if iscell(filename)
     combined = hdr;
   end
   hdr = combined.orig;
-  
+
   allhdr = cat(1, hdr{:});
   if numel(unique([allhdr.label]))==sum([allhdr.nChans])
     % each file has different channels, concatenate along the channel dimension
@@ -121,7 +121,7 @@ if iscell(filename)
       varargin = ft_setopt(varargin, 'header', hdr{i});
       event{i} = ft_read_event(filename{i}, varargin{:});
     end
-    
+
   else
     % each file has the same channels, concatenate along the time dimension
     % this requires careful bookkeeping of the sample indices
@@ -130,7 +130,7 @@ if iscell(filename)
       nsmp(i) = hdr{i}.nSamples*hdr{i}.nTrials;
     end
     offset = [0 cumsum(nsmp(1:end-1))];
-    
+
     event = cell(size(filename));
     for i=1:numel(filename)
       varargin = ft_setopt(varargin, 'header', hdr{i});
@@ -274,17 +274,17 @@ event = [];
 % please maintain this list in alphabetical order
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 switch eventformat
-  
+
   case {'4d' '4d_pdf', '4d_m4d', '4d_xyz'}
     if isempty(hdr)
       hdr = ft_read_header(filename, 'headerformat', eventformat);
     end
-    
+
     if isempty(chanindx)
       % auto-detect the trigger channels
       chanindx = match_str(hdr.label, 'TRIGGER');
     end
-    
+
     % read the trigger channels and do flank detection
     if isfield(hdr, 'orig') && isfield(hdr.orig, 'config_data') && (strcmp(hdr.orig.config_data.site_name, 'Glasgow') || strcmp(hdr.orig.config_data.site_name, 'Marseille')),
       trigger = read_trigger(filename, 'header', hdr, 'dataformat', '4d', 'begsample', flt_minsample, 'endsample', flt_maxsample, 'chanindx', chanindx, 'detectflank', detectflank, 'denoise', denoise, 'trigshift', trigshift, 'trigpadding', trigpadding, 'fix4d8192', true);
@@ -292,21 +292,21 @@ switch eventformat
       trigger = read_trigger(filename, 'header', hdr, 'dataformat', '4d', 'begsample', flt_minsample, 'endsample', flt_maxsample, 'chanindx', chanindx, 'detectflank', detectflank, 'denoise', denoise, 'trigshift', trigshift, 'trigpadding', trigpadding, 'fix4d8192', false);
     end
     event   = appendstruct(event, trigger);
-    
+
     respindx = match_str(hdr.label, 'RESPONSE');
     if ~isempty(respindx)
       response = read_trigger(filename, 'header', hdr, 'dataformat', '4d', 'begsample', flt_minsample, 'endsample', flt_maxsample, 'chanindx', respindx, 'detectflank', detectflank, 'denoise', denoise, 'trigshift', trigshift, 'trigpadding', trigpadding);
       event    = appendstruct(event, response);
     end
-    
+
   case 'bci2000_dat'
     % this requires the load_bcidat mex file to be present on the path
     ft_hastoolbox('BCI2000', 1);
-    
+
     if isempty(hdr)
[…]
