### 2023-09-13 2722e13
improve the parsing of the value for 'trial' events
