### 2023-09-13 28bc946
ENH - the trialinfo can also be a table
