### 2023-09-13 2816928
ENH - detect more of the known channel types from fif files, including CTF gradiometers
