+++ fileio/ft_chantype.m
@@ -166,41 +166,22 @@ elseif isheader && issubfield(input, 'orig.channames')
   end
   
 elseif isheader && issubfield(input, 'orig.chs.coil_type')
-  % this is for Neuromag or Babysquid systems
-  % all the chs.kinds and chs.coil_types are obtained from the MNE manual, p.210-211
-  for sel=find([input.orig.chs.kind]==1 & [input.orig.chs.coil_type]==2)'    % planar gradiometers
-    chantype(sel) = {'megplanar'}; % Neuromag-122 planar gradiometer
-  end
-  for sel=find([input.orig.chs.kind]==1 & [input.orig.chs.coil_type]==3012)' % planar gradiometers
-    chantype(sel) = {'megplanar'}; % Type T1 planar grad
-  end
-  for sel=find([input.orig.chs.kind]==1 & [input.orig.chs.coil_type]==3013)' % planar gradiometers
-    chantype(sel) = {'megplanar'}; % Type T2 planar grad
-  end
-  for sel=find([input.orig.chs.kind]==1 & [input.orig.chs.coil_type]==3014)' % planar gradiometers
-    chantype(sel) = {'megplanar'}; % Type T3 planar grad
-  end
-  for sel=find([input.orig.chs.kind]==1 & [input.orig.chs.coil_type]==3022)' % magnetometers
-    chantype(sel) = {'megmag'};    % Type T1 magenetometer
-  end
-  for sel=find([input.orig.chs.kind]==1 & [input.orig.chs.coil_type]==3023)' % magnetometers
-    chantype(sel) = {'megmag'};    % Type T2 magenetometer
-  end
-  for sel=find([input.orig.chs.kind]==1 & [input.orig.chs.coil_type]==3024)' % magnetometers
-    chantype(sel) = {'megmag'};    % Type T3 magenetometer
-  end
-  for sel=find([input.orig.chs.kind]==1 & [input.orig.chs.coil_type]==7001)' % axial gradiometer
-    chantype(sel) = {'megaxial'};
-  end
-  for sel=find([input.orig.chs.kind]==301)' % MEG reference channel, located far from head
-    chantype(sel) = {'ref'};
-  end
-  for sel=find([input.orig.chs.kind]==2)'   % EEG channels
-    chantype(sel) = {'eeg'};
-  end
-  for sel=find([input.orig.chs.kind]==201)' % MCG channels
-    chantype(sel) = {'mcg'};
-  end
+  % this is for fif files imported using neuromag_mne
+
+  % the kinds and coil_types were initially obtained from the MNE manual (p.210-211) and updated from the coil_def.dat file
+  chantype([input.orig.chs.kind]==1 & ismember([input.orig.chs.coil_type], [2, 3012, 3013, 3014, 3015])) = {'megplanar'};
+  chantype([input.orig.chs.kind]==1 & ismember([input.orig.chs.coil_type], [2000, 3022, 3023, 3024, 3025, 4001, 7002, 7003, 8001, 8002, 8101, 8201])) = {'megmag'};
+  chantype([input.orig.chs.kind]==1 & ismember([input.orig.chs.coil_type], [4002, 5001, 6001, 7001, 7501, 9001, 9101, 9102])) = {'megaxial'};
+  chantype([input.orig.chs.kind]==1 & ismember([input.orig.chs.coil_type], [4003, 5002, 6002, 7004, 7502, 4004, 4005, 5003, 5004, 7503])) = {'ref'};
+
+  chantype([input.orig.chs.kind]==2)   = {'eeg'};
+  chantype([input.orig.chs.kind]==201) = {'mcg'};
+  chantype([input.orig.chs.kind]==202) = {'eog'};
[…]
