### 2023-09-13 0962f5d
ENH - make code more robust for CTF data that originates from a fif file
