### 2023-09-13 0aa4735
ENH - made tests more consistent
