### 2023-09-18 #5566 @gflofst
I've had two bad weeks including technology failure. Hopefully this week I'll be able to get it done.

Best,

Jay
