### 2023-09-19 #5566 @crsl4
Thank you, Jay! No worries! We are experiencing delays on several fronts.
