### 2023-09-26 a2c40e5
give an explicit error when cfg.trl is incorrect
