### 2023-09-26 777d573
ENH - changed docstring, and fixed small code typo
