### 2023-10-05 690e5c7
FIX #2326 - fixed time range selection for select_topoplotER in ft_multiplotER
- renamed viewmode layout into topographic
- do not recursively store the topographic layout in the butterfly layout
- added check and informative error for labelcharN
- make chanindx selection in pointsymbol, pointsize, pointcolor
