+++ ft_prepare_layout.m
@@ -72,8 +72,8 @@ function [layout, cfg] = ft_prepare_layout(cfg, data)
 %                      specificies channels to use for determining channel box size (default = 'all', recommended for MEG/EEG, a selection is recommended for iEEG)
 %   cfg.skipscale   = 'yes' or 'no', whether the scale should be included in the layout or not (default = 'no')
 %   cfg.skipcomnt   = 'yes' or 'no', whether the comment should be included in the layout or not (default = 'no')
-%   cfg.color       = 'spatial', Nx3 matrix, or [] (default). If not empty, an Nx3 color matrix will be added to the layout, based on the
-%                     positions of the electrodes, or based on the specified matrix
+%   cfg.color       = empty, 'spatial', or Nx3 matrix, if non-empty, an Nx3 color matrix based on the position 
+% %                   of the sensors will be added (default = [])
 %
 % If you use cfg.headshape or cfg.mri to create a headshape outline, the input
 % geometry should be expressed in the same units and coordinate system as the input
@@ -94,12 +94,6 @@ function [layout, cfg] = ft_prepare_layout(cfg, data)
 %   cfg.layout = 'circular'   will distribute the channels on a circle
 %   cfg.width  = scalar (default is automatic)
 %   cfg.height = scalar (default is automatic)
-%   cfg.layouttopo = filename or struct (default is empty)
-%
-% For a butterfly layout, the option cfg.layouttopo will add an extra field to the layout, containing the spatial layout
-% of the sensor array. This can be used to plot the spatial distribution of the color-coded channels, as in ft_multiplotER
-% with cfg.viewmode = 'butterfly'. If it's defined empty, but if the input data argument contains a sensor description, then it
-% will be created from this
 %
 % For an sEEG shaft the option cfg.layout='vertical' or 'horizontal' is useful to
 % represent the channels in a linear sequence . In this case you can also specify the
@@ -208,8 +202,7 @@ cfg.width        = ft_getopt(cfg, 'width',      []);
 cfg.height       = ft_getopt(cfg, 'height',     []);
 cfg.commentpos   = ft_getopt(cfg, 'commentpos', 'layout');
 cfg.scalepos     = ft_getopt(cfg, 'scalepos',   'layout');
-cfg.color        = ft_getopt(cfg, 'color');
-cfg.layouttopo   = ft_getopt(cfg, 'layouttopo', []);
+cfg.color        = ft_getopt(cfg, 'color',      []);
 
 if isempty(cfg.skipscale)
   if ischar(cfg.layout) && any(strcmp(cfg.layout, {'ordered', 'vertical', 'horizontal', 'butterfly', 'circular'}))
@@ -371,32 +364,26 @@ elseif isequal(cfg.layout, 'butterfly')
   layout.mask    = {};
   layout.outline = {};
   
-  layout.label{end+1}  = 'SCALE';
-  layout.pos(end+1, :) = layout.pos(1,:);
-  layout.width(end+1)  = layout.width(1);
-  layout.height(end+1) = layout.height(1);
-  
-  tmpcfg = removefields(cfg, {'layout' 'layouttopo'});
-  tmpcfg.skipscale = 'yes';
-  tmpcfg.skipcmnt  = 'yes';
-  if ~isempty(cfg.layouttopo)
-    tmpcfg.layout = cfg.layouttopo;
-  end
-  try
-    % failure may happen if the tmpcfg.layout = [], and if the data does
-    % not contain a sensor description.
-    tmplayout = ft_prepare_layout(tmpcfg, data);
-    layout.layout = tmplayout;
[…]
