### 2023-10-05 b8f1e70
more explicit loading of test data
