### 2023-10-06 6ddee7d
FIX #2328 - don't realign empty fiducial positions, and don't add them
