+++ external/mayo_mef/med/@MultiscaleElectrophysiologyData_1p0/importSignal.m
@@ -1,0 +1,186 @@
+function [x, t] = importSignal(this, start_end, st_unit, filepath, filename, options)
+    % MULTISCALEELECTROPHYSIOLOGYDATA_1P0.IMPORTSIGNAL imports MED 1.0 signal into MATLAB
+    %
+    % Syntax:
+    %   [x, t] = importSignal(this)
+    %   [x, t] = importSignal(__, start_end)
+    %   [x, t] = importSignal(__, start_end, st_unit)
+    %   [x, t] = importSignal(__, start_end, st_unit, filepath)
+    %   [x, t] = importSignal(__, start_end, st_unit, filepath, filename)
+    %   [x, t] = importSignal(__, 'Level1Password', level_1_pw)
+    %   [x, t] = importSignal(__, 'Level2Password', level_2_pw)
+    %   [x, t] = importSignal(__, 'AccessLevel', access_level)
+    %
+    % Input(s):
+    %   this            - [obj] MultiscaleElectrophysiologyData_1p0 object
+    %   start_end       - [1 x 2 array] (opt) [start time/index, end time/index] of
+    %                     the signal to be extracted fromt the file (default:
+    %                     the entire signal)
+    %   st_unit         - [str] (opt) unit of start_end: 'Index' (default), 'uUTC',
+    %                     'Second', 'Minute', 'Hour', and 'Day'
+    %   filepath        - [str] (opt) directory of the session
+    %   filename        = [str] (opt) filename of the channel
+    %   level_1_pw      - [str] (para) password of level 1 (default = this.Level1Password)
+    %   level_2_pw      - [str] (para) password of level 2 (default = this.Level2Password)
+    %   access_level    - [str] (para) data decode level to be used
+    %                     (default = this.AccessLevel)
+    %
+    % Output(s):
+    %   x               - [num array] extracted signal
+    %   t               - [num array] time indices of the signal in the file
+    %
+    % Example:
+    %
+    % Note:
+    %   Import data from one channel of MED 1.0 file into MatLab.
+    %
+    % References:
+    %
+    % See also .
+
+    % Copyright 2023 Richard J. Cui. Created: Thu 04/20/2023 12:24:33.818 AM
+    % $Revision: 0.7 $  $Date: Fri 10/06/2023 12:19:28.864 AM $
+    %
+    % Rocky Creek Dr. NE
+    % Rochester, MN 55906, USA
+    %
+    % Email: richard.cui@utoronto.ca
+
+    % ======================================================================
+    % parse inputs
+    % ======================================================================
+    arguments
+        this (1, 1) MultiscaleElectrophysiologyData_1p0
+        start_end (1, 2) double {mustBeNonnegative} = [0, inf]
+        st_unit (1, 1) string {mustBeMember(st_unit, ["index", "uutc", "second", "minute", "hour", "day"])} = "index"
+        filepath (1, :) string = string.empty(1, 0)
+        filename (1, :) string = string.empty(1, 0)
+    end % positional
+
+    arguments
+        options.Level1Password (1, 1) string = this.Level1Password
+        options.Level2Password (1, 1) string = this.Level2Password
+        options.AccessLevel (1, 1) double = this.AccessLevel
+    end % optional
+
+    % ======================================================================
+    % main
+    % ======================================================================
+    % parameters
+    % ----------
+    l1_pw = options.Level1Password;
+    l2_pw = options.Level2Password;
+    al = options.AccessLevel;
+
+    % password
+    % --------
+    if l1_pw == ""
+        l1_pw = this.Level1Password;
+    else
+        this.Level1Password = l1_pw;
+    end % if
+
+    if l2_pw == ""
+        l2_pw = this.Level2Password;
+    else
[…]
