### 2023-10-10 d03ffce
FIX - fixed typo in tutorial path in test script
