### 2023-10-10 c1e5d50
FIX - fixed regression error in test_pull2099 with topographic linecolors
