### 2023-10-11 4c4b1d4
fixed bugs for code example
