+++ external/mayo_mef/med/@MEDSession/relative2absTimePoint.m
@@ -22,7 +22,7 @@ function abs_time = relative2absTimePoint(this, rel_time, unit)
     % See also .
 
     % Copyright 2023 Richard J. Cui. Created: Sun 09/17/2023 10:40:37.047 PM
-    % $Revision: 0.2 $  $Date: Mon 10/09/2023 12:22:37.852 AM $
+    % $Revision: 0.3 $  $Date: Tue 10/10/2023 10:13:35.497 PM $
     %
     % Rocky Creek Dr. NE
     % Rochester, MN 55906, USA
@@ -35,14 +35,14 @@ function abs_time = relative2absTimePoint(this, rel_time, unit)
     arguments
         this (1, 1) MEDSession
         rel_time (1, :) double
-        unit (1, :) char {mustBeMember(unit, {'index', 'uutc', 'msec', 'second', 'minute', 'hour', 'day'})} = 's'
+        unit (1, :) char {mustBeMember(unit, {'index', 'uutc', 'msec', 'second', 'minute', 'hour', 'day'})} = 'index'
     end % positional
 
     % ======================================================================
     % main
     % ======================================================================
     offset = this.getSessionRecordOffset(unit);
-    abs_time = rel_time + double(offset);
+    abs_time = rel_time + offset;
 
 end % function relative2absTimePoint
 
