### 2023-10-11 ffca3bf
updated testing script
