### 2023-10-12 18db425
ENH - use local data, don't use parpool
