+++ test/test_mayo_mef.m
@@ -95,35 +95,40 @@ function test_mayo_mef()
     mef30_pw = struct('Level1Password', 'password1', 'Level2Password', ...
         'password2', 'AccessLevel', 2);
 
-    % low-level testing
-    fprintf('low-level testing...\n')
-    hdr = ft_read_header(mef30_data, 'password', mef30_pw);
+    try
+        % low-level testing
+        fprintf('low-level testing...\n')
+        hdr = ft_read_header(mef30_data, 'password', mef30_pw);
 
-    if isempty(hdr)
-        warning('failed to read header of MEF 3.0 sample dataset')
-    end
+        if isempty(hdr)
+            warning('failed to read header of MEF 3.0 sample dataset')
+        end
 
-    dat = ft_read_data(mef30_data, 'password', mef30_pw);
+        dat = ft_read_data(mef30_data, 'password', mef30_pw);
 
-    if isempty(dat)
-        warning('failed to read data of MEF 3.0 sample dataset')
-    end
+        if isempty(dat)
+            warning('failed to read data of MEF 3.0 sample dataset')
+        end
 
-    evt = ft_read_event(mef30_data, 'password', mef30_pw);
+        evt = ft_read_event(mef30_data, 'password', mef30_pw);
 
-    if isempty(evt)
-        warning('failed to read event of MEF 3.0 sample dataset')
-    end
+        if isempty(evt)
+            warning('failed to read event of MEF 3.0 sample dataset')
+        end
 
-    % high-level testing
-    fprintf('high-level testing...\n')
-    cfg = [];
-    cfg.dataset = mef30_data;
-    cfg.password = mef30_pw;
-    data = ft_preprocessing(cfg);
+        % high-level testing
+        fprintf('high-level testing...\n')
+        cfg = [];
+        cfg.dataset = mef30_data;
+        cfg.password = mef30_pw;
+        data = ft_preprocessing(cfg);
 
-    if isempty(data)
-        warning('failed to read data of MEF 3.0 sample dataset in high-level testing')
+        if isempty(data)
+            warning('failed to read data of MEF 3.0 sample dataset in high-level testing')
+        end
+
+    catch
+        warning('failed to read MEF 3.0 sample dataset')
     end
 
     % MED version 1.0
@@ -133,44 +138,49 @@ function test_mayo_mef()
     disp('testing MED version 1.0...')
     disp('--------------------------')
 
-    % the password of MED 1.0 sample dataset
-    med10_pw = struct('Level1Password', 'L1_password', 'Level2Password', ...
-        'L2_password', 'AccessLevel', 2);
+    try
+        % the password of MED 1.0 sample dataset
+        med10_pw = struct('Level1Password', 'L1_password', 'Level2Password', ...
+            'L2_password', 'AccessLevel', 2);
 
-    % low-level testing
-    fprintf('low-level testing...\n')
-    hdr = ft_read_header(med10_data, 'password', med10_pw);
+        % low-level testing
+        fprintf('low-level testing...\n')
+        hdr = ft_read_header(med10_data, 'password', med10_pw);
 
-    if isempty(hdr)
-        warning('failed to read header of MED 1.0 sample dataset')
+        if isempty(hdr)
+            warning('failed to read header of MED 1.0 sample dataset')
+        end
+
+        dat = ft_read_data(med10_data, 'password', med10_pw);
+
+        if isempty(dat)
+            warning('failed to read data of MED 1.0 sample dataset')
+        end
+
+        evt = ft_read_event(med10_data, 'password', med10_pw);
+
+        if isempty(evt)
+            warning('failed to read event of MED 1.0 sample dataset')
+        end
+
+        % high-level testing
+        fprintf('high-level testing...\n')
+        cfg = [];
[…]
