+++ external/mayo_mef/mex_mef/make_mex_mef.m
@@ -1,7 +1,7 @@
 % Compile mex files required to process MEF files
 
 % Copyright 2019-2020 Richard J. Cui. Created: Wed 05/29/2019  9:49:29.694 PM
-% $Revision: 0.8 $  $Date: Fri 06/05/2020 10:27:44.467 AM $
+% $Revision: 0.9 $  $Date: Thu 10/12/2023 17:24:54.023 PM $
 %
 % Multimodel Neuroimaging Lab (Dr. Dora Hermes)
 % Mayo Clinic St. Mary Campus
@@ -14,7 +14,7 @@
 % =========================================================================
 % go to mex_mef
 mex_mef = fileparts(mfilename('fullpath')); % directory of make_mex_mef.m assumed in mex_mef
-mayo_mef = fileparts(mex_mef); % mayo_mef directory 
+mayo_mef = fileparts(mex_mef); % mayo_mef directory
 
 % =========================================================================
 % processing mex for MEF 2.0 file
@@ -23,54 +23,62 @@ mayo_mef = fileparts(mex_mef); % mayo_mef directory
 % =========================================================================
 % processing mex for MEF 2.1 file
 % =========================================================================
-cd([mex_mef,filesep,'mef_2p1']) % assume 'mef_2p1' is the subdirectory
+try
+    cd([mex_mef, filesep, 'mef_2p1']) % assume 'mef_2p1' is the subdirectory
 
-fprintf('===== Compiling c-mex for MEF 2.1 data =====\n')
-fprintf('Building read_mef_header_2p1.mex*\n')
-mex -output read_mef_header_2p1 ...
-    read_mef_header_mex_2p1.c ...
-    mef_lib_2p1.c
-movefile('read_mef_header_2p1.mex*',mayo_mef)
+    fprintf('===== Compiling c-mex for MEF 2.1 data =====\n')
+    fprintf('Building read_mef_header_2p1.mex*\n')
+    mex -output read_mef_header_2p1 ...
+        read_mef_header_mex_2p1.c ...
+        mef_lib_2p1.c
+    movefile('read_mef_header_2p1.mex*', mayo_mef)
 
-fprintf('\n')
-fprintf('Building decompress_mef_2p1.mex*\n')
-mex -output decompress_mef_2p1 ...
-    decompress_mef_mex_2p1.c ...
-    mef_lib_2p1.c
-movefile('decompress_mef_2p1.mex*',mayo_mef)
+    fprintf('\n')
+    fprintf('Building decompress_mef_2p1.mex*\n')
+    mex -output decompress_mef_2p1 ...
+        decompress_mef_mex_2p1.c ...
+        mef_lib_2p1.c
+    movefile('decompress_mef_2p1.mex*', mayo_mef)
 
-cd(mayo_mef)
+    cd(mayo_mef)
+catch
+    warning('Cannot build MEX file for MEF 2.1 data.')
+end
 
 % =========================================================================
 % processing mex for MEF 3.0 file
 % =========================================================================
-fprintf('\n')
-fprintf('===== Compiling c-mex for MEF 3.0 data =====\n')
+try
+    fprintf('\n')
+    fprintf('===== Compiling c-mex for MEF 3.0 data =====\n')
 
-% cd([mex_mef, filesep, 'mef_3p0'])
-% fprintf('Building read_mef_header_mex_3p0.mex*\n')
-% mex -output read_mef_header_mex_3p0 ...
-%     read_mef_header_mex_3p0.c ...
-%     matmef/matmef_mapping.c ...
-%     matmef/mex_datahelper.c
[…]
