### 2023-10-13 c522193
ENH - add useinside option
