+++ plotting/ft_plot_mesh.m
@@ -90,7 +90,6 @@ if numel(mesh)>1
 end
 
 % get the optional input arguments
-vertexcolor  = ft_getopt(varargin, 'vertexcolor');
 if isfield(mesh, 'tri') && size(mesh.tri,1)>10000
   facecolor    = ft_getopt(varargin, 'facecolor',   'cortex_light');
   edgecolor    = ft_getopt(varargin, 'edgecolor',   'none');
@@ -98,24 +97,26 @@ else
   facecolor    = ft_getopt(varargin, 'facecolor',   'white');
   edgecolor    = ft_getopt(varargin, 'edgecolor',   'k');
 end
-faceindex     = ft_getopt(varargin, 'faceindex',   false);
-vertexindex   = ft_getopt(varargin, 'vertexindex', false);
-vertexsize    = ft_getopt(varargin, 'vertexsize',  10);
-vertexmarker  = ft_getopt(varargin, 'vertexmarker', '.');
-facealpha     = ft_getopt(varargin, 'facealpha',   1);
-edgealpha     = ft_getopt(varargin, 'edgealpha',   1);
-edgelinewidth = ft_getopt(varargin, 'edgelinewidth', .5);
-material_     = ft_getopt(varargin, 'material');        % note the underscore, there is also a material function
-tag           = ft_getopt(varargin, 'tag',         '');
-surfaceonly   = ft_getopt(varargin, 'surfaceonly');     % default is handled below
+vertexcolor   = ft_getopt(varargin, 'vertexcolor');
+vertexindex   = ft_getopt(varargin, 'vertexindex',   false);
+vertexsize    = ft_getopt(varargin, 'vertexsize',    10);
+vertexmarker  = ft_getopt(varargin, 'vertexmarker',  '.');
+faceindex     = ft_getopt(varargin, 'faceindex',     false);
+facealpha     = ft_getopt(varargin, 'facealpha',     1);
+edgealpha     = ft_getopt(varargin, 'edgealpha',     1);
+edgelinewidth = ft_getopt(varargin, 'edgelinewidth', 0.5);
+material_     = ft_getopt(varargin, 'material');         % note the underscore, there is also a material function
+tag           = ft_getopt(varargin, 'tag',           '');
+surfaceonly   = ft_getopt(varargin, 'surfaceonly');      % default is handled below
 unit          = ft_getopt(varargin, 'unit');
-axes_         = ft_getopt(varargin, 'axes', false);     % do not confuse with built-in function
-clim          = ft_getopt(varargin, 'clim');
+axes_         = ft_getopt(varargin, 'axes',      false); % do not confuse with built-in function
 alphalim      = ft_getopt(varargin, 'alphalim');
-alphamapping  = ft_getopt(varargin, 'alphamap', 'rampup');
-cmap          = ft_getopt(varargin, 'colormap');
+alphamapping  = ft_getopt(varargin, 'alphamap',  'rampup');
 maskstyle     = ft_getopt(varargin, 'maskstyle', 'opacity');
-contour       = ft_getopt(varargin, 'contour',   []);
+cmap          = ft_getopt(varargin, 'colormap');
+clim          = ft_getopt(varargin, 'clim');
+contour       = ft_getopt(varargin, 'contour');
+insideonly    = ft_getopt(varargin, 'insideonly', false);
 
 % these have to do with the font
 fontcolor       = ft_getopt(varargin, 'fontcolor', 'k');  % default is black
@@ -135,14 +136,20 @@ hastet   = isfield(mesh, 'tet');   % tetraheders as a Mx4 matrix with vertex ind
 hashex   = isfield(mesh, 'hex');   % hexaheders  as a Mx8 matrix with vertex indices
[…]
