### 2023-10-14 9f73f3e
WIP - some adjustments to improve code flow, to be finalized after discussing
