+++ inverse/ft_inverse_sam.m
@@ -30,6 +30,7 @@ function [estimate] = ft_inverse_sam(sourcemodel, sens, headmodel, dat, C, varar
 %
 % See also FT_SOURCEANALYSIS, FT_PREPARE_HEADMODEL, FT_PREPARE_SOURCEMODEL
 
+% Copyright (C) 2023,      Malte Hoeltershinken
 % Copyright (C) 2009,      Robert Oostenveld
 % Copyright (C) 2005-2009, Arjan Hillebrand
 % Copyright (C) 2005-2009, Gareth Barnes
@@ -63,12 +64,24 @@ leadfieldopt = ft_setopt(leadfieldopt, 'normalize',      ft_getopt(varargin, 'no
 leadfieldopt = ft_setopt(leadfieldopt, 'normalizeparam', ft_getopt(varargin, 'normalizeparam'));
 leadfieldopt = ft_setopt(leadfieldopt, 'weight',         ft_getopt(varargin, 'weight'));
 
-if strcmp(fixedori, 'moiseev') && isempty(noisecov)
-  ft_error('fixedori method ''moiseev'' requires a noise covariance matrix to be specified');
-end
+switch fixedori
+  case 'robert'
+    ft_error('fixedori=''robert'' is not supported anymore. The exact same functionality is implemented in ft_inverse_lcmv, using fixedori = ''yes''');
+  case 'gareth'
+    % placeholder: give some user feedback, either error or not, to be discussed, depends on whether it will be merged
+  case 'moiseev'
+    % placeholder: give some user feedback, either error or not, to be discussed, depends on whether it will/can be merged
+    % (apart from the toi functionality)
+    if isempty(noisecov)
+      ft_error('fixedori method ''moiseev'' requires a noise covariance matrix to be specified');
+    end
+  case 'spinning'
+    ft_info('fixedori=''spinning'' has been adjusted to provide an exact estimate of the optimal orientation, see the code for details');
+    leadfieldopt = ft_setopt(leadfieldopt, 'backproject', 'no');
 
-if ~strcmp(fixedori, 'spinning')
-  leadfieldopt = ft_setopt(leadfieldopt, 'backproject', 'no');
+
+  otherwise
+    ft_error('unsupported value for fixedori option');
 end
 
 if strcmp(fixedori, 'spinning')
@@ -117,7 +130,7 @@ origpos    = sourcemodel.pos;
 sourcemodel.pos = sourcemodel.pos(originside,:);
 
 if hasmom
-  ft_warning('this probably will not work because of the spinning/nonspinning source orientation estimate being part of this code');
+  ft_warning('a prespecified mom in the sourcemodel will probably not work because of the spinning/nonspinning source orientation estimate being part of this code');
   sourcemodel.mom = sourcemodel.mom(:, sourcemodel.inside);
 end
 
@@ -174,14 +187,10 @@ else
 end
 
 % the inverse only has to be computed once for all dipoles
-inv_cov   = ft_inv(C, invopt{:});
+invC = ft_inv(C, invopt{:});
 if isempty(noisecov)
   noisecov = noise * eye(size(C));
 end
-
-
-% the angles are the same for all dipole locations
-all_angles = 0:pi/72:pi;
 
 % start the scanning
 ft_progress('init', feedback, 'scanning grid');
@@ -191,8 +200,8 @@ for i=1:size(sourcemodel.pos,1)
   vox_pos = sourcemodel.pos(i,:);
   
   if hasfilter
-    SAMweights = sourcemodel.filter{i};
-    if size(SAMweights,1)>1
+    filt = sourcemodel.filter{i};
+    if size(filt,1)>1
[…]
