+++ inverse/ft_inverse_sam.m
@@ -15,11 +15,15 @@ function [estimate] = ft_inverse_sam(sourcemodel, sens, headmodel, dat, C, varar
 %   estimate    contains the estimated source parameters
 %
 % Additional input arguments should be specified as key-value pairs and can include
-%   'meansphereorigin'
 %   'feedback'
-%   'fixedori'
+%   'fixedori'          deprecated, control behaviour via 'reducerank' instead
 %   'noisecov'
 %   'toi'
+%
+% If no orientation is specified, the SAM beamformer will try to estimate the orientation from the data.
+% The beamformer will either try to estimate the whole orientation, or only its tangential component.
+% This is controlled by the 'reducerank' parameter. For reducerank=3, the whole orientation is estimated,
+% and for reducerank=2 only the tangential component is estimated.
 %
 % These options influence the forward computation of the leadfield
 %   'reducerank'      = 'no' or number  (default = 3 for EEG, 2 for MEG)
@@ -43,9 +47,8 @@ if mod(nargin-5,2)
 end
 
 % get the optional input arguments, or use defaults
-meansphereorigin  = ft_getopt(varargin, 'meansphereorigin');
 feedback          = ft_getopt(varargin, 'feedback', 'text');
-fixedori          = ft_getopt(varargin, 'fixedori', 'spinning');
+fixedori          = ft_getopt(varargin, 'fixedori');
 noisecov          = ft_getopt(varargin, 'noisecov');
 toi               = ft_getopt(varargin, 'toi');
 
@@ -56,50 +59,19 @@ invopt = ft_setopt(invopt, 'kappa',     ft_getopt(varargin, 'kappa'));
 invopt = ft_setopt(invopt, 'tolerance', ft_getopt(varargin, 'tol'));
 invopt = ft_setopt(invopt, 'method',    ft_getopt(varargin, 'invmethod'));
 
-% construct the low-level options for the leadfield computation as key-value pairs, these are passed to FT_COMPUTE_LEADFIELD
-leadfieldopt = {};
-leadfieldopt = ft_setopt(leadfieldopt, 'reducerank',     ft_getopt(varargin, 'reducerank'));
-leadfieldopt = ft_setopt(leadfieldopt, 'backproject',    ft_getopt(varargin, 'backproject'));
-leadfieldopt = ft_setopt(leadfieldopt, 'normalize',      ft_getopt(varargin, 'normalize'));
-leadfieldopt = ft_setopt(leadfieldopt, 'normalizeparam', ft_getopt(varargin, 'normalizeparam'));
-leadfieldopt = ft_setopt(leadfieldopt, 'weight',         ft_getopt(varargin, 'weight'));
-
-switch fixedori
-  case 'robert'
-    ft_error('fixedori=''robert'' is not supported anymore. The exact same functionality is implemented in ft_inverse_lcmv, using fixedori = ''yes''');
-  case 'gareth'
-    % placeholder: give some user feedback, either error or not, to be discussed, depends on whether it will be merged
-  case 'moiseev'
-    % placeholder: give some user feedback, either error or not, to be discussed, depends on whether it will/can be merged
-    % (apart from the toi functionality)
-    if isempty(noisecov)
-      ft_error('fixedori method ''moiseev'' requires a noise covariance matrix to be specified');
-    end
-  case 'spinning'
-    ft_info('fixedori=''spinning'' has been adjusted to provide an exact estimate of the optimal orientation, see the code for details');
-    leadfieldopt = ft_setopt(leadfieldopt, 'backproject', 'no');
-
-
-  otherwise
-    ft_error('unsupported value for fixedori option');
-end
-
-if strcmp(fixedori, 'spinning')
-  % determine the mean sphere origin, required for spinning
[…]
