### 2023-10-19 32f6393
Update SAM test
