+++ test/test_beamformer_sam.m
@@ -103,26 +103,24 @@ cfg.method              = 'sam';
 cfg.grid                = grid;
 cfg.headmodel           = vol;
 cfg.grad                = data_signal.grad;
-%cfg.sam.toi            = [.1 .2];
 cfg.sam.lambda          = '5%';
 cfg.sam.projectmom      = 'no';
 cfg.sam.keepfilter      = 'yes';
 cfg.sam.keepori         = 'yes';
 cfg.sam.projectnoise    = 'yes';
-cfg.sam.noisecov        = tlck_N.cov;
-cfg.sam.fixedori        = 'moiseev';
-source_new_ft           = ft_sourceanalysis(cfg, tlck_S);
-ori_newImpl             = cat(2,source_new_ft.avg.ori{:});
 
-cfg.sam.fixedori            = 'gareth';
+% default case, equivalent to the fixedori approach previously named 'moiseev'
+cfg.sam.noisecov        = tlck_N.cov;
+source_moiseev_ft       = ft_sourceanalysis(cfg, tlck_S);
+ori_moiseev             = cat(2,source_moiseev_ft.avg.ori{:});
+
+% if no noise covariance is specified, it is estimated using a multiple of the identity matrix.
+% this is equivalent to the fixedori approach previously named 'gareth'
+cfg.sam                 = rmfield(cfg.sam, 'noisecov');
 source_gar_ft           = ft_sourceanalysis(cfg, tlck_S);
 ori_gar                 = cat(2,source_gar_ft.avg.ori{:});
 
-cfg.sam.fixedori            = 'robert';
-source_rob_ft           = ft_sourceanalysis(cfg, tlck_S);
-ori_rob                 = cat(2,source_rob_ft.avg.ori{:});
-
-ori_All = [ori_newImpl; ori_gar; ori_rob];
+ori_All = [ori_moiseev; ori_gar];
 
 
 %% estimates moms per trials
@@ -132,7 +130,6 @@ cfg.method              = 'sam';
 cfg.grid                = grid;
 cfg.headmodel           = vol;
 cfg.grad                = data_signal.grad;
-%cfg.sam.toi            = [.1 .2];
 cfg.sam.lambda          = '5%';
 cfg.sam.keepfilter      = 'yes';
 cfg.sam.keepori         = 'yes';
@@ -141,20 +138,22 @@ cfg.sam.keepmom         = 'yes';  % saves the time series
 cfg.sam.keeptrials      = 'yes'; %'no' or 'yes
 cfg.sam.keepleadfield   = 'no';
 cfg.sam.keepfilter      = 'no';
-cfg.sam.fixedori        = 'moiseev';
+
+% default case, equivalent to the fixedori approach previously named 'moiseev'
 cfg.sam.noisecov        = tlck_N.cov;
-cfg.grid.filter         = source_new_ft.avg.filter;
-source_new_mom          = ft_sourceanalysis(cfg, tlck_S);
+cfg.grid.filter         = source_moiseev_ft.avg.filter;
+source_moiseev_mom      = ft_sourceanalysis(cfg, tlck_S);
  
-cfg.sam.fixedori        = 'gareth';
+
+% if no noise covariance is specified, it is estimated using a multiple of the identity matrix.
+% this is equivalent to the fixedori approach previously named 'gareth'
+cfg.sam                 = rmfield(cfg.sam, 'noisecov');
 cfg.grid.filter         = source_gar_ft.avg.filter;
 source_gar_mom          = ft_sourceanalysis(cfg, tlck_S);
 
[…]
