### 2023-10-23 71fde0a
ENH - small updates to survive more strict cfg checking
