### 2023-10-25 5005906
implemented realtime interface to sensys fluxgate magnetometer
