+++ realtime/example/ft_realtime_sensysproxy.m
@@ -1,0 +1,160 @@
+function ft_realtime_sensysproxy(cfg)
+
+% FT_REALTIME_SENSYSPROXY reads real-time continuous fluxgate magnetometer
+% data from the serial interface provided by the Sensis FGM3D TD
+% application and writes the data to the FieldTrip buffer.
+%
+% Use as
+%   ft_realtime_sensysproxy(cfg)
+%
+% The configuration should contain
+%   cfg.filename    = string, name of the serial port (default = 'COM2')
+%   cfg.blocksize   = number, in seconds (default = 0.1)
+%   cfg.fsample     = sampling frequency (default = 1000)
+%
+% The target to write the data to is configured as
+%   cfg.target.datafile      = string, target destination for the data (default = 'buffer://localhost:1972')
+%   cfg.target.dataformat    = string, default is determined automatic
+%
+% To stop this realtime function, you have to press Ctrl-C
+%
+% See also FT_REALTIME_SIGNALPROXY, FT_REALTIME_SIGNALVIEWER
+
+% Copyright (C) 2023, Robert Oostenveld
+%
+% This file is part of FieldTrip, see http://www.fieldtriptoolbox.org
+% for the documentation and details.
+%
+%    FieldTrip is free software: you can redistribute it and/or modify
+%    it under the terms of the GNU General Public License as published by
+%    the Free Software Foundation, either version 3 of the License, or
+%    (at your option) any later version.
+%
+%    FieldTrip is distributed in the hope that it will be useful,
+%    but WITHOUT ANY WARRANTY; without even the implied warranty of
+%    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+%    GNU General Public License for more details.
+%
+%    You should have received a copy of the GNU General Public License
+%    along with FieldTrip. If not, see <http://www.gnu.org/licenses/>.
+%
+% $Id$
+
+cfg = ft_checkconfig(cfg);
+
+% set the defaults
+cfg.filename          = ft_getopt(cfg, 'filename', 'COM2');
+cfg.baudrate          = ft_getopt(cfg, 'baudrate', 921600);
+cfg.blocksize         = ft_getopt(cfg, 'blocksize', 0.1);
+cfg.fsample           = ft_getopt(cfg, 'fsample', 400);
+cfg.target            = ft_getopt(cfg, 'target');
+cfg.target.datafile   = ft_getopt(cfg.target, 'datafile', 'buffer://localhost:1972');
+cfg.target.dataformat = ft_getopt(cfg.target, 'dataformat', []);
+
+% these are not used at the moment, the defaults seem to work fine
+cfg.databits    = 8;
+cfg.flowcontrol = 'none';
+cfg.stopbits    = 1;
+cfg.parity      = 'none';
+
+% ensure that each block is an integer number of samples
+cfg.blocksize = round(cfg.blocksize*cfg.fsample)/cfg.fsample;
+
+% share the configuration details with the callback function
+setappdata(0, 'cfg', cfg)
+
+% the unicorn uses the SPP protocol, i.e. serial-over-bluetooth
+% open the serial device
+sensys = serialport(cfg.filename, cfg.baudrate);
+
+%%
+% this takes care of cleanup
+
+cleanup = onCleanup(@()myCleanupFun(sensys));
+
+% make sure the persistent variables are not reused from the last call
[…]
