+++ realtime/online_meg/ft_realtime_fieldnulling.m
@@ -1,0 +1,250 @@
+function ft_realtime_fieldnulling(cfg)
+
+% FT_REALTIME_FIELDNULLING is a real-time application to drive the nulling
+% coils in the magnetically shielded room.
+%
+% Use as
+%   ft_realtime_fieldnulling(cfg)
+%
+% The configuration should contain
+%   cfg.serialport  = string, name of the serial port (default = 'COM2')
+%   cfg.fsample     = sampling frequency (default = 1000)
+
+% these are used by the ft_preamble/ft_postamble function and scripts
+ft_revision = '$Id$';
+ft_nargin   = nargin;
+ft_nargout  = nargout;
+
+% do the general setup of the function
+ft_defaults
+ft_preamble init
+ft_preamble debug
+ft_preamble provenance
+
+% the ft_abort variable is set to true or false in ft_preamble_init
+if ft_abort
+  % do not continue function execution in case the outputfile is present and the user indicated to keep it
+  return
+end
+
+% set the defaults
+cfg.serialport        = ft_getopt(cfg, 'serialport', 'COM2');
+cfg.baudrate          = ft_getopt(cfg, 'baudrate', 921600);
+cfg.fsample           = ft_getopt(cfg, 'fsample', 400);
+
+% these are not used at the moment, the defaults seem to work fine
+cfg.databits    = 8;
+cfg.flowcontrol = 'none';
+cfg.stopbits    = 1;
+cfg.parity      = 'none';
+
+%% set up the serial connection to the fluxgate sensor
+
+% % open the serial device
+% fluxgate = serialport(cfg.serialport, cfg.baudrate);
+% cleanup = onCleanup(@()serial_cleanup(fluxgate));
+% 
+% %  make sure the persistent variables are not reused from the last call
+% clear serial_callback
+% 
+% configureTerminator(fluxgate, "LF");
+% configureCallback(fluxgate, "terminator", @serial_callback);
+
+%% set up the digital-to-analog converter
+
+% % Create a DataAcquisition object for the specified vendor.
+% dac = daq("ni");
+% 
+% % Add channels and set channel properties, if any.
+% addoutput(dac,"cDAQ1Mod1","ao0","Voltage");
+% addoutput(dac,"cDAQ1Mod1","ao1","Voltage");
+% addoutput(dac,"cDAQ1Mod1","ao2","Voltage");
+% addoutput(dac,"cDAQ1Mod1","ao3","Voltage");
+% addoutput(dac,"cDAQ1Mod1","ao4","Voltage");
+% addoutput(dac,"cDAQ1Mod1","ao5","Voltage");
+% 
+% % Output the specified DC amplitude on each channel.
+% dcOutput = [1 2 3 4 5 6];
+% write(dac, dcOutput);
+
+%%
+
+% open a new figure with the specified settings
+fig = open_figure(keepfields(cfg, {'figure', 'position', 'visible', 'renderer', 'figurename', 'title'}));
+
+% store the configuration details and in the application
+setappdata(fig, 'cfg', cfg);
+setappdata(fig, 'current', [0 0 0 0 0 0 ]);
+
[…]
