### 2023-10-27 563b3d9
duplicated private function
