### 2023-10-27 148a9d0
extended the design and implementation of the GUI
