+++ realtime/online_meg/ft_realtime_fieldnulling.m
@@ -28,9 +28,9 @@ if ft_abort
 end
 
 % set the defaults
-cfg.serialport        = ft_getopt(cfg, 'serialport', 'COM2');
-cfg.baudrate          = ft_getopt(cfg, 'baudrate', 921600);
-cfg.fsample           = ft_getopt(cfg, 'fsample', 400);
+cfg.serialport  = ft_getopt(cfg, 'serialport', 'COM2');
+cfg.baudrate    = ft_getopt(cfg, 'baudrate', 921600);
+cfg.fsample     = ft_getopt(cfg, 'fsample', 400);
 
 % these are not used at the moment, the defaults seem to work fine
 cfg.databits    = 8;
@@ -43,26 +43,26 @@ cfg.parity      = 'none';
 % % open the serial device
 % fluxgate = serialport(cfg.serialport, cfg.baudrate);
 % cleanup = onCleanup(@()serial_cleanup(fluxgate));
-% 
+%
 % %  make sure the persistent variables are not reused from the last call
 % clear serial_callback
-% 
-% configureTerminator(fluxgate, "LF");
-% configureCallback(fluxgate, "terminator", @serial_callback);
+%
+% configureTerminator(fluxgate, 'LF');
+% configureCallback(fluxgate, 'terminator', @serial_callback);
 
 %% set up the digital-to-analog converter
 
 % % Create a DataAcquisition object for the specified vendor.
-% dac = daq("ni");
-% 
+% dac = daq('ni');
+%
 % % Add channels and set channel properties, if any.
-% addoutput(dac,"cDAQ1Mod1","ao0","Voltage");
-% addoutput(dac,"cDAQ1Mod1","ao1","Voltage");
-% addoutput(dac,"cDAQ1Mod1","ao2","Voltage");
-% addoutput(dac,"cDAQ1Mod1","ao3","Voltage");
-% addoutput(dac,"cDAQ1Mod1","ao4","Voltage");
-% addoutput(dac,"cDAQ1Mod1","ao5","Voltage");
-% 
+% addoutput(dac,'cDAQ1Mod1','ao0','Voltage');
+% addoutput(dac,'cDAQ1Mod1','ao1','Voltage');
+% addoutput(dac,'cDAQ1Mod1','ao2','Voltage');
+% addoutput(dac,'cDAQ1Mod1','ao3','Voltage');
+% addoutput(dac,'cDAQ1Mod1','ao4','Voltage');
+% addoutput(dac,'cDAQ1Mod1','ao5','Voltage');
+%
 % % Output the specified DC amplitude on each channel.
 % dcOutput = [1 2 3 4 5 6];
 % write(dac, dcOutput);
@@ -79,7 +79,7 @@ setappdata(fig, 'current', [0 0 0 0 0 0 ]);
 % add the callbacks
 set(fig, 'CloseRequestFcn',     @cb_quit);
 set(fig, 'WindowKeypressfcn',   @cb_keyboard);
-set(fig, 'WindowButtondownfcn', @cb_buttonpress);
+set(fig, 'WindowButtondownfcn', @cb_click);
 
 cb_creategui(fig);
 cb_redraw(fig);
@@ -105,7 +105,7 @@ function serial_callback(fluxgate, varargin)
 persistent counter
 
 if isempty(counter)
-    counter = 0;
+  counter = 0;
 end
 
 % The format is as follows:
@@ -127,7 +127,7 @@ end % function
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 function serial_cleanup(fluxgate)
[…]
