### 2023-10-27 08f59d8
completed version 1
- this includes a working voltage output and fluxgate input
