+++ realtime/online_meg/ft_realtime_fieldnulling.m
@@ -1,7 +1,7 @@
 function ft_realtime_fieldnulling(cfg)
 
 % FT_REALTIME_FIELDNULLING is a real-time application to drive the nulling
-% coils in the magnetically shielded room.
+% coilss in the magnetically shielded room.
 %
 % Use as
 %   ft_realtime_fieldnulling(cfg)
@@ -9,6 +9,31 @@ function ft_realtime_fieldnulling(cfg)
 % The configuration should contain
 %   cfg.serialport  = string, name of the serial port (default = 'COM2')
 %   cfg.fsample     = sampling frequency (default = 1000)
+%
+% When clicking the + and - buttons, you can use the shift and ctrl
+% modifier keys to make small and even smaller steps.
+%
+% See also FT_REALTIME_SENSYSPROXY
+
+% Copyright (C) 2023, Robert Oostenveld
+%
+% This file is part of FieldTrip, see http://www.fieldtriptoolbox.org
+% for the documentation and details.
+%
+%    FieldTrip is free software: you can redistribute it and/or modify
+%    it under the terms of the GNU General Public License as published by
+%    the Free Software Foundation, either version 3 of the License, or
+%    (at your option) any later version.
+%
+%    FieldTrip is distributed in the hope that it will be useful,
+%    but WITHOUT ANY WARRANTY; without even the implied warranty of
+%    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+%    GNU General Public License for more details.
+%
+%    You should have received a copy of the GNU General Public License
+%    along with FieldTrip. If not, see <http://www.gnu.org/licenses/>.
+%
+% $Id$
 
 % these are used by the ft_preamble/ft_postamble function and scripts
 ft_revision = '$Id$';
@@ -23,74 +48,87 @@ ft_preamble provenance
 
 % the ft_abort variable is set to true or false in ft_preamble_init
 if ft_abort
-  % do not continue function execution in case the outputfile is present and the user indicated to keep it
-  return
+    % do not continue function execution in case the outputfile is present and the user indicated to keep it
+    return
 end
 
 % set the defaults
 cfg.serialport  = ft_getopt(cfg, 'serialport', 'COM2');
 cfg.baudrate    = ft_getopt(cfg, 'baudrate', 921600);
-cfg.fsample     = ft_getopt(cfg, 'fsample', 400);
+cfg.fsample     = ft_getopt(cfg, 'fsample', 40);
+cfg.databits    = ft_getopt(cfg, 'databits', 8);
+cfg.flowcontrol = ft_getopt(cfg, 'flowcontrol', 'none');
+cfg.stopbits    = ft_getopt(cfg, 'stopbits', 1);
+cfg.parity      = ft_getopt(cfg, 'parity', 'none');
+cfg.position    = ft_getopt(cfg, 'position'); % default is handled below
 
-% these are not used at the moment, the defaults seem to work fine
-cfg.databits    = 8;
-cfg.flowcontrol = 'none';
-cfg.stopbits    = 1;
[…]
