### 2023-10-27 af629bd
made a start with calibration and auto-nulling
