### 2023-10-28 6d4147e
completed version 2 with auto-nulling
- not yet tested as I don't have the hardware here
- implemented calibration measurement and computation
- implemented auto-nulling on basis of the residual measured field
- allow the input and output to be disabled (for development+testing without hardware)
- moved configurable calibration settings to the input cfg
- added two functions to centrally enable and disable the GUI elements
