+++ realtime/online_meg/ft_realtime_fieldnulling.m
@@ -48,23 +48,34 @@ ft_preamble provenance
 
 % the ft_abort variable is set to true or false in ft_preamble_init
 if ft_abort
-    % do not continue function execution in case the outputfile is present and the user indicated to keep it
-    return
+  % do not continue function execution in case the outputfile is present and the user indicated to keep it
+  return
 end
 
 % set the defaults
-cfg.serialport  = ft_getopt(cfg, 'serialport', 'COM2');
-cfg.baudrate    = ft_getopt(cfg, 'baudrate', 921600);
-cfg.fsample     = ft_getopt(cfg, 'fsample', 40);
-cfg.databits    = ft_getopt(cfg, 'databits', 8);
-cfg.flowcontrol = ft_getopt(cfg, 'flowcontrol', 'none');
-cfg.stopbits    = ft_getopt(cfg, 'stopbits', 1);
-cfg.parity      = ft_getopt(cfg, 'parity', 'none');
-cfg.position    = ft_getopt(cfg, 'position'); % default is handled below
+cfg.serialport    = ft_getopt(cfg, 'serialport', 'COM2');
+cfg.baudrate      = ft_getopt(cfg, 'baudrate', 921600);
+cfg.fsample       = ft_getopt(cfg, 'fsample', 40);
+cfg.databits      = ft_getopt(cfg, 'databits', 8);
+cfg.flowcontrol   = ft_getopt(cfg, 'flowcontrol', 'none');
+cfg.stopbits      = ft_getopt(cfg, 'stopbits', 1);
+cfg.parity        = ft_getopt(cfg, 'parity', 'none');
+cfg.position      = ft_getopt(cfg, 'position'); % default is handled below
+cfg.enableinput   = ft_getopt(cfg, 'enableinput', 'yes');
+cfg.enableoutput  = ft_getopt(cfg, 'enableoutput', 'yes');
+cfg.offset        = ft_getopt(cfg, 'offset', [0 0 0 0 0 0]); % initial offset on the coils
+
+% these determine the calibration signal
+cfg.calibration           = ft_getopt(cfg, '', []);
+cfg.calibration.duration  = ft_getopt(cfg, 'duration', 5); % in seconds
+cfg.calibration.nchan     = ft_getopt(cfg, 'nchan', 6);
+cfg.calibration.amplitude = ft_getopt(cfg, 'amplitude', 1);
+cfg.calibration.frequency = ft_getopt(cfg, 'frequency', [5 6 7 8 9 10]); % in Hz
+cfg.calibration.fsample   = ft_getopt(cfg, 'fsample', 1000); % in Hz
 
 if isempty(cfg.position)
-    cfg.position = get(groot, 'defaultFigurePosition');
-    cfg.position(4) = 240;
+  cfg.position = get(groot, 'defaultFigurePosition');
+  cfg.position(4) = 240;
 end
 
 %%
@@ -77,11 +88,13 @@ tmpcfg.figurename = 'ft_realtime_fieldnulling';
 fig = open_figure(tmpcfg);
 drawnow
 
-% store the configuration details and in the application
+% store the configuration details in the application
 setappdata(fig, 'cfg', cfg);
-setappdata(fig, 'field', [0 0 0 0]); % x, y, z, abs
-setappdata(fig, 'offset', [0 0 0 0 0 0]);
[…]
