### 2023-10-28 19a0cab
small change and bugfix to cb_redraw
