### 2023-10-28 e67f035
completed version 3 with continuous closed-loop nulling
- this code has not been tested with real hardware
- replaced the backslash with pinv to get minimum-norm solution
- improve dis/enabling GUI elements when auto nulling and closed-loop nulling
