+++ realtime/online_meg/ft_realtime_fieldnulling.m
@@ -64,6 +64,7 @@ cfg.position      = ft_getopt(cfg, 'position'); % default is handled below
 cfg.enableinput   = ft_getopt(cfg, 'enableinput', 'yes');
 cfg.enableoutput  = ft_getopt(cfg, 'enableoutput', 'yes');
 cfg.offset        = ft_getopt(cfg, 'offset', [0 0 0 0 0 0]); % initial offset on the coils
+cfg.calib         = ft_getopt(cfg, 'calib'); % user-specified calibration values
 
 % these determine the calibration signal
 cfg.calibration           = ft_getopt(cfg, '', []);
@@ -75,6 +76,7 @@ cfg.calibration.fsample   = ft_getopt(cfg, 'fsample', 1000); % in Hz
 
 if isempty(cfg.position)
   cfg.position = get(groot, 'defaultFigurePosition');
+  cfg.position(3) = 800;
   cfg.position(4) = 240;
 end
 
@@ -90,9 +92,9 @@ drawnow
 setappdata(fig, 'cfg', cfg);
 setappdata(fig, 'field', [nan nan nan nan]);  % set the initial measured field: x, y, z, abs
 setappdata(fig, 'offset', cfg.offset);        % set the initial offset
-setappdata(fig, 'running', false);
-setappdata(fig, 'calib', []);
-setappdata(fig, 'calibration', 'off');
+setappdata(fig, 'calib', cfg.calib);
+setappdata(fig, 'running', 'off');            % can be 'off', 'manual', 'closedloop'
+setappdata(fig, 'calibration', 'off');        % can be 'init', 'on', 'off'
 
 % add the callbacks
 set(fig, 'CloseRequestFcn',     @cb_quit);
@@ -101,13 +103,13 @@ set(fig, 'WindowButtondownfcn', @cb_click);
 
 %% set up the serial connection to the fluxgate sensor
 
-measure_callback(fig); % call it once to pass the figure handle
+measure_sample(fig); % call it once to pass the figure handle
 
 if istrue(cfg.enableinput)
   fluxgate = serialport(cfg.serialport, cfg.baudrate);
   cleanup = onCleanup(@()measure_cleanup(fluxgate));
   configureTerminator(fluxgate, 'LF');
-  configureCallback(fluxgate, 'terminator', @measure_callback);
+  configureCallback(fluxgate, 'terminator', @measure_sample);
 else
   fluxgate = [];
 end
@@ -133,7 +135,7 @@ setappdata(fig, 'coils', coils);
 
 %% activate the graphical user interface
 
-setappdata(fig, 'running', true);
+setappdata(fig, 'running', 'manual');
 cb_create_gui(fig);
 cb_redraw(fig);
 
@@ -175,16 +177,22 @@ offset = getappdata(fig, 'offset');
 field = getappdata(fig, 'field');
 calib = getappdata(fig, 'calib');
 
-% only keep the x, y, and z component
-residual = field(1:3);
-residual = residual(:); % it should be a column
-offset = offset(:); % it should be a column
 
 if isempty(calib)
-  warning('cannot auto-null, calibration has not yet been performed')
+  warning('cannot auto null, calibration has not yet been performed')
 
 else
-  disp('auto-null')
+  disp('auto null')
+
[…]
