### 2023-10-28 1144d4b
some improvements to the GUI
- more consistent use of cb_click for all buttons
- avoid get() and use h.Xxxx
