+++ realtime/online_meg/ft_realtime_fieldnulling.m
@@ -15,6 +15,13 @@ function ft_realtime_fieldnulling(cfg)
 %
 % See also FT_REALTIME_SENSYSPROXY
 
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+% Some ideas to implement:
+%  - dithering of output DAC values
+%  - clipping of output values between safe range
+%  - smooth padded transition to prevent overshoots
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
 % Copyright (C) 2023, Robert Oostenveld
 %
 % This file is part of FieldTrip, see http://www.fieldtriptoolbox.org
@@ -90,7 +97,7 @@ drawnow
 
 % store the configuration details in the application
 setappdata(fig, 'cfg', cfg);
-setappdata(fig, 'field', [nan nan nan nan]);  % set the initial measured field: x, y, z, abs
+setappdata(fig, 'field', [0 0 0 0]);          % set the initial measured field: x, y, z, abs
 setappdata(fig, 'offset', cfg.offset);        % set the initial offset
 setappdata(fig, 'calib', cfg.calib);
 setappdata(fig, 'running', 'off');            % can be 'off', 'manual', 'closedloop'
@@ -171,19 +178,16 @@ end % function
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 % SUBFUNCTION
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-function control_auto_null(h, eventdata)
-fig = getparent(h);
+function control_auto_null(fig)
+disp('auto null')
 offset = getappdata(fig, 'offset');
 field = getappdata(fig, 'field');
 calib = getappdata(fig, 'calib');
-
 
 if isempty(calib)
   warning('cannot auto null, calibration has not yet been performed')
 
 else
-  disp('auto null')
-
   % this is incompatible with pairwise linked coils
   set(findobj(fig, 'tag', 'c1x'), 'value', 0);
   set(findobj(fig, 'tag', 'c3x'), 'value', 0);
@@ -213,16 +217,6 @@ else
   control_offset(fig);
 
 end
-end % function
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% SUBFUNCTION
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-function calibration_init(h, eventdata)
-fig = getparent(h);
-cb_disable_gui(fig);
-setappdata(fig', 'calibration', 'init');
-disp('initiated calibration')
 end % function
 
[…]
