### 2023-10-30 b6e5624
implemented user-specified clipping of output
- tested with fluxgate and DAC and fixed some bugs and typos
