+++ realtime/online_meg/ft_realtime_fieldnulling.m
@@ -18,7 +18,6 @@ function ft_realtime_fieldnulling(cfg)
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 % Some ideas to implement:
 %  - dithering of output DAC values
-%  - clipping of output values between safe range
 %  - smooth padded transition to prevent overshoots
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
@@ -72,6 +71,7 @@ cfg.enableinput   = ft_getopt(cfg, 'enableinput', 'yes');
 cfg.enableoutput  = ft_getopt(cfg, 'enableoutput', 'yes');
 cfg.offset        = ft_getopt(cfg, 'offset', [0 0 0 0 0 0]); % initial offset on the coils
 cfg.calib         = ft_getopt(cfg, 'calib'); % user-specified calibration values
+cfg.clip          = ft_getopt(cfg, 'clip', [-10 10]); % clipping for the output
 
 % these determine the calibration signal
 cfg.calibration           = ft_getopt(cfg, '', []);
@@ -165,8 +165,20 @@ end % main function
 % SUBFUNCTION
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 function control_offset(fig)
-coils = getappdata(fig, 'coils');
-offset = getappdata(fig, 'offset');
+cfg     = getappdata(fig, 'cfg');
+coils   = getappdata(fig, 'coils');
+offset  = getappdata(fig, 'offset');
+
+% ensure it is a row vector
+offset = offset(:)';
+
+% clip to the allowed range
+if any(offset<cfg.clip(1) | offset>cfg.clip(2))
+  warning('clipping output')
+  offset(offset>cfg.clip(2)) = cfg.clip(2);
+  offset(offset<cfg.clip(1)) = cfg.clip(1);
+  setappdata(fig, 'offset', offset);
+end
 
 if ~isempty(coils)
   % output the specified DC offset on each of the coils
@@ -180,9 +192,10 @@ end % function
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 function control_auto_null(fig)
 disp('auto null')
-offset = getappdata(fig, 'offset');
-field = getappdata(fig, 'field');
-calib = getappdata(fig, 'calib');
+cfg     = getappdata(fig, 'cfg');
+offset  = getappdata(fig, 'offset');
+field   = getappdata(fig, 'field');
+calib   = getappdata(fig, 'calib');
 
 if isempty(calib)
   warning('cannot auto null, calibration has not yet been performed')
@@ -210,6 +223,13 @@ else
 
   correction = - pinv(calib) * residual;
   offset = offset + correction;
+
+  % clip to the allowed range
+  if any(offset<cfg.clip(1) | offset>cfg.clip(2))
+    warning('clipping output')
+    offset(offset>cfg.clip(2)) = cfg.clip(2);
+    offset(offset<cfg.clip(1)) = cfg.clip(1);
+  end
 
   % show and apply the updated offset values
   setappdata(fig, 'offset', offset);
@@ -246,7 +266,7 @@ end
 if ~isempty(coils)
   preload(coils, signal');
   start(coils);
-  disp('started calibration signal');
[…]
