### 2023-10-30 19b5d31
various fixes and improvements
- fixed manual offset, it was not sent to the DAC
- use disp instead of warning
- changed default clip range to [-1 1]
- automatically update the step size to the clip range
