### 2023-10-31 635c8b6
changed default clipping back to [-10 10], consistent with the hardware
