### 2023-10-31 0a386b5
remove unnecessary reassignment
