### 2023-11-03 562b810
FIX: added abs to prob calculation
