### 2023-11-03 9a4413c
FIX: added abs to prob calculation
