### 2023-11-03 eb69827
ENH - added interpreter option, (as already in singleplotER) for title
