### 2023-11-06 f005570
ENH - deal with upper/lower case lay files
