### 2023-11-06 38b1c36
FIX - only set new offset when the calibration is not running
