### 2023-11-08 5396640
Update ft_realtime_fieldnulling.m
- added help
- fixed z-label for fluxgate
- made cfg.fsample required
- renamed cfg.clip to cfg.range
- don't write DAC values when the same as previous time
- set the output to zero when stopping
