+++ realtime/online_meg/ft_realtime_fieldnulling.m
@@ -7,11 +7,22 @@ function ft_realtime_fieldnulling(cfg)
 %   ft_realtime_fieldnulling(cfg)
 %
 % The configuration should contain
-%   cfg.serialport  = string, name of the serial port (default = 'COM2')
-%   cfg.fsample     = sampling frequency (default = 1000)
+%   cfg.serialport   = string, name of the serial port (default = 'COM2')
+%   cfg.baudrate     = number (default = 921600)
+%   cfg.fsample      = number, sampling rate of the fluxgate AFTER averaging (no default provided)
+%   cfg.enableinput  = string, 'yes' or 'no' to enable the fluxgate input (default = 'yes')
+%   cfg.enableoutput = string, 'yes' or 'no' to enable the NI-DAQ output (default = 'yes')
+%   cfg.range        = [min max], values outside this range will be clipped (default [-10 10])
+%   cfg.offset       = initial offset on the coils (default = [0 0 0 0 0 0])
+%
+% Furthermore, these options determine the calibration signal
+%   cfg.calibration.duration  = number in seconds (default = 5)
+%   cfg.calibration.amplitude = number in Volt (default = 1)
+%   cfg.calibration.frequency = number in Hz for each of the coils (default = [5 6 7 8 9 10])
 %
 % When clicking the + and - buttons, you can use the shift and ctrl
-% modifier keys to make small and even smaller steps.
+% modifier keys to make small and even smaller steps. The step size depends
+% on the range.
 %
 % See also FT_REALTIME_SENSYSPROXY
 
@@ -59,10 +70,14 @@ if ft_abort
   return
 end
 
+% check if the input cfg is valid for this function
+cfg = ft_checkconfig(cfg, 'required', 'fsample');
+cfg = ft_checkconfig(cfg, 'renamed', {'clip', 'range'});
+
 % set the defaults
 cfg.serialport    = ft_getopt(cfg, 'serialport', 'COM2');
 cfg.baudrate      = ft_getopt(cfg, 'baudrate', 921600);
-cfg.fsample       = ft_getopt(cfg, 'fsample', 40);
+cfg.fsample       = ft_getopt(cfg, 'fsample'); % this must be specified
 cfg.databits      = ft_getopt(cfg, 'databits', 8);
 cfg.flowcontrol   = ft_getopt(cfg, 'flowcontrol', 'none');
 cfg.stopbits      = ft_getopt(cfg, 'stopbits', 1);
@@ -72,7 +87,7 @@ cfg.enableinput   = ft_getopt(cfg, 'enableinput', 'yes');
 cfg.enableoutput  = ft_getopt(cfg, 'enableoutput', 'yes');
 cfg.offset        = ft_getopt(cfg, 'offset', [0 0 0 0 0 0]); % initial offset on the coils
 cfg.calib         = ft_getopt(cfg, 'calib'); % user-specified calibration values
-cfg.clip          = ft_getopt(cfg, 'clip', [-10 10]); % clipping for the output
+cfg.range         = ft_getopt(cfg, 'range', [-10 10]); % values outside this range will be clipped
 
 % these determine the calibration signal
 cfg.calibration           = ft_getopt(cfg, '', []);
@@ -87,6 +102,11 @@ if isempty(cfg.position)
   cfg.position(3) = 800;
   cfg.position(4) = 240;
 end
+
+% clear the subfunctions with persistent values
[…]
