### 2023-11-08 b8b64c1
Update ft_realtime_fieldnulling.m
- added cfg.polarity
- renamed offset or output to voltage
- renamed dat or input to field
- fixed bug causing calibration signal to last for just 1s
- fixed bug in computation of calibration values
