+++ realtime/online_meg/ft_realtime_fieldnulling.m
@@ -1,7 +1,7 @@
 function ft_realtime_fieldnulling(cfg)
 
 % FT_REALTIME_FIELDNULLING is a real-time application to drive the nulling
-% coilss in the magnetically shielded room.
+% coils in the magnetically shielded room.
 %
 % Use as
 %   ft_realtime_fieldnulling(cfg)
@@ -13,7 +13,8 @@ function ft_realtime_fieldnulling(cfg)
 %   cfg.enableinput  = string, 'yes' or 'no' to enable the fluxgate input (default = 'yes')
 %   cfg.enableoutput = string, 'yes' or 'no' to enable the NI-DAQ output (default = 'yes')
 %   cfg.range        = [min max], values outside this range will be clipped (default [-10 10])
-%   cfg.offset       = initial offset on the coils (default = [0 0 0 0 0 0])
+%   cfg.voltage      = initial voltage on the coils (default = [0 0 0 0 0 0])
+%   cfg.polarity     = vector with +1 or -1 values to indicate polarity of coils (default = [1 1 1 1 1 1])
 %
 % Furthermore, these options determine the calibration signal
 %   cfg.calibration.duration  = number in seconds (default = 5)
@@ -28,9 +29,9 @@ function ft_realtime_fieldnulling(cfg)
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 % Some ideas to implement:
-%  - dithering of output DAC values
-%  - smooth padded transition to prevent overshoots
-%  - logging of input and output values
+%  - dithering of output DAC values to achive
+%  - smooth padded transitions, to prevent overshoots
+%  - logging of input and output
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 % Copyright (C) 2023, Robert Oostenveld
@@ -66,7 +67,7 @@ ft_preamble provenance
 
 % the ft_abort variable is set to true or false in ft_preamble_init
 if ft_abort
-  % do not continue function execution in case the outputfile is present and the user indicated to keep it
+  % do not continue function execution if something is wrong
   return
 end
 
@@ -85,14 +86,14 @@ cfg.parity        = ft_getopt(cfg, 'parity', 'none');
 cfg.position      = ft_getopt(cfg, 'position'); % default is handled below
 cfg.enableinput   = ft_getopt(cfg, 'enableinput', 'yes');
 cfg.enableoutput  = ft_getopt(cfg, 'enableoutput', 'yes');
-cfg.offset        = ft_getopt(cfg, 'offset', [0 0 0 0 0 0]); % initial offset on the coils
+cfg.voltage       = ft_getopt(cfg, 'voltage', [0 0 0 0 0 0]); % initial voltage on each of the coils
+cfg.polarity      = ft_getopt(cfg, 'polarity', [1 1 1 1 1 1]); % +1 or -1 for each of the coils coils
 cfg.calib         = ft_getopt(cfg, 'calib'); % user-specified calibration values
 cfg.range         = ft_getopt(cfg, 'range', [-10 10]); % values outside this range will be clipped
 
 % these determine the calibration signal
 cfg.calibration           = ft_getopt(cfg, '', []);
 cfg.calibration.duration  = ft_getopt(cfg, 'duration', 5); % in seconds
-cfg.calibration.nchan     = ft_getopt(cfg, 'nchan', 6);
 cfg.calibration.amplitude = ft_getopt(cfg, 'amplitude', 1);
[…]
