### 2023-11-09 b1884a8
do not recompute lf if already existing
