### 2023-11-09 22fb3b7
get the lf of the optimal dipole
