### 2023-11-09 4ef8ff0
use the leadfield of the optimal dipole
