### 2023-11-09 da6cfd0
ENH - Replace non-linear orientation optimization in the SAM spinning… (#2325)
* ENH - Replace non-linear orientation optimization in the SAM spinning case with an equivalent approach based on eigendecomposition, and reorganized the code such that the  'spinning', 'gareth', and 'moiseev' cases of 'fixedori' are merged. The three fixedori-approaches 'spinning', 'gareth', and 'moiseev' all essentially perform the same procedure for orientation estimation. They only differ in whether the whole orientation or only its tangential component are estimated, and in how the noise covariance matrix is es
[…]
