+++ ft_sourceanalysis.m
@@ -71,8 +71,9 @@ function [source] = ft_sourceanalysis(cfg, data, baseline)
 %   cfg.numpermutation     = number, e.g. 500 or 'all'
 %
 % If you have not specified a sourcemodel with pre-computed leadfields, the leadfield
-% for each source position will be computed on the fly. In that case you can modify
-% the leadfields by reducing the rank (i.e. remove the weakest orientation), or by
+% for each source position will be computed on the fly, in the lower level function that
+% is called for the heavy lifting. In that case you can modify parameters for the forward
+% computation, e.g. by reducing the rank (i.e. remove the weakest orientation), or by
 % normalizing each column.
 %   cfg.reducerank      = 'no', or number (default = 3 for EEG, 2 for MEG)
 %   cfg.backproject     = 'yes' or 'no',  determines when reducerank is applied whether the
@@ -87,20 +88,24 @@ function [source] = ft_sourceanalysis(cfg, data, baseline)
 %   cfg.channel       = Nx1 cell-array with selection of channels (default = 'all'), see FT_CHANNELSELECTION for details
 %   cfg.frequency     = single number (in Hz)
 %   cfg.latency       = single number in seconds, for time-frequency analysis
-%   cfg.lambda        = number or empty for automatic default
-%   cfg.kappa         = number or empty for automatic default
-%   cfg.tol           = number or empty for automatic default
 %   cfg.refchan       = reference channel label (for coherence)
 %   cfg.refdip        = reference dipole location (for coherence)
 %   cfg.supchan       = suppressed channel label(s)
 %   cfg.supdip        = suppressed dipole location(s)
 %   cfg.keeptrials    = 'no' or 'yes'
 %   cfg.keepleadfield = 'no' or 'yes'
-%   cfg.projectnoise  = 'no' or 'yes'
-%   cfg.keepfilter    = 'no' or 'yes'
-%   cfg.keepcsd       = 'no' or 'yes'
-%   cfg.keepmom       = 'no' or 'yes'
-%   cfg.feedback      = 'no', 'text', 'textbar', 'gui' (default = 'text')
+%
+% Some options need to be specified as method specific options, and determine the low-level computation of the inverse operator.
+% The functionality (and applicability) of the (sub-)options are documented in the lower-level ft_inverse_<method> functions. 
+% Replace <method> with one of the supported methods.  
+%   cfg.<method>.lambda        = number or empty for automatic default
+%   cfg.<method>.kappa         = number or empty for automatic default
+%   cfg.<method>.tol           = number or empty for automatic default
+%   cfg.<method>.projectnoise  = 'no' or 'yes'
+%   cfg.<method>.keepfilter    = 'no' or 'yes'
+%   cfg.<method>.keepcsd       = 'no' or 'yes'
+%   cfg.<method>.keepmom       = 'no' or 'yes'
+%   cfg.<method>.feedback      = 'no', 'text', 'textbar', 'gui' (default = 'text')
 %
 % The volume conduction model of the head should be specified as
 %   cfg.headmodel     = structure with volume conduction model, see FT_PREPARE_HEADMODEL
@@ -124,7 +129,7 @@ function [source] = ft_sourceanalysis(cfg, data, baseline)
 % cfg.numcomponents
 % cfg.refchannel
 % cfg.trialweight   = 'equal' or 'proportional'
-% cfg.powmethod     = 'lambda1' or 'trace'
+% cfg.<method>.powmethod     = 'lambda1' or 'trace'
 
 % Copyright (c) 2003-2008, F.C. Donders Centre, Robert Oostenveld
 %
