### 2023-11-09 #5566 @matsvanes
Hi Jay,

The installation instructions can be found here: https://www.fieldtriptoolbox.org/download/
In short, after [downloading](https://www.fieldtriptoolbox.org/download.php) the toolbox, you should add it to your path with `addpath <path_to_fieldtrip>`, and then run `ft_defaults`, which will setup the rest.

For requirements, see https://www.fieldtriptoolbox.org/faq/requirements/. Depending on the fieldtrip functionality you want to use, you may be required to have (licensed) toolboxes installed.
