### 2023-11-09 a7348e7
Update ft_realtime_fieldnulling.m
- improved state change between calibration on and off
- don't redraw with every new fluxgate sample, but only once per second
- fixed parsing of cfg.calibration
- added graphical feedback when computing calibration
- optional padding in calibration
