+++ realtime/online_meg/ft_realtime_fieldnulling.m
@@ -18,6 +18,7 @@ function ft_realtime_fieldnulling(cfg)
 %
 % Furthermore, these options determine the calibration signal
 %   cfg.calibration.duration  = number in seconds (default = 5)
+%   cfg.calibration.pad       = number in seconds (default = 0)
 %   cfg.calibration.amplitude = number in Volt (default = 1)
 %   cfg.calibration.frequency = number in Hz for each of the coils (default = [5 6 7 8 9 10])
 %
@@ -92,11 +93,12 @@ cfg.calib         = ft_getopt(cfg, 'calib'); % user-specified calibration values
 cfg.range         = ft_getopt(cfg, 'range', [-10 10]); % values outside this range will be clipped
 
 % these determine the calibration signal
-cfg.calibration           = ft_getopt(cfg, '', []);
-cfg.calibration.duration  = ft_getopt(cfg, 'duration', 5); % in seconds
-cfg.calibration.amplitude = ft_getopt(cfg, 'amplitude', 1);
-cfg.calibration.frequency = ft_getopt(cfg, 'frequency', [5 6 7 8 9 10]); % in Hz
-cfg.calibration.fsample   = ft_getopt(cfg, 'fsample', 1000); % in Hz
+cfg.calibration           = ft_getopt(cfg, 'calibration', []);
+cfg.calibration.duration  = ft_getopt(cfg.calibration, 'duration', 5); % in seconds
+cfg.calibration.pad       = ft_getopt(cfg.calibration, 'pad', 0); % in seconds
+cfg.calibration.amplitude = ft_getopt(cfg.calibration, 'amplitude', 1);
+cfg.calibration.frequency = ft_getopt(cfg.calibration, 'frequency', [5 6 7 8 9 10]); % in Hz
+cfg.calibration.fsample   = ft_getopt(cfg.calibration, 'fsample', 1000); % in Hz
 
 if isempty(cfg.position)
   cfg.position = get(groot, 'defaultFigurePosition');
@@ -169,7 +171,12 @@ setappdata(fig, 'coils', coils);
 ft_info('creating gui');
 setappdata(fig, 'running', 'manual');
 cb_create_gui(fig);
-cb_redraw(fig);
+
+t = timer;
+t.Period = 1;
+t.ExecutionMode = 'fixedRate';
+t.TimerFcn = {@cb_redraw, fig};
+t.start();
 
 while ishandle(fig)
   pause(0.1);
@@ -197,7 +204,6 @@ voltage = getappdata(fig, 'voltage');
 
 % clip the voltage to the specified range
 if any(voltage<cfg.range(1) | voltage>cfg.range(2))
-  ft_info('clipping voltage')
   voltage(voltage>cfg.range(2)) = cfg.range(2);
   voltage(voltage<cfg.range(1)) = cfg.range(1);
   setappdata(fig, 'voltage', voltage);
@@ -256,9 +262,8 @@ if ~isempty(calib)
     voltage(voltage<cfg.range(1)) = cfg.range(1);
   end
 
-  % show and apply the updated voltage values
+  % output the updated voltage
   setappdata(fig, 'voltage', voltage);
-  cb_redraw(fig);
   control_voltage(fig);
 
 else
@@ -295,7 +300,10 @@ end
 if ~isempty(coils) && ~coils.Running
   preload(coils, voltage');
   start(coils, 'repeatoutput');
-  fprintf('started calibration signal for %.1f seconds\n', cfg.calibration.duration);
+  while ~coils.Running
+    pause(0.1);
+  end
[…]
