### 2023-11-13 6ca88cc
ENH - speed up computation outside for-loop (#2342)
