### 2023-11-13 445e92d
ENH - pass on optional montage into callbacks (#2340)
