### 2023-11-13 15beeab
ENH - deal with delay and phase difference between input voltage and output field
- changed some defaults
- reduce startup delay for calibration signal
