+++ realtime/online_meg/ft_realtime_fieldnulling.m
@@ -20,7 +20,7 @@ function ft_realtime_fieldnulling(cfg)
 %   cfg.calibration.duration  = number in seconds (default = 5)
 %   cfg.calibration.pad       = number in seconds (default = 0)
 %   cfg.calibration.amplitude = number in Volt (default = 1)
-%   cfg.calibration.frequency = number in Hz for each of the coils (default = [5 6 7 8 9 10])
+%   cfg.calibration.frequency = number in Hz for each of the coils (default = [1 2 3 4 5 6])
 %
 % When clicking the + and - buttons, you can use the shift and ctrl
 % modifier keys to make small and even smaller steps. The step size depends
@@ -97,7 +97,7 @@ cfg.calibration           = ft_getopt(cfg, 'calibration', []);
 cfg.calibration.duration  = ft_getopt(cfg.calibration, 'duration', 5); % in seconds
 cfg.calibration.pad       = ft_getopt(cfg.calibration, 'pad', 0); % in seconds
 cfg.calibration.amplitude = ft_getopt(cfg.calibration, 'amplitude', 1);
-cfg.calibration.frequency = ft_getopt(cfg.calibration, 'frequency', [5 6 7 8 9 10]); % in Hz
+cfg.calibration.frequency = ft_getopt(cfg.calibration, 'frequency', [1 2 3 4 5 6]); % in Hz
 cfg.calibration.fsample   = ft_getopt(cfg.calibration, 'fsample', 1000); % in Hz
 
 if isempty(cfg.position)
@@ -124,7 +124,7 @@ setappdata(fig, 'cfg', cfg);
 setappdata(fig, 'field', [0 0 0 0]);          % set the initial value for the field: x, y, z, abs
 setappdata(fig, 'voltage', cfg.voltage);      % set the initial voltage
 setappdata(fig, 'calib', cfg.calib);
-setappdata(fig, 'running', 'off');            % can be 'off', 'manual', 'closedloop'
+setappdata(fig, 'closedloop', 'off');         % can be 'off', 'on'
 setappdata(fig, 'calibration', 'off');        % can be 'init', 'on', 'off'
 
 % add the callbacks
@@ -169,7 +169,7 @@ setappdata(fig, 'coils', coils);
 %% activate the graphical user interface
 
 ft_info('creating gui');
-setappdata(fig, 'running', 'manual');
+setappdata(fig, 'closedloop', 'off');
 cb_create_gui(fig);
 
 t = timer;
@@ -276,7 +276,7 @@ end % function
 % SUBFUNCTION
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 function calibration_start_signal(fig)
-% output a sine wave with a different frequency on each of the coils
+% output a sine-wave with a different frequency on each of the coils
 cfg     = getappdata(fig, 'cfg');
 coils   = getappdata(fig, 'coils');
 voltage = getappdata(fig, 'voltage');
@@ -290,7 +290,7 @@ assert(numel(cfg.calibration.amplitude)==1);
 
 coils.Rate = fsample;
 
-% create one second of signal, it will loop until finished
+% create one second of a sine-wave signal, it will loop until finished
 voltage = zeros(ncoil, nsample);
 time = (0:(nsample-1))/fsample;
 for i=1:ncoil
@@ -301,7 +301,7 @@ if ~isempty(coils) && ~coils.Running
   preload(coils, voltage');
   start(coils, 'repeatoutput');
   while ~coils.Running
-    pause(0.1);
+    pause(0.05);
   end
[…]
