+++ realtime/online_meg/ft_realtime_fieldnulling.m
@@ -1,870 +1,0 @@
-function ft_realtime_fieldnulling(cfg)
-
-% FT_REALTIME_FIELDNULLING is a real-time application to drive the nulling
-% coils in the magnetically shielded room.
-%
-% Use as
-%   ft_realtime_fieldnulling(cfg)
-%
-% The configuration should contain
-%   cfg.serialport   = string, name of the serial port (default = 'COM2')
-%   cfg.baudrate     = number (default = 921600)
-%   cfg.fsample      = number, sampling rate of the fluxgate AFTER averaging (no default provided)
-%   cfg.enableinput  = string, 'yes' or 'no' to enable the fluxgate input (default = 'yes')
-%   cfg.enableoutput = string, 'yes' or 'no' to enable the NI-DAQ output (default = 'yes')
-%   cfg.range        = [min max], values outside this range will be clipped (default [-10 10])
-%   cfg.voltage      = initial voltage on the coils (default = [0 0 0 0 0 0])
-%   cfg.polarity     = vector with +1 or -1 values to indicate polarity of coils (default = [1 1 1 1 1 1])
-%
-% Furthermore, these options determine the calibration signal
-%   cfg.calibration.duration  = number in seconds (default = 5)
-%   cfg.calibration.pad       = number in seconds (default = 0)
-%   cfg.calibration.amplitude = number in Volt (default = 1)
-%   cfg.calibration.frequency = number in Hz for each of the coils (default = [1 2 3 4 5 6])
-%
-% When clicking the + and - buttons, you can use the shift and ctrl
-% modifier keys to make small and even smaller steps. The step size depends
-% on the range.
-%
-% See also FT_REALTIME_SENSYSPROXY
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% Some ideas to implement:
-%  - dithering of output DAC values to achive
-%  - smooth padded transitions, to prevent overshoots
-%  - logging of input and output
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-% Copyright (C) 2023, Robert Oostenveld
-%
-% This file is part of FieldTrip, see http://www.fieldtriptoolbox.org
-% for the documentation and details.
-%
-%    FieldTrip is free software: you can redistribute it and/or modify
-%    it under the terms of the GNU General Public License as published by
-%    the Free Software Foundation, either version 3 of the License, or
-%    (at your option) any later version.
-%
-%    FieldTrip is distributed in the hope that it will be useful,
-%    but WITHOUT ANY WARRANTY; without even the implied warranty of
-%    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-%    GNU General Public License for more details.
-%
-%    You should have received a copy of the GNU General Public License
-%    along with FieldTrip. If not, see <http://www.gnu.org/licenses/>.
-%
-% $Id$
-
-% these are used by the ft_preamble/ft_postamble function and scripts
-ft_revision = '$Id$';
-ft_nargin   = nargin;
-ft_nargout  = nargout;
-
[…]
