### 2023-11-14 #5566 @gflofst
I'm at SC23 this week. I doubt I'll have time to do this before Saturday.

From: Mats ***@***.***>
Sent: Thursday, November 9, 2023 7:12 AM
Cc: Lofstead, Jay F ***@***.***>; Mention ***@***.***>
Subject: [EXTERNAL] Re: [openjournals/joss-reviews] [REVIEW]: Reducing the efforts to create reproducible analysis code with FieldTrip (Issue #5566)

Hi Jay,

The installation instructions can be found here: https://www.fieldtriptoolbox.org/download/
In short, after downloading<https://www.fieldtriptoolbox.org/download.php> the toolbox, you should add it to your path with addpath <path_to_fieldtrip>, and then run ft_defaults, which will setup the rest.

For requirements, see https://www.fieldtriptoolbox.org/faq/requirements/. Depending on the fieldtrip functionality you want to use, you may be required to have (licensed) toolboxes installed.

-
You are receiving this because you were mentioned.Message ID: ***@***.******@***.***>>
