### 2023-12-01 #5566 @gflofst
It looks good and technically meets the requirements. I'd prefer if there was more in the github rather than their own website (e.g., installation instructions), but it is all finable without too much effort. I vote to accept.
