### 2023-12-01 #5566 @robertoostenveld
> I'd prefer if there was more in the github rather than their own website (e.g., installation instructions), but it is all finable without too much effort. I vote to accept.

Note that the website itself is also on github (separate from the code), including documentation on reproduce script [here](https://github.com/fieldtrip/website/blob/master/example/reproducescript.md), [here](https://github.com/fieldtrip/website/blob/master/example/reproducescript_group.md) and [here](https://github.com/fieldtrip/website/blob/master/example/reproducescript_andersen.md). It is just that the rendering as html on our own server looks better and it makes us independent from github/microsoft as commercial service provider.

Having independent installation instructions in the git repository, i.e., also on github, [bitbucket](https://bitbucket.org/fieldtriptoolbox/fieldtrip) and [gitlab](https://gitlab.com/fieldtrip/fieldtrip), separate from those on the website, is a good idea and simple to implement. I will add that to the [README](https://github.com/fieldtrip/fieldtrip/blob/master/README.md) file.
