### 2023-12-01 20044da
ENH - added installation instructions to README, see https://github.com/openjournals/joss-reviews/issues/5566#issuecomment-1835726996
