### 2024-01-15 #5566 @matsvanes
> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
>
>     * [x]  Double check authors and affiliations (including ORCIDs)
>
>     * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. - _Version 20231220 (Note that it doesn't appear to be the last version on Zenodo, but that's incorrect and Zenodo is aware of the issue - this version/DOI can be used however).
>_
>     * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here. - https://doi.org/10.5281/zenodo.10495308
>
>     * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper. - @crsl4 please advise what to do here, since the software is an extension the author list is not the same as in the original. The authors of the original do match the ones in the archive
>
>     * [x]  Make sure that the license listed for the archive is the same as the software license.
>
>
> ## Editor Tasks Prior to Acceptance
>
>     * [ ]  Read the text of the paper and offer comments/corrections (as either a list or a PR)
>
>     * [ ]  Check the references in the paper for corrections (e.g. capitalization)
>
>     * [ ]  Check that the archive title, author list, version tag, and the license are correct
>
>     * [ ]  Set archive DOI with `@editorialbot set <DOI here> as archive`
>
>     * [ ]  Set version with `@editorialbot set <version here> as version`
>
>     * [ ]  Double check rendering of paper with `@editorialbot generate pdf`
>
>     * [ ]  Specifically check the references with `@editorialbot check references` and ask author(s) to update as needed
>
>     * [ ]  Recommend acceptance with `@editorialbot recommend-accept`
