### 2024-01-30 #5566 @crsl4
@matsvanes Just for a stylistic perspective, can you make sure that the snippets of code (or comments within the code) are not truncated in the PDF?
