### 2024-02-05 #5566 @crsl4
Thanks @matsvanes ! Honestly, this appears to be something going wrong in the rendering of the PDF, so I am not sure you can do something on your end. Let's see what @Kevin-Mattheus-Moerman thinks.
