### 2024-02-09 #5566 @matsvanes
@crsl4 I updated the Matlab reference to "software" in bibtex. Hopefully that will sort it, since it doesn't have a DOI?
