### 2024-02-09 #5566 @crsl4
@Kevin-Mattheus-Moerman We need help! The references do not appear in the PDF, but the check of references via the editorialbot is ok. Anything we are missing?
