### 2024-02-10 #5566 @xuanxu
The paper file includes text formatting for the references and so they are not read as proper citations.
[This PR](https://github.com/fieldtrip/fieldtrip/pull/2390) should fix it.
