### 2024-02-20 #5566 @robertoostenveld
We have the release https://github.com/fieldtrip/fieldtrip/releases/tag/20231220, right? That is the release which on Zenodo corresponds to https://doi.org/10.5281/zenodo.10495308.

Later tags such as `20240214` are in-between stable versions but not proper releases, and hence not optimal to refer to.
