### 2024-02-22 #5566 @ashahide
Congratulations on the acceptance, and all of your hard work! It was a
pleasure to read.

On Thu, Feb 22, 2024 at 3:55 PM Kevin Mattheus Moerman <
***@***.***> wrote:

> @matsvanes <https://github.com/matsvanes> congratulations on this JOSS
> publication!
>
> Thanks for editing @crsl4 <https://github.com/crsl4> ! And a special
> thanks to the reviewers: @gflofst <https://github.com/gflofst>, @ashahide
> <https://github.com/ashahide> !!!
>
